### 2021-10-12 #3817 @whedon
[PRE REVIEW]: COMPAS: A rapid binary population synthesis suite
**Submitting author:** @ilyamandel (<a href="http://orcid.org/0000-0002-6134-8946">Ilya Mandel</a>)
**Repository:** <a href="https://github.com/TeamCOMPAS/COMPAS" target ="_blank">https://github.com/TeamCOMPAS/COMPAS</a>
**Version:** v02.21.00
