### 2021-10-12 d55abb7
Removed paper.tex
