+++ paper.tex
@@ -1,445 +1,0 @@
-\documentclass[10pt,a4paper,onecolumn]{article}
-\usepackage{marginnote}
-\usepackage{graphicx}
-\usepackage{xcolor}
-\usepackage{authblk,etoolbox}
-\usepackage{titlesec}
-\usepackage{calc}
-\usepackage{tikz}
-\usepackage{hyperref}
-\hypersetup{colorlinks,breaklinks,
-            urlcolor=[rgb]{0.0, 0.5, 1.0},
-            linkcolor=[rgb]{0.0, 0.5, 1.0}}
-\usepackage{caption}
-\usepackage{tcolorbox}
-\usepackage{amssymb,amsmath}
-\usepackage{ifxetex,ifluatex}
-\usepackage{seqsplit}
-\usepackage{xstring}
-
-\usepackage{float}
-\let\origfigure\figure
-\let\endorigfigure\endfigure
-\renewenvironment{figure}[1][2] {
-    \expandafter\origfigure\expandafter[H]
-} {
-    \endorigfigure
-}
-
-\usepackage{fixltx2e} % provides \textsubscript
-\usepackage[
-  backend=biber,
-%  style=alphabetic,
-%  citestyle=numeric
-]{biblatex}
-\bibliography{paper.bib}
-
-% --- Splitting \texttt --------------------------------------------------
-
-\let\textttOrig=\texttt
-\def\texttt#1{\expandafter\textttOrig{\seqsplit{#1}}}
-\renewcommand{\seqinsert}{\ifmmode
-  \allowbreak
-  \else\penalty6000\hspace{0pt plus 0.02em}\fi}
-
-
-% --- Pandoc does not distinguish between links like [foo](bar) and
-% --- [foo](foo) -- a simplistic Markdown model.  However, this is
-% --- wrong:  in links like [foo](foo) the text is the url, and must
-% --- be split correspondingly.
-% --- Here we detect links \href{foo}{foo}, and also links starting
-% --- with https://doi.org, and use path-like splitting (but not
-% --- escaping!) with these links.
-% --- Another vile thing pandoc does is the different escaping of
-% --- foo and bar.  This may confound our detection.
-% --- This problem we do not try to solve at present, with the exception
-% --- of doi-like urls, which we detect correctly.
-
-
-\makeatletter
-\let\href@Orig=\href
-\def\href@Urllike#1#2{\href@Orig{#1}{\begingroup
-    \def\Url@String{#2}\Url@FormatString
-    \endgroup}}
-\def\href@Notdoi#1#2{\def\tempa{#1}\def\tempb{#2}%
-  \ifx\tempa\tempb\relax\href@Urllike{#1}{#2}\else
-  \href@Orig{#1}{#2}\fi}
-\def\href#1#2{%
-  \IfBeginWith{#1}{https://doi.org}%
-  {\href@Urllike{#1}{#2}}{\href@Notdoi{#1}{#2}}}
-\makeatother
-
-
-% --- Page layout -------------------------------------------------------------
-\usepackage[top=3.5cm, bottom=3cm, right=1.5cm, left=1.0cm,
-            headheight=2.2cm, reversemp, includemp, marginparwidth=4.5cm]{geometry}
-
-% --- Default font ------------------------------------------------------------
-% \renewcommand\familydefault{\sfdefault}
-
-% --- Style -------------------------------------------------------------------
-\renewcommand{\bibfont}{\small \sffamily}
-\renewcommand{\captionfont}{\small\sffamily}
-\renewcommand{\captionlabelfont}{\bfseries}
-
-% --- Section/SubSection/SubSubSection ----------------------------------------
-\titleformat{\section}
[…]
