### 2021-10-12 #3817 @ilyamandel
> @ilyamandel – am I right in thinking that this submission accompanies a paper in a AAS journal?

@arfon -- Correct.  A full methods paper been submitted to ApJS and is under review.  The AAS editor is on board with our submitting it to JOSS under the JOSS/AAS collaboration.
