### 2021-10-12 29ba931
fixed semi-major axis help string consistencies, and clarified BH neutrino mass loss value string
