### 2021-10-12 590497d
removed ambiguous properties MASS_[12]_FINAL
