### 2021-10-12 abaa0b5
removed ambiguous properties MASS_[12]_FINAL
