### 2021-10-12 4743db0
cleaned help string for MCBUR1
