### 2021-10-13 e74b3f6
Minor fixes (#657)
* Minor fixes:
(a) Fixed a few typos in header strings
(b) Changed true_anomaly to mean_anomaly in SSE SN output

* Delete settings.json

I need to add this to .gitignore...
