+++ examples/methods_paper_plots/detailed_evolution/detailed_evol_plotter.py
@@ -156,6 +156,7 @@ def printEvolutionaryHistory(Data):
     if Data['Time'][i]<14000 and not((Data['Stellar_Type(1)'][i]==13 or Data['Stellar_Type(1)'][i]==14) and (Data['Stellar_Type(2)'][i]==13 or Data['Stellar_Type(2)'][i]==14)):     #proxy for unrecorded meregr
         print(Data['Time'][i], '   Stellar merger')
 
+    #Merger time calculation follows https://iopscience.iop.org/article/10.3847/2515-5172/ac2d35
     if ((Data['Stellar_Type(1)'][i]==13 or Data['Stellar_Type(1)'][i]==14) and (Data['Stellar_Type(2)'][i]==13 or Data['Stellar_Type(2)'][i]==14)):
         Msunkg=1.98892e30
         c=299792458
