### 2021-10-13 fddfd2f
Add disruption check to merger line
