### 2021-10-14 530c3a0
some tweaks to the formatting of printed lines, to improve clarity
