### 2021-10-15 #3817 @christinahedges
@katiebreivik and @HeloiseS, just so you are aware this JOSS submission/paper is a joint publication with AAS Journals, i.e. this JOSS paper will be published together with another paper in one of the AAS journals. As part of this collaboration, AAS publishing makes a small donation to the running costs of JOSS. (as an example, in 2019, JOSS received $200 as a result of this collaboration).

If this is unacceptable to you, please let me know and I can look for alternative reviewers.

If you'd like to read more about the JOSS costs and funding from this collaboration, please see below:

Blog post describing the collaboration: https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
Summary of JOSS financial model: https://joss.theoj.org/about#costs
