### 2021-10-16 9063168
Brief evolutionary history printing (#644)
* Brief evolutionary history printing

* Add reference to https://iopscience.iop.org/article/10.3847/2515-5172/ac2d35

* Add disruption check to merger line

* Add check for DWD

* some tweaks to the formatting of printed lines, to improve clarity

Co-authored-by: Reinhold Willcox <reinhold.willcox@mail.mcgill.ca>
