+++ examples/methods_paper_plots/detailed_evolution/detailed_evol_plotter.py
@@ -81,11 +81,14 @@ def main():
     axes[1][0].text( axes[1][0].get_xlim()[1]*-.18, axes[1][0].get_ylim()[1]*0.87,  'c)', fontweight='bold')
     axes[1][1].text( axes[1][1].get_xlim()[1]*-.25, axes[1][1].get_ylim()[1]*0.87,  'd)', fontweight='bold')
     
+    ### Print out a synopsys of the evolutionary history to the command line
+    printEvolutionaryHistory(Data)
+    
     ### Finalize the boundaries, save, and show
     fig.subplots_adjust(left=0.05)  #adjusting boundaries of the plotter
     fig.subplots_adjust(wspace=.3)
-    plt.savefig('gw151226evol.eps', format='eps') 
-    plt.savefig('gw151226evol.png', format='png') 
+    plt.savefig('gw151226evol.eps', format='eps')
+    plt.savefig('gw151226evol.png', format='png')
     plt.show()
 
 
@@ -111,10 +114,85 @@ def getStellarTypes(Data):
 
     return stellarTypes, useTypes, typeNameMap
 
-""
+
+def printFormattedEvolutionLine(time, event, m1, t1, m2, t2, a, e):
+    # All values are floats except event which is a string and t1, t2 which are ints
+    print("{:10.6f}   {:21}  {:7.3f}    {:2}    {:7.3f}    {:2}   {:8.3f}  {:5.3f}" .format(time, event, m1, t1, m2, t2, a, e))
+
+def printEvolutionaryHistory(Data):
+    """
+    This function prints a synopsys of the evolutionary history to the command line; it can eventually include cartoons as well.
+    """
+
+    print('Time (Myr), Event,                  M1 (M_o), type1, M2 (M_o), type2, a (R_o),   e')
+
+    printFormattedEvolutionLine( Data['Time'][0], 'start:Z='+str(Data['Metallicity@ZAMS(1)'][0]),  
+                                 Data['Mass(1)'][0], Data['Stellar_Type(1)'][0],  
+                                 Data['Mass(2)'][0], Data['Stellar_Type(2)'][0],  
+                                 Data['SemiMajorAxis'][0],  Data['Eccentricity'][0])
+
+    for i in range(1,Data['Time'].size):
+        if Data['MT_History'][i]>0:         #mass transfer happened
+            if Data['MT_History'][i]==Data['MT_History'][i-1]:
+                continue    #repeated entry
+            if Data['MT_History'][i]==1:
+                MTstring='Stable MT: 1 to 2'
+            elif Data['MT_History'][i]==2:
+                MTstring='Stable MT: 2 to 1'
+            elif Data['MT_History'][i]==3:
+                MTstring='CE: 1 to 2'
+            elif Data['MT_History'][i]==4:
+                MTstring='CE: 2 to 1'
+            elif Data['MT_History'][i]==5:
+                MTstring='CE: Double Core'
+            elif Data['MT_History'][i]==6:
+                MTstring='CE: both MS'
+            elif Data['MT_History'][i]==7:
+                MTstring='CE: MS with CO'
+            else:
+                MTstring='Unknown MT'
+
+            printFormattedEvolutionLine( Data['Time'][i], MTstring, 
+                                         Data['Mass(1)'][i], Data['Stellar_Type(1)'][i], 
+                                         Data['Mass(2)'][i], Data['Stellar_Type(2)'][i], 
[…]
