### 2021-10-20 #3838 @whedon
[REVIEW]: COMPAS: A rapid binary population synthesis suite
**Submitting author:** @ilyamandel (<a href="http://orcid.org/0000-0002-6134-8946">Ilya Mandel</a>)
**Repository:** <a href="https://github.com/TeamCOMPAS/COMPAS" target ="_blank">https://github.com/TeamCOMPAS/COMPAS</a>
**Version:** v02.27.00

- [x] #3952
