+++ src/Options.cpp
@@ -108,6 +108,9 @@ Options* Options::Instance() {
 #define COMPLAIN(complainStr)           { std::stringstream _ss; _ss << complainStr; throw _ss.str(); }
 #define COMPLAIN_IF(cond, complainStr)  { if (cond) COMPLAIN(complainStr) }
 
+#define WARNUSER(warnStr)               { std::cerr << warnStr << std::endl; }
+#define WARNUSER_IF(cond, warnStr)      { if (cond) WARNUSER(warnStr) }
+
 
 /*
  * Initialise option values
@@ -153,6 +156,11 @@ void Options::OptionValues::Initialise() {
     m_StoreInputFiles                                               = true;
 
     m_SwitchLog                                                     = false;
+
+
+    // annotations
+    m_Notes.clear();
+    m_NotesHdrs.clear();
 
 
     // Evolution mode: SSE or BSE
@@ -558,14 +566,28 @@ bool Options::AddOptions(OptionValues *p_Options, po::options_description *p_Opt
     ss << "";
     for (auto debugClass = p_Options->m_DebugClasses.begin(); debugClass != p_Options->m_DebugClasses.end(); ++debugClass) ss << *debugClass << ",";
     defaultDebugClasses = ss.str();
-    if (defaultDebugClasses.size() > 0) defaultDebugClasses.erase(defaultDebugClasses.size() - 1);
+    if (defaultDebugClasses.length() > 0) defaultDebugClasses.erase(defaultDebugClasses.length() - 1);
 
     // log classes
     std::string defaultLogClasses;
     ss << "";
     for (auto logClass = p_Options->m_LogClasses.begin(); logClass != p_Options->m_LogClasses.end(); ++logClass) ss << *logClass << ",";
     defaultLogClasses = ss.str();
-    if (defaultLogClasses.size() > 0) defaultLogClasses.erase(defaultLogClasses.size() - 1);
+    if (defaultLogClasses.length() > 0) defaultLogClasses.erase(defaultLogClasses.length() - 1);
+
+    // annotations
+    std::string defaultNotes;
+    ss << "";
+    for (auto note = p_Options->m_Notes.begin(); note != p_Options->m_Notes.end(); ++note) ss << *note << ",";
+    defaultNotes = ss.str();
+    if (defaultNotes.length() > 0) defaultNotes.erase(defaultNotes.length() - 1);
+
+    // annotation headers
+    std::string defaultNotesHdrs;
+    ss << "";
+    for (auto noteHdr = p_Options->m_NotesHdrs.begin(); noteHdr != p_Options->m_NotesHdrs.end(); ++noteHdr) ss << *noteHdr << ",";
+    defaultNotesHdrs = ss.str();
+    if (defaultNotesHdrs.length() > 0) defaultNotesHdrs.erase(defaultNotesHdrs.length() - 1);
 
 
     // add options
@@ -1561,10 +1583,22 @@ bool Options::AddOptions(OptionValues *p_Options, po::options_description *p_Opt
             po::value<vector<std::string>>(&p_Options->m_DebugClasses)->multitoken()->default_value(p_Options->m_DebugClasses),                                                                        
             ("Debug classes enabled (default = " + defaultDebugClasses + ")").c_str()
         )
+
         (
             "log-classes",                                                 
             po::value<vector<std::string>>(&p_Options->m_LogClasses)->multitoken()->default_value(p_Options->m_LogClasses),                                                                            
             ("Logging classes enabled (default = " + defaultLogClasses + ")").c_str()
+        )
+
+        (
+            "notes",                                                 
+            po::value<vector<std::string>>(&p_Options->m_Notes)->multitoken()->default_value(p_Options->m_Notes),                                                                            
[…]
