+++ src/Options.h
@@ -33,7 +33,7 @@ namespace po = boost::program_options;
 
 // OPT_VALUE macro
 //
-// getter functions return the value of the class member variable - the class
+// Getter functions return the value of the class member variable - the class
 // member variable is set to a value depending upon the value of the corresponding
 // option enetered by the user.
 // 
@@ -115,7 +115,10 @@ class Options {
 private:
 
     // The following vectors are used to constrain which options can be specified
-    // when.  There are three groups of two vectors:
+    // when:
+    //
+    // m_ShorthandAllowed records option strings that may be specified using the
+    // shorthand notation described in Options::PreprocessOptionValues()
     //
     // m_GridLineExcluded records option strings that may not be specified on a grid line
     //
@@ -128,6 +131,38 @@ private:
     // Each of these is described in more detail below
 
 
+    // m_ShorthandAllowed records option strings that may be specified using the
+    // shorthand notation described in Options::ExpandShorthandOptionValues().
+    //
+    // Furthermore, the vector records whether option values for such options can be
+    // defaulted (i.e. some or all values need not be specified), and if so, what
+    // string should be substituted for the unspecified values (at this stage there is
+    // no type associated with the options - what is being manipulated is the command-line
+    // or grid-line string that will be parsed to determine options specified and their
+    // values).
+    //
+    // This vector is checked immediately prior to the command line or grid line being parsed
+    // for ranges and sets - i.e. before the command line or grid line is passed to boost
+    // for final parsing.  If any option strings not in this vector are specified using
+    // shorthand notation, boost will parse them as usual and likely (though not necessarily)
+    // complain (boost will only complain if the option/value pair is malformed or unknown,
+    // which would almost certainly be the case - but it isn't guaranteed to be). 
+
+    typedef std::tuple<std::string, bool, std::string> SHORTHAND_ENTRY;         // option name, default allowed (i.e. can be omitted), default string
+    std::vector<SHORTHAND_ENTRY> m_ShorthandAllowed = {
+
+        // trying to keep entries alphabetical so easier to find specific entries
+
+        // option name          default allowed     default string
+        { "debug-classes",      false,              "" },
+
+        { "log-classes",        false,              "" },
+
+        { "notes",              true,               "" },
+        { "notes-hdrs",         false,              "" }
+    };
+
+
     // m_GridLineExcluded records option strings that may not be specified on a grid line
     //
     // This vector is checked when the grid line is parsed - if any option strings are 
@@ -137,7 +172,7 @@ private:
     // think of a good reason to exclude options from the commandline, so I haven't 
     // implemented that functionality (though it wouldn't be too difficult to add it).
     //
-    // I coupld probably have done this using a different set of options in Boost for
+    // I could probably have done this using a different set of options in Boost for
     // the commandline and gridfile, but in the end I decided this way was actually
     // easier, cleaner, and gives us a bit more control.
 
@@ -185,6 +220,7 @@ private:
         "maximum-number-timestep-iterations",
         "mode",
 
+        "notes-hdrs",
         "number-of-systems",
 
         "output-container", "c",
@@ -209,7 +245,7 @@ private:
     // m_SSEOnly records option strings that apply to SSE only
     // m_BSEOnly records option strings that apply to BSE only
     //
-    // These vectors are checked when the commandline or grid line are parsed for
+    // These vectors are checked when the commandline or grid line is parsed for
     // ranges and sets.  Ranges and sets are played out, and stars/binaries evolved
     // based on the grid of options defined by any ranges and sets specified by the
     // user.
@@ -357,7 +393,7 @@ private:
     // m_RangeExcluded records option strings that apply to SSE only
     // m_SetExcluded records option strings that apply to BSE only
     //
[…]

[…]
