+++ src/grid.txt
@@ -1,0 +1,2 @@
+--log-classes cLaSS1 clAss1a cLAss2 cLASS3 --notes-hdrs class1 class2 --notes note1 note2 note3
+
