### 2021-10-26 6c6333e
initial changes to allow for stellar mergers to be printed as CEE output, some typo fixes
