### 2021-10-26 d4ef69b
initial changes to allow for stellar mergers to be printed as CEE output, some typo fixes
