### 2021-10-27 3bafe72
Enhancements and minor fixes: (a) Added ability for users to annotate log files via new program options '--notes-hdrs' and '--notes'.  See docs for details. (b) Added a shorthand notation for vector program options (e.g. annotations, log-classes, debug-classes).  See docs for details. (c) Added HDF5 support to Log::GetLogStandardRecord() (return value) and Log::LogStandardRecord() (input parameter).  This only matters     to SSE Supernovae file - for delayed writes.  The original implementation may have resulted in minor discrepanicies in SSE Supernovae     log records, (because of when the va
[…]
