+++ src/Log.h
@@ -231,6 +231,90 @@ using std::string;
  *
  * 
  * JR, January 2021
+ * 
+ * 
+ * 
+ * Annotations
+ * ===========
+ * 
+ * We have added functionality to allow users to annotate log files.  The original motivation for 
+ * annotation functionality was to enable users to describe the contents of custom grid files, but
+ * annotations can be used for any reason.  With the ability to annotate the log files, the users
+ * can indicate the origin of various input data - e.g. the user could indicate what IMF was used 
+ * to draw initial mass values, or what distribution was used to draw mass ratio (q), etc.  
+ * 
+ * Annotations are written to log files as columns of data (aka datasets in HDF5 files).
+ * Annotations are specified via program options, and so can be specified on the command line as
+ * well as in grid files.  Because annotations are written to log files as columns of data, we
+ * have provided functionality to provide headers for the columns, as well as the actual data
+ * for the columns (the annotations).  Two new program options are provided:
+ * 
+ * --notes-hdrs 
+ * Allows users to specify header strings for annotation columns.  Can only be specified on the
+ * command line.  Allows users to specify one or more annotation header strings: this is a vector
+ * program option.  Usage is:
+ * 
+ *      --notes-hdrs headerstr1 headerstr2 headerstr3 ... headerstrN
+ * 
+ * Note that header strings are separated by a space character.  There is no limit to the number
+ * of header strings specified, and the number specified defines the number of annotations allowed.
+ *
+ * 
+ * --notes
+ * Allows users to specify annotations.  Can be specified on the command line and in a grid file.
+ * Allows users to specify one or more annotation strings: this is a vector program option.  Usage
+ * is:
+ * 
+ *      --notes annotation1 annotation2 annotation3 ... annotationN
+ * 
+ * Note that annotation strings are separated by a space character.  The number of annotation
+ * strings is limited to the number of annotation header strings specified (via the --notes-hdrs
+ * program option).  If more annotation strings are specified than header strings, the excess
+ * annotation strings will be ignored (and a warning displayed).  Note that when using this notation
+ * all annotation strings must be provided: there is no mechanism to allow a default annotation
+ * using the fallback method for program options (to the command-line value, then to the COMPAS
+ * default) - leaving an annotation string blank would be ambiguous (as to which annotation string
+ * had been left blank), and specifying "" as an annotation string would be ambiguous (as to whether
+ * the use wanted the annotation string to default, or just be a blank string).
+ * 
+ * Because this notation could become awkward, and to allow for default annotations, a shorthand
+ * notation for vector program options has been provided (see notes in Options.h for details).
+ * 
+ * Usage using the shorthand notation is:
+ * 
+ *     --notes-hdrs [headerstr1,headerstr2,headerstr3,...,headerStrN]
+ * 
+ *     --notes [annotation1,annotation2,annotation3,...,annotationN]
+ * 
+ * Because the parameters are bounded by the brackets, and delimited by commas (and so are now
+ * positional), users can omit specific annotations:
+ * 
+ *     --notes [,,annotation3,,annotation5]
+ * 
+ * In the example above, annotations 1, 2, 4, and those beyond annotation 5 have been omitted.
+ * Annotations 1, 2 & 4 will default - if they are specified in this manner on a grid line they will
+ * default to the correspodning annotation specified on the command line; if they are specified in 
+ * this manner on the command line they will default to the COMPAS default annotation (the empty 
+ * string).  If the number of annotations expected, as defined by the number of annotation headers 
+ * specified via the --notes-hdrs program option is more than 5, then anotations beyond annotation 5
+ * (the last annotation actually specified by the user) will default in the same manner as described
+ * above.
+ * 
[…]
