+++ src/logDefs.txt
@@ -1,4 +1,0 @@
-bse_sysparms_rec -= { program_option::notes[2] }
-
-bse_sne_rec += {program_option::notes[1], program_option::notes[2]}
-
