+++ src/Options.h
@@ -25,11 +25,11 @@
 #include "Rand.h"
 #include "changelog.h"
 
-using std::string;
-using std::vector;
-using std::get;
-
 namespace po = boost::program_options;
+
+
+const std::string NOT_PROVIDED = std::to_string(255);
+
 
 // OPT_VALUE macro
 //
@@ -132,7 +132,8 @@ private:
 
 
     // m_ShorthandAllowed records option strings that may be specified using the
-    // shorthand notation described in Options::ExpandShorthandOptionValues().
+    // shorthand notation described in Options::ExpandShorthandOptionValues() and
+    // in Log.h.
     //
     // Furthermore, the vector records whether option values for such options can be
     // defaulted (i.e. some or all values need not be specified), and if so, what
@@ -154,12 +155,12 @@ private:
         // trying to keep entries alphabetical so easier to find specific entries
 
         // option name          default allowed     default string
-        { "debug-classes",      false,              "" },
+        { "debug-classes",      false,              "" },                       // don't allow defaults - we don't know how many classes to specify
 
-        { "log-classes",        false,              "" },
+        { "log-classes",        false,              "" },                       // don't allow defaults - we don't know how many classes to specify
 
-        { "notes",              true,               "" },
-        { "notes-hdrs",         false,              "" }
+        { "notes",              true,               "" },                       // allow defaults - number of notes is 0..#notes-hdrs
+        { "notes-hdrs",         false,              "" }                        // don't allow defaults - we don't know how many headers to specify
     };
 
 
@@ -183,7 +184,7 @@ private:
         "add-options-to-sysparms",
 
         "debug-level",
-        "debug_classes",
+        "debug-classes",
         "debug-to-file",
         "detailed-output",
 
@@ -424,7 +425,7 @@ private:
         "common-envelope-lambda-prescription",
         "common-envelope-mass-accretion-prescription",
 
-        "debug_classes",
+        "debug-classes",
         "debug-level",
         "debug-to-file",
         "detailed-output",
@@ -521,7 +522,7 @@ private:
 
         "add-options-to-sysparms",
 
-        "debug_classes",
+        "debug-classes",
         "debug-level",
         "debug-to-file",
         "detailed-output",
@@ -611,8 +612,8 @@ public:
 
             bool                                                m_EnableWarnings;                                               // Flag used to determine if warnings (via SHOW_WARN macros) should be displayed
 
-            vector<string>                                      m_Notes;                                                        // Notes contents - for user-defined annotations
-            vector<string>                                      m_NotesHdrs;                                                    // Notes header strings - for user-defined annotations
+            std::vector<std::string>                            m_Notes;                                                        // Notes contents - for user-defined annotations
+            std::vector<std::string>                            m_NotesHdrs;                                                    // Notes header strings - for user-defined annotations
 
 	        bool                                                m_BeBinaries;													// Flag if we want to print BeBinaries (main.cpp)
             bool                                                m_EvolvePulsars;                                                // Whether to evolve pulsars or not
@@ -750,10 +751,10 @@ public:
 	        double                                              m_MaximumNeutronStarMass;						                // Maximum mass of a neutron star allowed, set to default in StarTrack
 
             // Setup default output directory and desired output directory
-            string                                              m_OutputPathString;                                             // String to hold the output directory
+            std::string                                         m_OutputPathString;                                             // String to hold the output directory
             boost::filesystem::path                             m_DefaultOutputPath;                                            // Default output location
             boost::filesystem::path                             m_OutputPath;                                                   // Desired output location
-            string                                              m_OutputContainerName;                                          // Name of output container (directory)
+            std::string                                         m_OutputContainerName;                                          // Name of output container (directory)
 
             // Mass loss options
[…]

[…]
