+++ docs/readthedocs/docs/post-processing-hdf5-tools.rst
@@ -1,0 +1,14 @@
+COMPAS HDF5 tools
+=================
+
+COMPAS provides a number of tools to manipulate ``HDF5`` files.  They are:
+
+.. toctree::
+   :maxdepth: 1
+
+   Copying and concatenating: h5copy.py <post-processing-h5copy>
+   View summary, header details, and contents: h5view.py <post-processing-h5view>
+   Reduce and rewrite: rewrite_H5.py <./notebooks/H5/3_rewritingH5.ipynb>
+   ./notebooks/H5/1_creatingH5.ipynb
+   ./notebooks/H5/2_readingH5.ipynb
+
