### 2021-10-30 ca31e0a
Enhancements and minor fixes: (a) Added ability for users to annotate log files via new program options '--notes-hdrs' and '--notes'.  See docs for details. (b) Added a shorthand notation for vector program options (e.g. annotations, log-classes, debug-classes).  See docs for details. (c) Added '--notes-hdrs' and '--notes' to pythonSubmit.py (default = None for both) (d) Added HDF5 support to Log::GetLogStandardRecord() (return value) and Log::LogStandardRecord() (input parameter).  This only matters     to SSE Supernovae file - for delayed writes.  The original implementation may have resulte
[…]
