+++ src/Options.cpp
@@ -2461,63 +2461,81 @@ std::tuple<std::string, int, std::vector<std::string>> Options::ExpandShorthandO
 
     std::vector<std::string> strargs = {std::string(p_ArgStrings[0])};                                                      // new args vector - command name is arg[0]
 
+    std::string  argString  = "";                                                                                           // argument string
     std::string  optionName = "";                                                                                           // option name
-    for (size_t iArg = 0; iArg < (size_t)p_ArgCount; iArg++) {                                                              // for each arg string
+    for (size_t iArg = 1; iArg < (size_t)p_ArgCount; iArg++) {                                                              // for each arg string
 
-        if (iArg <= 1) continue;                                                                                            // step over the executable name
+        argString = p_ArgStrings[iArg];                                                                                     // the argument we're processing
 
-        optionName = p_ArgStrings[iArg - 1];                                                                                // get the option name for the argument we're processing
-        optionName = utils::ToLower(utils::trim(optionName));                                                               // downshift and trim whitespace
-        if (optionName[0] != '-') {                                                                                         // is it actually an option name?
-            strargs.push_back(std::string(p_ArgStrings[iArg - 1]));                                                         // no - just add it to the new vector (unadulterated)
-            if (iArg == (size_t)p_ArgCount - 1) strargs.push_back(std::string(p_ArgStrings[iArg]));                         // and add option value to the new vector if last arg
-        }
-        else {                                                                                                              // have option name
-            strargs.push_back(p_ArgStrings[iArg - 1]);                                                                      // add option name to the new vector
+        if (argString[0] == '-') {                                                                                          // is it actually an option name?
+            strargs.push_back(argString);                                                                                   // yes - add it to the new vector (unadulterated)
+
+            optionName = argString;                                                                                         // get the option name for the argument we'll be processing
+            optionName = utils::ToLower(utils::trim(optionName));                                                           // downshift and trim whitespace
             optionName.erase(0, optionName.find_first_not_of("-"));                                                         // remove the "-" or "--"
-
-            // check if option is on the shorthand allowed list, and process it if it is
-            // presumably invalid options won't be on the list...
-            auto it = std::find_if(m_ShorthandAllowed.begin(), m_ShorthandAllowed.end(), [&optionName](const SHORTHAND_ENTRY& e) { return std::get<0>(e) == optionName; });
-            if (it == m_ShorthandAllowed.end()) {                                                                           // option in shorthand allowed list?
-                if (iArg == (size_t)p_ArgCount - 1) strargs.push_back(std::string(p_ArgStrings[iArg]));                     // no - add option value to the new vector if last arg
+        }
+        else {                                                                                                              // argument is not an option name - process it
+            if (iArg == 1) {                                                                                                // first arg (i.e. no option name)?
+                strargs.push_back(argString);                                                                               // yes - add it to the new vector (unadulterated)
             }
-            else {                                                                                                          // yes - shorthand allowed - process it
-                if (p_ArgStrings[iArg] != nullptr) {                                                                        // null arg?
+            else {                                                                                                          // no - not first arg - process it
+                // check if option is on the shorthand allowed list, and process it if it is
+                // presumably invalid options won't be on the list...
+                auto it = std::find_if(m_ShorthandAllowed.begin(), m_ShorthandAllowed.end(), [&optionName](const SHORTHAND_ENTRY& e) { return std::get<0>(e) == optionName; });
+                if (it == m_ShorthandAllowed.end()) {                                                                       // option in shorthand allowed list?
+                    strargs.push_back(argString);                                                                           // no - add option value to the new vector
+                }
+                else {                                                                                                      // yes - shorthand allowed - process it
+                    if (!argString.empty()) {                                                                               // null arg?
[…]
