+++ docs/readthedocs/docs/developer-guide.rst
@@ -1,0 +1,24 @@
+Developer guide
+===============
+
+TeamCOMPAS welcomes the active involvement of colleagues and others interested in the ongoing development and improvement of
+the COMPAS software. We hope this Developer Guide helps anyone interested in contributing to the COMPAS software. We expect
+this guide to be a living document and improve along with the improvements made to the software.
+
+This guide covers the C++ COMPAS population synthesis code.  Post-processing code (Python scrpts) are not (yet) covered
+by this guide.
+
+
+.. toctree::
+   :maxdepth: 1
+
+   SSE
+   BSE
+   object-ids
+   services
+   floating-point-comparisons
+   constants-dot-h
+   changelog
+   programming-style
+   developer-building-compas
+   

[…]
