+++ docs/readthedocs/docs/programming-style-style-indentation.rst
@@ -1,9 +1,0 @@
-Indentation
-===========
-
-There is ongoing debate in the programming community as to whether indentation should be achieved using spaces or tabs (strange, but true...). 
-The use of spaces is more common. Unfortunately a mix of spaces and tabs doesn’t work well with some editors - we should settle on one method and 
-try to stick to it.  COMPAS (so far) has a mix of both – no matter which convention we choose, once there are instances of both in the code it's 
-hard to recover.  So while using tabs is probably preferred, use whatever is convenient (pragmatism is your friend...). 
-
-COMPAS (mostly) uses an indentation size of 4 spaces - again we should settle on a size and stick to it.
