+++ docs/readthedocs/docs/services-random-numbers-seed.rst
@@ -1,8 +1,0 @@
-Rand::Seed(p_Seed)
-==================
-
-::
-
-    UNSIGNED LONG INT Rand::Seed(const UNSIGNED LONG p_Seed)
-
-Sets the seed for the ``gsl`` random number generator to ``p_Seed``. The return value is the seed.
