+++ docs/readthedocs/docs/services-base-level-logging-func-squawk.rst
@@ -1,12 +1,0 @@
-Log::Squawk(p_SquawkStr)
-========================
-
-::
-
-    VOID Log::Squawk(
-        STRING p_SquawkStr  // the string to be written to stderr.
-    )
-
-Writes ``p_SquawkStr`` to stderr.
-
-This function does not return a value.
