+++ docs/readthedocs/docs/references.rst
@@ -1,6 +1,0 @@
-References
-==========
-
-.. bibliography::
-   :all:
-

[…]
