+++ docs/readthedocs/docs/standard-logfile-default-record-specifications-BSE-supernovae.rst
@@ -1,40 +1,0 @@
-BSE supernovae
-==============
-
-Default record definition for the BSE Supernovae log file::
-
-    const ANY_PROPERTY_VECTOR BSE_SUPERNOVAE_REC = {
-        BINARY_PROPERTY::RANDOM_SEED,
-        SUPERNOVA_PROPERTY::DRAWN_KICK_MAGNITUDE,
-        SUPERNOVA_PROPERTY::KICK_MAGNITUDE,
-        SUPERNOVA_PROPERTY::FALLBACK_FRACTION,
-        BINARY_PROPERTY::ORBITAL_VELOCITY_PRE_SUPERNOVA,
-        BINARY_PROPERTY::DIMENSIONLESS_KICK_MAGNITUDE, 
-        SUPERNOVA_PROPERTY::MEAN_ANOMALY,
-        SUPERNOVA_PROPERTY::SUPERNOVA_THETA,
-        SUPERNOVA_PROPERTY::SUPERNOVA_PHI,
-        SUPERNOVA_PROPERTY::SN_TYPE,
-        BINARY_PROPERTY::ECCENTRICITY_PRE_SUPERNOVA,  
-        BINARY_PROPERTY::ECCENTRICITY,
-        BINARY_PROPERTY::SEMI_MAJOR_AXIS_PRE_SUPERNOVA_RSOL,
-        BINARY_PROPERTY::SEMI_MAJOR_AXIS_RSOL,
-        BINARY_PROPERTY::TIME,
-        BINARY_PROPERTY::SUPERNOVA_STATE,
-        BINARY_PROPERTY::UNBOUND,
-        COMPANION_PROPERTY::STELLAR_TYPE,
-        SUPERNOVA_PROPERTY::STELLAR_TYPE,
-        SUPERNOVA_PROPERTY::STELLAR_TYPE_PREV,
-        COMPANION_PROPERTY::MASS,
-        SUPERNOVA_PROPERTY::TOTAL_MASS_AT_COMPACT_OBJECT_FORMATION,
-        SUPERNOVA_PROPERTY::CORE_MASS_AT_COMPACT_OBJECT_FORMATION,
-        SUPERNOVA_PROPERTY::CO_CORE_MASS_AT_COMPACT_OBJECT_FORMATION,
-        SUPERNOVA_PROPERTY::HE_CORE_MASS_AT_COMPACT_OBJECT_FORMATION,
-        SUPERNOVA_PROPERTY::MASS,
-        SUPERNOVA_PROPERTY::EXPERIENCED_RLOF,
-        SUPERNOVA_PROPERTY::MASS_TRANSFER_DONOR_HISTORY,
-        SUPERNOVA_PROPERTY::SPEED,
-        COMPANION_PROPERTY::SPEED,
-        BINARY_PROPERTY::SYSTEMIC_SPEED,
-        SUPERNOVA_PROPERTY::IS_HYDROGEN_POOR
-    };
-
