### 2021-10-30 TeamCOMPAS/COMPAS#664 @HeloiseS
TypeError when running pythonSubmitDemo.py 
_This issue is submitted as part of the JOSS review [#3838](https://github.com/openjournals/joss-reviews/issues/3838)_

### :bug: Description
I tried to run the demo and got the error below
```
>>> python pythonSubmitDemo.py

python_version = 3
compas_executable_override None
Traceback (most recent call last):
  File "pythonSubmitDemo.py", line 19, in <module>
    class pythonProgramOptions:
  File "pythonSubmitDemo.py", line 39, in pythonProgramOptions
    compas_executable = os.path.join(git_directory, 'src/COMPAS')
  File "/usr/lib/python3.8/posixpath.py", line 76, in join
    a = os.fspath(a)
TypeError: expected str, bytes or os.PathLike object, not NoneType
```

### :bug: Fixing
The bug occurs because I had not set the environment variable $COMPAS_ROOT_DIR and the demo file expects to find it.

**Requests and Comments**
-  You should say in the **tutorial** that at COMPAS run time the code expects an environment variable to be set. Now I went looking for the instruction, and sure enough I found it in "Running COMPAS via Python" - I am sure that many people, will, like myself, leave reading the 8 or so section of the user guide for later in their learning curve and start by trying to run the tutorial, if anything to see what happens. So I'd recommend making that first tutorial relatively self-contained.
-  Thanks for including a little how-to on setting environement variables. It sounds trivial but I knew no python and no linux when i started my PhD and that sort of stuff was a god-send in a sea of asssumed knowledge. Maybe link to it in the tutorial? You could say something to the effect of:

>  Note that the environment variable $COMPAS_ROOT_DIR needs to be set in order for this script to run properly - see page BLABLALA [link]

- using `assert` statements to check whether the variables that need to be set have been set properly would give a more useful error message for people that still missed the memo on setting  $COMPAS_ROOT_DIR (or anything else they need to do). e.g.
```
assert git_directory is not None, "It seems that the environment variable $COMPAS_ROOT_DIR as not been set."
```

**Label the issue**
`urgency_low`    `severity_minor`

**Versioning (please complete the following information):**
 - OS: Ubuntu 20.04
 - COMPAS - v2.21
