### 2021-10-30 TeamCOMPAS/COMPAS#665 @HeloiseS
/bin/sh: 1: Permission denied | pythonSubmitDemo.py
_This issue is submitted as part of the JOSS review [#3838](https://github.com/openjournals/joss-reviews/issues/3838)_

### :bug: Description and fix attempts
I don't seem to be able to run `pythonSubmitDemo.py` - I have now set my `$COMPAS_EXECUTABLE_PATH` (maybe i did it wrong?)

```
python_version = 3
compas_executable_override None
/home/fste075/COMPAS/src/COMPAS
[...]
/bin/sh: 1: /home/fste075/COMPAS/src/COMPAS: not found
```
I have tried setting my environement variable to `/home/fste075/` and removed the `src/` on line 39 of but i then get a `/bin/sh: 1: /home/fste075/COMPAS: Permission denied` error.  I performed an aggressive chmod 777 on the COMPAS directory, to no success.

**Label the issue**
`urgency_low`
`severity_minor`  ?

**Versioning (please complete the following information):**
 - OS: Ubunutu 20.03
 - COMPAS v02.21

### Further comment / question
On inspecting the code to check where `compas_executable` was being used i realised that there is no `__init__` method for that class(which isn't  :sparkles: _stricktly_ :sparkles:  necessary but good form)  and I am a little confused about the overall construction. For example, why is the following method not a class attribute instead or an @property ?
```
def stringCommands(self):
        stringCommands = [
            '--mode',
            '--case-BB-stability-prescription',
            '--chemically-homogeneous-evolution',
            '--luminous-blue-variable-prescription',
            '--mass-loss-prescription',
            '--mass-transfer-angular-momentum-loss-prescription',
            '--mass-transfer-accretion-efficiency-prescription',
            '--mass-transfer-rejuvenation-prescription',
            '--initial-mass-function',
            '--semi-major-axis-distribution',
            '--orbital-period-distribution',
            '--mass-ratio-distribution',
            '--eccentricity-distribution',
            '--metallicity-distribution',
            '--rotational-velocity-distribution',
            '--remnant-mass-prescription',
            '--fryer-supernova-engine',
            '--black-hole-kicks',
            '--kick-magnitude-distribution',
            '--kick-direction',
            '--output-path',
            '--output-container',
            '--common-envelope-lambda-prescription',
            '--stellar-zeta-prescription',
            '--mass-transfer-thermal-limit-accretor',
            '--pulsational-pair-instability-prescription',
            '--neutron-star-equation-of-state',
            '--pulsar-birth-magnetic-field-distribution',
            '--pulsar-birth-spin-period-distribution',
            '--common-envelope-mass-accretion-prescription',
            '--envelope-state-prescription',
            '--logfile-name-prefix',
            '--logfile-type',
            '--logfile-definitions',
            '--grid',
            '--logfile-common-envelopes',
            '--logfile-detailed-output',
            '--logfile-double-compact-objects',
            '--logfile-pulsar-evolution',
            '--logfile-rlof-parameters',
            '--logfile-supernovae',
            '--logfile-switch-log',
            '--logfile-system-parameters',
            '--neutrino-mass-loss-BH-formation'
        ]
   return stringCommands
```
