### 2021-10-30 TeamCOMPAS/COMPAS#665 @HeloiseS
### UPDATE

I found the problem haza! - my initial env variable was correct but **I missed another step** - the Build COMPAS step to be precise, which indeed is rather essential.

The demo runs great (haven't check outputs yet tho) - I have but one request based on this experience:
**Could you put the steps to install + demo instructions on 1 single page?** If the install recipe was a top-to-bottom read it would harder to forget a step.

From looking at your menu, it all looks in the right order (and all the instructions are clearly there) but for some reason your subsections are 1) individual pages i need to click through and 2) some of them are hidden (I guess it's based on the number of levels you choose to display at once). The 'Build COMPAS' was hidden under dependencies - and naively i thought that referred to the packages needed for the code to work (e.g. numpy), not something I **had to do** to complete the install -- so i didn't click that link until much later in the debugging process.

You could take all the "Getting started" material, stick the demo to end of it, and have it as one big Quick Start/Install recipe, at the end of which you can direct the n00bs to the User Guide once they've successfully run the default demo once (so they know the code works) to modify the demo and properly understand it :)
