### 2021-10-30 #3838 @HeloiseS
Hello all!

I made a start on the review - plenty left to do code-side. For now here are a few comments on the paper:

>  State of the field: Do the authors describe how this software compares to other commonly-used packages?

Some text should be added to the paper to address this crateria - now since the word limit won't let you do such a discussion justice AND there is the apj paper which does have a rather extensive list in the intro, is it worth citing the sister paper? I don't know how the ApJ/JOSS link works and whether I should consider the two paper as one and tick that box or ask you to add a sentence saying you expand on that in the ApJ publication. @christinahedges  can you advise?

Now for 2 very minor comments:
- stating that massive stars are born with similarly massive companions based on Sana et al. is very 2012 of you ;) Moe and Di Stefano 2017  has updated observations and I'm pretty sure it doesn't show that massive stars have an excess of systems with mass ratios ~1  (they have nifty distributions and table - see figure 2 and table 13).  It's saturday as i'm doing this so I'll  check that and amend as necessary; also the updated reference would be good :)
- "embarassingly" -> although i appreciate the emphasis the emotive language maybe doesn't belong here ;) pick another word or drop it?
