### 2021-10-30 #3838 @ilyamandel
Hi @HeloiseS ,

First of all, thank you very much for pointing out the documentation (and python submit script) improvements as issues on the COMPAS github page.  Much appreciated -- we'll aim to address them ASAP.

On your two "minor points":
1. Happy to add a citation to Moe & DiStefano (2017), of course, who actually include a specific twin subpopulation; however, we just meant the more modest statement that massive stars are more likely to have massive companions than if the companions were independently drawn from the IMF, rather than focus on details of the mass distribution.
2. See https://en.wikipedia.org/wiki/Embarrassingly_parallel .
