### 2021-10-30 08e4a15
Update BaseBinaryStar.cpp
