### 2021-10-31 #3838 @HeloiseS
### Further discussion on the `pythonSubmit.py` files

Okay so I think I have put my finger on why these files feels so odd:

1 - first, as i mentioned before there are some weird constructions:  for example methods are called to create list  when (at the moment) it seems you could just make the lists attributes without the need to call a function and return something. Another thing that caught my eye is the function `combineCommandLineOptionsDictIntoShellCommand` which exists as a separate entity and not a method of the class, when it seems like its only job is to operate on the things within the class.

2- The definition of the class itself isn't very pythonic (to me at least). I would expect that class to have the architecture to contain the necessary data and the tools to put it together in a command compas can read, but **not the specific numbers and options for a specific simulation**. It seems (from skimming through the folders) that you are redefining the class over and over again in each pythonSubmit.py file. What I would expect instead is for these options to be parsed to each **instances** of the class, for example with a yaml file. So basically my thinking is that COMPAS should come with a *unique* `pythonProgramOptions` class that can be imported, and it's the config files that the user would tweak and then give to the pipeline to run compas. For example:

```
from compas import pythonProgramOptions

myoptions = pythonProgramOptions(myconfig.yaml)
myoptions.call_shell_command() # if you want to call directly
myoptions.shell_command # if you want to retrieve the string and do somthing with it
```

now the question is... how much of a pain would that be regarding backwards compatibility? If you're curious to see what i would do with this class i can fork a suggestion for the class and yaml file (unless Jeff prefers JSON ;) )in the next couple of days for you to look at and see if it would work with the general workflow of compas.
