+++ docs/COMPAS_LaTeX/sections/RevisionHistory.tex
@@ -1,104 +1,0 @@
-\newpage\section{Revision History}\label{sec:RevisionHistory}
-
-\small
-
-\begin{tabularx}{\linewidth}{
-    |>{\hsize=0.8\hsize}X
-    |>{\hsize=0.3\hsize}X
-    |>{\hsize=2.1\hsize}X
-    |>{\hsize=0.8\hsize}X
-    |
-}
-
-\hline                                                                      % 
-
-\revisionHistoryRow{\textbf{Date}}{\textbf{Version}}{\textbf{Description}}{\textbf{Author}}
-
-\revisionHistoryRow{2012-2019}{}{Original COMPAS manual}{Team COMPAS}
-
-\revisionHistoryRow{1 September 2019}{0.1}{Initial draft}{Jeff Riley}
-
-\revisionHistoryRow{20 September 2019}{0.2}{Updated compilation requirements}{Jeff Riley}
-
-\revisionHistoryRow{1 October 2019}{0.3}{Added (minimal) CHE documentation}{Jeff Riley}
-
-\revisionHistoryRow{21 October 2019}{0.4}{Added Grids documentation}{}
-\revisionHistoryRow{}{}{Added Programming Style and Conventions}{}
-\revisionHistoryRow{}{}{Added Appendices A-E}{}
-\revisionHistoryRow{}{}{Reformatted for User Guide, Developer Guide etc.}{Jeff Riley}
-
-\revisionHistoryRow{18 December 2019}{0.5}{Updated Grids documentation for kick values}{Jeff Riley}
-
-\revisionHistoryRow{20 December 2019}{0.6}{Updated Appendices A, B, D, E}{Jeff Riley}
-
-\revisionHistoryRow{23 January 2020}{0.7}{Updated Grids documentation for kick values}{Jeff Riley}
-
-\revisionHistoryRow{31 March 2020}{1.0}{Updated for Beta release}{Jeff Riley}
-
-\revisionHistoryRow{22 April 2020}{1.1}{Updated to reflect pre2ndSN $\rightarrow$ preSN variable changes}{Simon Stevenson}
-
-\revisionHistoryRow{27 April 2020}{1.2}{Updated to reflect SSE Grid file changes}{Jeff Riley}
-
-\revisionHistoryRow{7 September 2020}{2.0}{Created LaTex version}{Jeff Riley}
-
-\revisionHistoryRow{9 September 2020}{2.1}{Added SSE \& BSE Switch Log files documentation}{}
-
-\revisionHistoryRow{}{}{Added SSE Supernova log file documentation}{Jeff Riley}
-
-\revisionHistoryRow{11 September 2020}{2.2}{Added description for RLOF output parameters}{Mike Lau}
-
-\revisionHistoryRow{3 November 2020}{2.3}{Updated for new grid file functionality}{Jeff Riley}
-
-\revisionHistoryRow{4 November 2020}{2.4}{Updated for changed switchlog functionality}{Jeff Riley}
-
-\revisionHistoryRow{10 November 2020}{2.5}{Updated for added SSE System Parameters file}{}
-
-\revisionHistoryRow{}{}{Clean up/fix for past code changes}{Jeff Riley}
-
-\revisionHistoryRow{10 November 2020}{2.6}{Added description of Schneider remnant mass and MassTransferDonorHistory}{Reinhold Willcox}
-
-\revisionHistoryRow{13 November 2020}{2.7}{Added description of metallicity-distribution and related options}{}
-\revisionHistoryRow{}{}{Removed references to AIS; minor fixes and cleanup}{Jeff Riley}
-
-\revisionHistoryRow{19 November 2020}{2.8}{Added description of option \mbox{luminous-blue-variable-prescription}.}{}
[…]
