+++ docs/COMPAS_LaTeX/sections/DeveloperGuide/constants_h.tex
@@ -1,94 +1,0 @@
-\subsection{Constants File -- constants.h}\label{sec:ConstantsFile}
-
-As well as plain constant values, many distribution and prescription identifiers are declared in constants.h.  These are mostly declared as enum classes, with each enum class having a corresponding map of labels.  The benefit is that the values of a particular (e.g.) prescription are limited to the values declared in the enum class, rather than any integer value, so the compiler will complain if an incorrect value is inadvertently used to reference that prescription.
-
-For example, the Common Envelope Accretion Prescriptions are declared in constants.h thus:
-
-\bigskip
-
-\hfill
-\begin{minipage}{\dimexpr\textwidth-2em}
-    enum class CE\_ACCRETION\_PRESCRIPTION: int \lcb \\
-    \tabto{2em}ZERO, CONSTANT, UNIFORM, MACLEOD \\
-    \rcb{;}
-
-    \medskip
-    const std::unordered\_map$<$CE\_ACCRETION\_PRESCRIPTION, std::string$>$ \\ CE\_ACCRETION\_PRESCRIPTION\_LABEL\ =\ \lcb \\
-    \tabto{2em}\lcb\ CE\_ACCRETION\_PRESCRIPTION::ZERO,\tabto{23.5em}"ZERO"\ \rcb, \\
-    \tabto{2em}\lcb\ CE\_ACCRETION\_PRESCRIPTION::CONSTANT,\tabto{23.5em}"CONSTANT"\ \rcb, \\
-    \tabto{2em}\lcb\ CE\_ACCRETION\_PRESCRIPTION::UNIFORM,\tabto{23.5em}"UNIFORM"\ \rcb, \\
-    \tabto{2em}\lcb\ CE\_ACCRETION\_PRESCRIPTION::MACLEOD,\tabto{23.5em}"MACLEOD"\ \rcb, \\
-    \rcb{;}
-\end{minipage}
-
-Note that the values allowed for variables of type CE\_ACCRETION\_PRESCRIPTION are limited to ZERO, CONSTANT, UNIFORM, and MACLEOD -- anything else will cause a compilation error.
-
-% Why do I need the \linebreaks in this paragraph?  
-% If I remove them I get 'overfull \hbox' and the line extends past the right margin...
-% (now I get 'underfull \hbox (badness xxxx)' - can't win)
-% I had to reword some sentences to get word spacing to look ok (only ok) - the wording
-% is now a little clumsy, but it at least looks halfway decent.
-% Maybe somebody smarter than me can fix this... JR
[…]
