### 2021-10-31 fa0a70c
Adding online docs in preparation for move to compas.readthedocs.io
