+++ docs/COMPAS_LaTeX/sections/UserGuide/COMPASOutput.tex
@@ -1,245 +1,0 @@
-\newpage
-\subsection{COMPAS Output}\label{sec:COMPASOutput}
-
-Summary and status information during the evolution of stars is written to stdout; how much is written depends upon the value of the \textit{\texttt{-{}-}quiet} program option.  
-
-Detailed information is written to log files (described below).  All COMPAS output files are created inside a container directory, specified by the \textit{\texttt{-{}-}output-container} program option.
-
-If Detailed Output log files (see the \textit{\texttt{-{}-}detailed-output} program option) are created, they will be created inside a containing directory named `Detailed\_Output' within the COMPAS output container directory.
-
-Also created in the COMPAS container directory is a file named `Run\_Details' in which COMPAS records some details of the run (COMPAS version, start time, program option values etc.). Note that the option values recorded in the Run details file are the values specified on the commandline, not the values specified in a grid file (if used).
-
-COMPAS defines several standard log files that may be produced depending upon the simulation type (Single Star Evolution (SSE), or Binary Star Evolution (BSE), and the value of various program options. The standard log files are:
-
-\begin{itemize}
-\itemsep0pt
-\item{Detailed Output log file \\ Records detailed information for a star, or binary star, during evolution. Enable with program option \textit{\texttt{-{}-{detailed-output}}}.}
-\item{SwitchLog log file \\ Records detailed information for all stars, or binary stars, at the time of each stellar type switch during evolution. Enable with program option \textit{\texttt{-{}-{switch-log}}}.}
-\item{Supernovae log file \\ Records summary information for all stars that experience a SN event during evolution.}
-\item{System Parameters log file \\ Records summary information for all stars, or binary stars, during evolution.}
-\item{Double Compact Objects log file \\ Records summary information for all binary systems that form DCOs during BSE.}
-\item{Common Envelopes log file \\ Records summary information for all binary systems that experience CEEs during BSE.}
-\item{Pulsar Evolution log file \\ Records detailed Pulsar evolution information during BSE.}
-\item{RLOF file \\ Records detailed information RLOF events during BSE. Enable with program option \textit{\texttt{-{}-{rlof-printing}}}.}
-\end{itemize}
-
-\newpage
-\subsubsection{Standard Log File Record Specifiers}\label{sec:StandardLogFileRecordSpecifiers}
-
-Each standard log file has an associated log file record specifier that defines what data are to be written to the log files. Each record specifier is a list of known properties that are to be written as the log record for the log file associated with the record specifier. Default record specifiers for each of the standard log files are shown in Appendix~\crossref{sec:DefaultLogFileRecordSpecs}. The standard log file record specifiers can be defined by the user at run-time (see Section~\crossref{sec:StandardLogFileRecordSpecification} below).
-
[…]
