+++ docs/online-docs/pages/Developer guide/Services/Logging and debugging/Base-level logging/services-base-level-logging-func-squawk.rst
@@ -1,0 +1,12 @@
+Log::Squawk(p_SquawkStr)
+========================
+
+::
+
+    VOID Log::Squawk(
+        STRING p_SquawkStr  // the string to be written to stderr.
+    )
+
+Writes ``p_SquawkStr`` to stderr.
+
+This function does not return a value.
