+++ docs/online-docs/pages/User guide/COMPAS output/standard-logfiles-default-record-specifications-BSE-pulsars.rst
@@ -1,0 +1,24 @@
+BSE pulsar evolution
+====================
+
+Default record definition for the BSE Pulsar Evolution log file::
+
+    const ANY_PROPERTY_VECTOR BSE_PULSAR_EVOLUTION_REC = {
+        BINARY_PROPERTY::RANDOM_SEED,
+        STAR_1_PROPERTY::MASS,
+        STAR_2_PROPERTY::MASS,
+        STAR_1_PROPERTY::STELLAR_TYPE,
+        STAR_2_PROPERTY::STELLAR_TYPE,
+        BINARY_PROPERTY::SEMI_MAJOR_AXIS_RSOL,
+        BINARY_PROPERTY::MASS_TRANSFER_TRACKER_HISTORY,
+        STAR_1_PROPERTY::PULSAR_MAGNETIC_FIELD,
+        STAR_2_PROPERTY::PULSAR_MAGNETIC_FIELD,
+        STAR_1_PROPERTY::PULSAR_SPIN_FREQUENCY,
+        STAR_2_PROPERTY::PULSAR_SPIN_FREQUENCY,
+        STAR_1_PROPERTY::PULSAR_SPIN_DOWN_RATE,
+        STAR_2_PROPERTY::PULSAR_SPIN_DOWN_RATE,
+        BINARY_PROPERTY::TIME,
+        BINARY_PROPERTY::DT
+    };
+
+    

[…]
