### 2021-10-31 7786fe1
Updates to docs for issues #664 and #665
