### 2021-10-31 aab9eaf
Updates to docs for issues #664 and #665
