### 2021-10-31 fe14eaf
Add checks per issue #664
