### 2021-10-31 884823a
Add checks per issue #664
