### 2021-10-31 e728c8e
Updated pythonSubmit.py per issue #664
