### 2021-10-31 bd59b15
Updated pythonSubmit.py per issue #664
