### 2021-10-31 TeamCOMPAS/COMPAS#664 @jeffriley
Thanks @HeloiseS.

I've added text to the tutorial page to ensure the environment variable is set.

I've update pythonSubmit.py to do some checks on the COMPAS executable (see PR #671).
