### 2021-10-31 TeamCOMPAS/COMPAS#665 @jeffriley
Thanks @HeloiseS.

> Could you put the steps to install + demo instructions on 1 single page?
> If the install recipe was a top-to-bottom read it would harder to forget a step.

We have different install instructions for different OSes, so putting the install instructions + demo on a single page would mean we either have install instructions for (say) Linux and MacOS on the same page, or we have different OS install pages and repeat the common content (e.g. the demo).  Not sure what's best there.  These online docs are fairly new, and we're still to go through them as a group and figure out the best presentation - but we will do that.

In the meantime I've brought the "Building COMPAS" page up a level - you're right, it shouldn't be buried in the dependencies - and I've added some text and links at the bottom of the "Getting started" page so users can jump straigh to the tutorial.

> On inspecting the code to check where compas_executable was being used i realised that there is no __init__ method
> for that class(which isn't ✨ stricktly ✨ necessary but good form) and I am a little confused about the overall
> construction. For example, why is the following method not a class attribute instead or an @Property ?

I don't know :-)  I don't actually use pythonSubmit.py very much - I usually run COMPAS from the command line.  pythonSubmit.py was written a long time ago (maybe by @SimonStevenson?), so someone more familiar with it might be able to tell you.

pythonSubmit.py could do with a tidy-up - I think it has grown over the years.  I could tidy-up the code, but I think it's better that someone who uses it more than I do should look at the structure - maybe we could reorder the variables/options so they're a bit easier to find.  I'll raise the issue with the group.
