+++ docs/online-docs/COMPAS-2021methodsPaper.bib
@@ -1,4 +1,4 @@
-@ This formats the citation for the COMPAS Methods Paper in our prefrred style
+@ This formats the citation for the COMPAS Methods Paper in our preferred style
 @ The following PREAMBLE entry must appear at the top of the bibtex file
 
 @PREAMBLE{ {\providecommand{\noopsort}[1]{}} }
