### 2021-10-31 TeamCOMPAS/COMPAS#664 @ilyamandel
Closing this issue.  Thanks again to @jeffriley for the fix and to @HeloiseS for raising it!
