### 2021-10-31 b3b47f4
Adding Moe+DiStefano citation
