+++ paper.md
@@ -101,7 +101,7 @@ aas-journal: Astrophysical Journal Supplements <- The name of the AAS journal.
 
 # Summary
 
-Most massive stars---those with initial masses greater than 8 $M_\odot$---are born with another similarly massive star as a companion [@Sana:2012Sci]. Massive binary stars are responsible for producing many exotic astrophysical phenomena, such as the observed diversity of supernovae, binary pulsars, X-ray binaries and merging compact objects. The latter are now regularly observed by the ground-based gravitational wave observatories Advanced LIGO and Virgo [@abbott2016observation;@GWTC2].  Population models of massive binary evolution make it possible to interpret existing observations and to make predictions for future observing campaigns.  
+Most massive stars---those with initial masses greater than 8 $M_\odot$---are born with another massive star as a companion [@Sana:2012Sci;@Moe:2017ApJS]. Massive binary stars are responsible for producing many exotic astrophysical phenomena, such as the observed diversity of supernovae, binary pulsars, X-ray binaries and merging compact objects. The latter are now regularly observed by the ground-based gravitational wave observatories Advanced LIGO and Virgo [@abbott2016observation;@GWTC2].  Population models of massive binary evolution make it possible to interpret existing observations and to make predictions for future observing campaigns.  
 
 # Statement of need
 
