### 2021-10-31 TeamCOMPAS/COMPAS#665 @ilyamandel
The instructions at https://compas-docs-readthedocs.readthedocs.io/en/latest/example-compas-run.html have been clarified, with "This tutorial assumes that you have already built the COMPAS executable as described in [Building COMPAS](https://compas-docs-readthedocs.readthedocs.io/en/latest/building-COMPAS.html)." added as a first sentence.
