### 2021-10-31 #3838 @ilyamandel
Hi @HeloiseS ,
Thank you very much for all of your suggestions so far.

A quick update:
1. We've updated the documentation to clarify some of the workflow issues you raised.
2. We included assert statements in pythonSubmit.py to make sure that the environment variables are set and the executable file exists.
3. We updated the awkward sentence about companion masses to "Most massive stars---those with initial masses greater than 8 $M_\odot$---are born with another massive star as a companion [@Sana:2012Sci;@Moe:2017ApJS]. "

We are grateful for your suggestion regarding an overhaul of pythonSubmit.py, which we agree is perhaps not optimal.  It is just meant as a simple and convenient way to update the COMPAS command line arguments -- so, indeed, the idea is that users would edit pythonSubmit itself, effectively treating it as an inlist text file.  Note that it is not necessary at all to use pythonSubmit to run COMPAS -- one could, say, store the desired command line in a text file.  So our feeling is that it's not the most urgent development priority at the moment.
