### 2021-11-01 #3838 @HeloiseS
Definitely not urgent - but I'd like to get the feeling of @katiebreivik and how she found handling the COMPAS calls. If you say it's intended for users to make their own python utility then it's worth noting that it was time consuming and fiddly to put the example into my own code since the config is embedded into a class and it took a while to get the reformatting to work properly without errors since there are so many parameters.

Let's put a pin in that for now, but i'd like to discuss that further later on :)
