### 2021-11-01 TeamCOMPAS/COMPAS#665 @HeloiseS
@jeffriley THANKS!

> pythonSubmit.py could do with a tidy-up - I think it has grown over the years. I could tidy-up the code, but I think it's better that someone who uses it more than I do should look at the structure - maybe we could reorder the variables/options so they're a bit easier to find. I'll raise the issue with the group.

I know @ilyamandel  is a little reluctant but i literally have it tidied up already so i can just send you my version and you can take a look - it would take you no time at all :)
