+++ src/Options.h
@@ -166,7 +166,6 @@ private:
         "log-level", 
         "log-classes",
 
-        // Serena
         //"logfile-be-binaries",
 
         "logfile-common-envelopes",
@@ -258,7 +257,6 @@ private:
         "allow-touching-at-birth",
         "angular-momentum-conservation-during-circularisation", 
 
-        // Serena
         //"be-binaries",
 
         "case-BB-stability-prescription",
@@ -276,7 +274,6 @@ private:
         "common-envelope-recombination-energy-density",
         "common-envelope-slope-kruckow",
 
-        // AVG
         /*
         "critical-mass-ratio-giant-degenerate-accretor",
         "critical-mass-ratio-giant-non-degenerate-accretor",
@@ -375,7 +372,6 @@ private:
         "allow-touching-at-birth",
         "angular-momentum-conservation-during-circularisation",
 
-        // Serena
         //"be-binaries",
 
         "black-hole-kicks",
@@ -418,7 +414,6 @@ private:
         "log-level", 
         "log-classes",
 
-        // Serena
         //"logfile-be-binaries",
 
         "logfile-common-envelopes",
@@ -502,7 +497,6 @@ private:
         "log-classes",
         "log-level", 
 
-        // Serena
         //"logfile-be-binaries",
 
         "logfile-common-envelopes",
@@ -788,6 +782,8 @@ public:
             double                                              m_CommonEnvelopeAlphaThermal;                                   // lambda = alpha_th*lambda_b + (1-alpha_th)*lambda_g
             double                                              m_CommonEnvelopeLambdaMultiplier;                               // Multiply common envelope lambda by some constant
             bool                                                m_AllowMainSequenceStarToSurviveCommonEnvelope;                 // Whether or not to allow a main sequence star to survive a common envelope event
+            bool                                                m_AllowRadiativeEnvelopeStarToSurviveCommonEnvelope;            // Whether or not to allow a radiative-envelope star to survive a common envelope event
+            bool                                                m_AllowImmediateRLOFpostCEToSurviveCommonEnvelope;              // Whether or not to allow Roche Lobe Overflow immediately after a CE to survive a common envelope event
     
             // Prescription for envelope state (radiative or convective)
             ENUM_OPT<ENVELOPE_STATE_PRESCRIPTION>               m_EnvelopeStatePrescription;
@@ -1047,11 +1043,13 @@ public:
     ADD_OPTIONS_TO_SYSPARMS                     AddOptionsToSysParms() const                                            { return m_CmdLine.optionValues.m_AddOptionsToSysParms.type; }
 
     bool                                        AllowMainSequenceStarToSurviveCommonEnvelope() const                    { return OPT_VALUE("common-envelope-allow-main-sequence-survive", m_AllowMainSequenceStarToSurviveCommonEnvelope, true); }
+    bool                                        AllowRadiativeEnvelopeStarToSurviveCommonEnvelope() const               { return OPT_VALUE("common-envelope-allow-radiative-envelope-survive", m_AllowRadiativeEnvelopeStarToSurviveCommonEnvelope, true); }
+    bool                                        AllowImmediateRLOFpostCEToSurviveCommonEnvelope() const                 { return OPT_VALUE("common-envelope-allow-immediate-rlof-post-ce-survive", m_AllowImmediateRLOFpostCEToSurviveCommonEnvelope, true); }
+    
     bool                                        AllowRLOFAtBirth() const                                                { return OPT_VALUE("allow-rlof-at-birth", m_AllowRLOFAtBirth, true); }
     bool                                        AllowTouchingAtBirth() const                                            { return OPT_VALUE("allow-touching-at-birth", m_AllowTouchingAtBirth, true); }
[…]
