### 2021-11-01 27a03cc
Update online documentation to match existing; add README for online docs; add online docs _build dir to .gitignore
