### 2021-11-01 f3ffdae
tweak to detailed_evol_plotter.py
