### 2021-11-01 6fe1e84
Update detailed_evol_plotter.py
