### 2021-11-01 e3c37ed
Minor fixes (a) reinstated _n suffix for BSE detailed filenames (inadvertently removed in v02.25.00) (b) updated pythonSubmit files: - preProcessing/pythonSubmit.py - examples/methods_paper_plots/detailed_evolution/pythonSubmitDemo.py - examples/methods_paper_plots/chirpmass_distribution/pythonSubmit.py - examples/methods_paper_plots/fig_5_HR_diagram/pythonSubmit.py - examples/methods_paper_plots/fig_6_max_R/pythonSubmit.py - examples/methods_paper_plots/fig_8_initial_core_final_mass_relations/pythonSubmitDefaults.py - examples/methods_paper_plots/fig_8_initial_core_final_mass_relations/python
[…]
