### 2021-11-01 #3838 @christinahedges
@HeloiseS and @ilyamandel, in response to your question, I believe that 1) the paper does need to discuss the state of the art (including other similar packages, and how this tool compares) 2) there should be enough room in the paper currently to add this discussion, it does not need to have extreme detail, but it should be present in the JOSS submission (not just the AAS submission).

I hope this helps clarify the question!
