### 2021-11-03 TeamCOMPAS/COMPAS#680 @HeloiseS
DOCKER - Logfile record specifications error &  command typos in the example pythonSubmit files
_This issue is submitted as part of the JOSS review [#3838](https://github.com/openjournals/joss-reviews/issues/3838)_

## :bug:  description

### The typo
so I was testing running compas in a docker image (it all installed beautifully, very easy) and i got the following error [Abridged] `Commandline Options error: unrecognised option '--common-envelope-allow-radiative-envelope-surive'`

I used one of your original pythonSubmit files to run this (the one in examples/\*/chirpmass_distribution), and when i go look in the file it does seem that the commands and the python variables associated with `--common-envelope-allow-radiative-envelope-survive` all have that typo in them (surive instead of survive). Same for the pythonSubmitDemo from looking at the github.

### The log file
After correcting the typo in the command I ran it again and got the following [Abridged]
```
Logfile record specifications error
File does not exist: /app/COMPAS/COMPAS_Output_Definitions.txt
```
I couldn't figure out why it wasn't happy, the COMPAS_Output_Definitions.txt is in the directory where i got the pythonSubmit file :shrug:  - I'll need your help to fix this one :)

### The Command I ran
```
sudo docker run
--rm
-it
-v $COMPAS_ROOT_DIR/compas-logs:/app/COMPAS/logs
-v $COMPAS_ROOT_DIR/examples/methods_paper_plots/chirpmass_distribution/pythonSubmit.py:/app/starts/pythonSubmit.py
-e COMPAS_EXECUTABLE_PATH=/app/COMPAS/bin/COMPAS
-e COMPAS_LOGS_OUTPUT_DIR_PATH=/app/COMPAS/logs
teamcompas/compas
python3 /app/starts/pythonSubmit.py
```

## Full outputs

### Full Output Bug 1
```
python_version = 3
/app/COMPAS/bin/COMPAS --enable-warnings False --use-mass-loss  --mass-transfer  --detailed-output False --evolve-unbound-systems False --population-data-printing False --rlof-printing  --circularise-binary-during-mass-transfer False --angular-momentum-conservation-during-circularisation False --pair-instability-supernovae  --pulsational-pair-instability  --quiet False --common-envelope-allow-main-sequence-survive  --common-envelope-allow-radiative-envelope-surive False --common-envelope-allow-immediate-rlof-post-ce-survive False --evolve-pulsars False --debug-to-file False --errors-to-file False --allow-rlof-at-birth  --allow-touching-at-birth False --switch-log False --check-photon-tiring-limit False --number-of-systems 100 --metallicity 0.0142 --common-envelope-alpha 1.0 --common-envelope-lambda 0.1 --common-envelope-slope-kruckow -0.8333333333333334 --common-envelope-alpha-thermal 1.0 --common-envelope-lambda-multiplier 1.0 --luminous-blue-variable-multiplier 1.5 --overall-wind-mass-loss-multiplier 1.0 --wolf-rayet-multiplier 1.0 --cool-wind-mass-loss-multiplier 1.0 --mass-transfer-fa 0.5 --mass-transfer-jloss 1.0 --maximum-evolution-time 13700.0 --maximum-number-timestep-iterations 99999 --timestep-multiplier 1.0 --initial-mass-min 10.0 --initial-mass-max 150.0 --initial-mass-power 0.0 --semi-major-axis-min 0.01 --semi-major-axis-max 1000.0 --mass-ratio-min 0.01 --mass-ratio-max 1.0 --minimum-secondary-mass 0.1 --eccentricity-min 0.0 --eccentricity-max 1.0 --metallicity-min 0.0001 --metallicity-max 0.03 --pulsar-birth-magnetic-field-distribution-min 11.0 --pulsar-birth-magnetic-field-distribution-max 13.0 --pulsar-birth-spin-period-distribution-min 10.0 --pulsar-birth-spin-period-distribution-max 100.0 --pulsar-magnetic-field-decay-timescale 1000.0 --pulsar-magnetic-field-decay-massscale 0.025 --pulsar-minimum-magnetic-field 8.0 --orbital-period-min 1.1 --orbital-period-max 1000 --kick-magnitude-sigma-CCSN-NS 265.0 --kick-magnitude-sigma-CCSN-BH 265.0 --fix-dimensionless-kick-magnitude -1 --kick-direction-power 0.0 --random-seed 0 --mass-transfer-thermal-limit-C 10.0 --eddington-accretion-factor 1 --pisn-lower-limit 60.0 --pisn-upper-limit 135.0 --ppi-lower-limit 35.0 --ppi-upper-limit 60.0 --maximum-neutron-star-mass 2.5 --kick-magnitude-sigma-ECSN 30.0 --kick-magnitude-sigma-USSN 30.0 --kick-scaling-factor 1.0 --maximum-mass-donor-nandez-ivanova 2.0 --common-envelope-recombination-energy-density 15000000000000.0 --common-envelope-mass-accretion-max 0.1 --common-envelope-mass-accretion-min 0.04 --zeta-main-sequence 2.0 --zeta-radiative-envelope-giant 6.5 --kick-magnitude-max -1.0 --muller-mandel-kick-multiplier-BH 200.0 --muller-mandel-kick-multiplier-NS 400.0 --log-level 0 --debug-level 0 --hdf5-chunk-size 100000 --hdf5-buffer-size 1 --neutrino-mass-loss-BH-formation-value 0.1 --mode BSE --case-BB-stability-prescription ALWAYS_STABLE --chemically-homogeneous-evolution PESSIMISTIC --luminous-blue-variable-prescription HURLEY_ADD --mass-loss-prescription VINK --mass-transfer-angular-momentum-loss-prescription ISOTROPIC --mass-transfer-accretion-efficiency-prescription THERMAL --mass-transfer-rejuvenation-prescription STARTRACK --initial-mass-function KROUPA --semi-major-axis-distribution FLATINLOG --orbital-period-distribution FLATINLOG --mass-ratio-distribution FLAT --eccentricity-distribution ZERO --metallicity-distribution ZSOLAR --rotational-velocity-distribution ZERO --remnant-mass-prescription FRYER2012 --fryer-supernova-engine DELAYED --black-hole-kicks FALLBACK --kick-magnitude-distribution MAXWELLIAN --kick-direction ISOTROPIC --output-path /app/COMPAS/logs --common-envelope-lambda-prescription LAMBDA_NANJING --stellar-zeta-prescription SOBERMAN --mass-transfer-thermal-limit-accretor CFACTOR --pulsational-pair-instability-prescription FARMER --neutron-star-equation-of-state SSE --pulsar-birth-magnetic-field-distribution ZERO --pulsar-birth-spin-period-distribution ZERO --common-envelope-mass-accretion-prescription ZERO --envelope-state-prescription LEGACY --logfile-type HDF5 --logfile-definitions /app/COMPAS/COMPAS_Output_Definitions.txt --neutrino-mass-loss-BH-formation FIXED_MASS --add-options-to-sysparms GRID

Commandline Options error: unrecognised option '--common-envelope-allow-radiative-envelope-surive'
Use option '-h' (or '--help') to see (descriptions of) available options
```

### Full Output bug 2

```
python_version = 3
/app/COMPAS/bin/COMPAS --enable-warnings False --use-mass-loss  --mass-transfer  --detailed-output False --evolve-unbound-systems False --population-data-printing False --rlof-printing  --circularise-binary-during-mass-transfer False --angular-momentum-conservation-during-circularisation False --pair-instability-supernovae  --pulsational-pair-instability  --quiet False --common-envelope-allow-main-sequence-survive  --common-envelope-allow-radiative-envelope-survive False --common-envelope-allow-immediate-rlof-post-ce-survive False --evolve-pulsars False --debug-to-file False --errors-to-file False --allow-rlof-at-birth  --allow-touching-at-birth False --switch-log False --check-photon-tiring-limit False --number-of-systems 100 --metallicity 0.0142 --common-envelope-alpha 1.0 --common-envelope-lambda 0.1 --common-envelope-slope-kruckow -0.8333333333333334 --common-envelope-alpha-thermal 1.0 --common-envelope-lambda-multiplier 1.0 --luminous-blue-variable-multiplier 1.5 --overall-wind-mass-loss-multiplier 1.0 --wolf-rayet-multiplier 1.0 --cool-wind-mass-loss-multiplier 1.0 --mass-transfer-fa 0.5 --mass-transfer-jloss 1.0 --maximum-evolution-time 13700.0 --maximum-number-timestep-iterations 99999 --timestep-multiplier 1.0 --initial-mass-min 10.0 --initial-mass-max 150.0 --initial-mass-power 0.0 --semi-major-axis-min 0.01 --semi-major-axis-max 1000.0 --mass-ratio-min 0.01 --mass-ratio-max 1.0 --minimum-secondary-mass 0.1 --eccentricity-min 0.0 --eccentricity-max 1.0 --metallicity-min 0.0001 --metallicity-max 0.03 --pulsar-birth-magnetic-field-distribution-min 11.0 --pulsar-birth-magnetic-field-distribution-max 13.0 --pulsar-birth-spin-period-distribution-min 10.0 --pulsar-birth-spin-period-distribution-max 100.0 --pulsar-magnetic-field-decay-timescale 1000.0 --pulsar-magnetic-field-decay-massscale 0.025 --pulsar-minimum-magnetic-field 8.0 --orbital-period-min 1.1 --orbital-period-max 1000 --kick-magnitude-sigma-CCSN-NS 265.0 --kick-magnitude-sigma-CCSN-BH 265.0 --fix-dimensionless-kick-magnitude -1 --kick-direction-power 0.0 --random-seed 0 --mass-transfer-thermal-limit-C 10.0 --eddington-accretion-factor 1 --pisn-lower-limit 60.0 --pisn-upper-limit 135.0 --ppi-lower-limit 35.0 --ppi-upper-limit 60.0 --maximum-neutron-star-mass 2.5 --kick-magnitude-sigma-ECSN 30.0 --kick-magnitude-sigma-USSN 30.0 --kick-scaling-factor 1.0 --maximum-mass-donor-nandez-ivanova 2.0 --common-envelope-recombination-energy-density 15000000000000.0 --common-envelope-mass-accretion-max 0.1 --common-envelope-mass-accretion-min 0.04 --zeta-main-sequence 2.0 --zeta-radiative-envelope-giant 6.5 --kick-magnitude-max -1.0 --muller-mandel-kick-multiplier-BH 200.0 --muller-mandel-kick-multiplier-NS 400.0 --log-level 0 --debug-level 0 --hdf5-chunk-size 100000 --hdf5-buffer-size 1 --neutrino-mass-loss-BH-formation-value 0.1 --mode BSE --case-BB-stability-prescription ALWAYS_STABLE --chemically-homogeneous-evolution PESSIMISTIC --luminous-blue-variable-prescription HURLEY_ADD --mass-loss-prescription VINK --mass-transfer-angular-momentum-loss-prescription ISOTROPIC --mass-transfer-accretion-efficiency-prescription THERMAL --mass-transfer-rejuvenation-prescription STARTRACK --initial-mass-function KROUPA --semi-major-axis-distribution FLATINLOG --orbital-period-distribution FLATINLOG --mass-ratio-distribution FLAT --eccentricity-distribution ZERO --metallicity-distribution ZSOLAR --rotational-velocity-distribution ZERO --remnant-mass-prescription FRYER2012 --fryer-supernova-engine DELAYED --black-hole-kicks FALLBACK --kick-magnitude-distribution MAXWELLIAN --kick-direction ISOTROPIC --output-path /app/COMPAS/logs --common-envelope-lambda-prescription LAMBDA_NANJING --stellar-zeta-prescription SOBERMAN --mass-transfer-thermal-limit-accretor CFACTOR --pulsational-pair-instability-prescription FARMER --neutron-star-equation-of-state SSE --pulsar-birth-magnetic-field-distribution ZERO --pulsar-birth-spin-period-distribution ZERO --common-envelope-mass-accretion-prescription ZERO --envelope-state-prescription LEGACY --logfile-type HDF5 --logfile-definitions /app/COMPAS/COMPAS_Output_Definitions.txt --neutrino-mass-loss-BH-formation FIXED_MASS --add-options-to-sysparms GRID

Logfile record specifications error
File does not exist: /app/COMPAS/COMPAS_Output_Definitions.txt

COMPAS v02.25.03
Compact Object Mergers: Population Astrophysics and Statistics
by Team COMPAS (http://compas.science/index.html)
A binary star simulator
```
