### 2021-11-03 TeamCOMPAS/COMPAS#680 @jeffriley
Thanks Heloise!  The funny thing is I noticed the typo in the C++ code - and fixed it - but I didn't think to look in pythonSubmit.  My guess is the author of that new part of the code copy-and-pasted from the C++ code before I noticed the typo and fixed it.

We'll push a fix.

@ilyamandel fyi.  We also should get the documentation for the new option online.
