### 2021-11-03 TeamCOMPAS/COMPAS#680 @HeloiseS
Aces. What about the log file record error? What did I do wrong or miss out?
