### 2021-11-03 TeamCOMPAS/COMPAS#680 @ilyamandel
Hi @HeloiseS,
surive is entirely my fault -- I made a typo in the latest version that I committed on Monday.  Thanks, @jeffriley!
BTW, do you want to review the living code or a particular tag (we tagged v02.21.00, which is the reference version for the ApJS submission)?
