### 2021-11-03 TeamCOMPAS/COMPAS#680 @jeffriley
> Aces. What about the log file record error? What did I do wrong or miss out?

Sorry - forgot to get back t o that.  I'll take a look later tonight or tomorrow.
