### 2021-11-03 TeamCOMPAS/COMPAS#680 @HeloiseS
@ilyamandel that's interesting. How did you end up with 4 identical typos across 2 files? did you accidentally upload old versions? You don't really need to asnwer here, the reason I'm asking is that ideally one would want to avoid that sort of stuff from happening and if it does (because we're all human after all) to get caught before it gets on the release branch, which brings me to continuous integration and unitest: one of JOSS boxes we'll have to discuss. Unitesting a giant C++ star evolution code might not be practical, but having a system to avoid that sort of bug in the surrounding applications (particularly the ones used to run it) would be important.

As for what version I'm reviewing: the living code :) currently on compas 2.25 I believe.
