### 2021-11-03 TeamCOMPAS/COMPAS#680 @jeffriley
> How did you end up with 4 identical typos across 2 files?

That would be partly my fault :-)  @ilya updated preProcessing/pythonSubmit.py, and presumably copy-pasted the incorrectly spelled option name from the C++ source.  Later I noticed thetypo in the C++ source and fixe it, but didn't realise it had been copy-pasted to preProcessing/pythonSubmit.py.  While making some mods to the pythonSubmit files in the examples directories, I thought I'd add the new options to those pyhtonSubmit files, so I copy-pasted the new (incorrectly spelled) options from preProcessing/pythonSubmit.py.
