### 2021-11-03 TeamCOMPAS/COMPAS#680 @jeffriley
> I couldn't figure out why it wasn't happy, the COMPAS_Output_Definitions.txt is in the directory where i got the pythonSubmit file

That's the problem.  When COMPAS is run in a docker image, it picks up the environment variables COMPAS_LOGS_OUTPUT_DIR_PATH and COMPAS_INPUT_DIR_PATH.

$COMPAS_LOGS_OUTPUT_DIR_PATH points to the directory to which the log files are written (so for COMPAS output files)
$COMPAS_INPUT_DIR_PATH points to the directory from which the grid file and logfile-definitions files are read (so for COMPAS input files)

pythonSubmit.py is not a COMPAS input file, so it (probably) won't be  in $COMPAS_INPUT_DIR_PATH (so putting the logfile-definitions file in the same directory as pythonSubmit.py almost certainly means it (the logfile-definitions file) is in the wrong directory (or at least is in a directory COMPAS doesn't know to look in).  The COMPAS executable knows nothing about pythonSubmit.py - all pythonSubmit.py does is construct the command line to execute COMPAS, and then executes it.  COMPAS, now running, goes and looks in $COMPAS_INPUT_DIR_PATH for its input files (grid and log-file-definitions), and puts its output in $COMPAS_LOGS_OUTPUT_DIR_PATH

So the command you should run for docker is:

    sudo docker run
    --rm
    -it
    -v $COMPAS_ROOT_DIR/compas-logs:/app/COMPAS/logs
    -v
     $COMPAS_ROOT_DIR/examples/methods_paper_plots/chirpmass_distribution/pythonSubmit.py:/app/starts/pythonSubmit.py
    -e COMPAS_EXECUTABLE_PATH=/app/COMPAS/bin/COMPAS
    -e COMPAS_INPUT_DIR_PATH=/app/COMPAS/config
    -e COMPAS_LOGS_OUTPUT_DIR_PATH=/app/COMPAS/logs
    teamcompas/compas
    python3 /app/starts/pythonSubmit.py

    (where /app/COMPAS/config is the directory COMPAS will look in to find the grid file and/or the logfile-definitions file)

Not your fault of course - our example doesn't show it, because it doesn't use a grid file or a logfile-definitions file.  I'll make a note to update the example.

Thanks for finding the omission :-)
