### 2021-11-03 TeamCOMPAS/COMPAS#680 @HeloiseS
I still get the same error so I have some follow up questions:

### Where is COMPAS_Output_Definitions.txt meant to go?

> goes and looks in $COMPAS_INPUT_DIR_PATH

Where is that? I went into the pythonSumbit file (I remembered seeing it in there) and there is a nicely detailed comment (see below), that I thought meant I was supposed to have the COMPAS_Output_Definitions.txt in the current working directory, but I still got the same error.
```
    # environment variable COMPAS_INPUT_DIR_PATH is used primarily for docker runs
    # if COMPAS_INPUT_DIR_PATH is set (!= None) it is prepended to input filenames
    # (such as grid_filename and logfile_definitions)
    # if COMPAS_INPUT_DIR_PATH is not set (== None) the current working directory
    # is prepended to input filenames
```

So let's get back to basics:
1) Where am i supposed to be? I assumed i could be anywhere on my computer given the environment variable points to compas, but maybe for the config files i need to be in a specific location?
2) What really is going on with COMPAS_INPUT_DIR_PATH and COMPAS_OUTPUT_DIR_PATH in this case? It looks like we are defining them in the command (and i actually removed the `-e COMPAS_INPUT_DIR_PATH=/app/COMPAS/config` at one point thinking, maybe if i didn't set it it would default to looking ind my cwd), but that looks like a location in the docker image not on my computer so I'm a tad confused.

### COMPAS_Outputs tucked away in compas-log
I initially copied the output definitions to $COMPAS_ROOT_DIR/compas-logs  (i don't know what i was trying to achieve :rofl:). It made me realise that  running docker created a directory compas-log in my $COMPAS_ROOT_DIR that is locked (I am  mainly a GUI user, and I see a little lock on the directory picture), the permisions were: drwxr-xr-x. I also realised that it wasn't empty!

I found a COMPAS_Output folder with the results from the docker run i did with the command line (btw I might as well tell you know, but I think [the docs](https://compas-docs-readthedocs.readthedocs.io/en/latest/running-via-docker-running.html) need updating because the command `--number-of-binaries=5` gave me an error saying that it didn't exist, I replaced it with `--number-of-systems` and it ran :ok_hand:

**Some Questions**
- Why are the permissions different on those directories than the ones created via not-docker? Not saying it needs changing just wondering, because I very aggressively `sudo chmod 777`-ed the directory - is that not recommended? I dunno if the permission setting is a fun and quirky docker thing that you're happy for users to update or if there's a reason for it and I  shouldn't mess with it.

- What's the time recorded in the log files? Is it different on docker runs and command line runs? The reason I ask is because in that log I have `Start generating binaries at Wed Nov  3 05:00:29 2021` and based on other runs that time is in 24 hour format, but I certainly didn't run compas at 5 AM so that's weird?
