### 2021-11-03 690c7fa
Typo fixes (suvive), doc updates
