### 2021-11-03 fac9fb4
Typo fixes (suvive), doc updates
