### 2021-11-03 TeamCOMPAS/COMPAS#680 @ilyamandel
PR #681 fixes the surive->survive typos and updates the documentation to include the new introduced CE options.  I leave the other questions for @jeffriley . :)

Aside: @jeffriley , do we still need the documentation in docs (rather than docs/online-docs)?  For example, https://github.com/TeamCOMPAS/COMPAS/blob/dev/docs/getting_started.md is out of date [see, e.g., reference to non-existing .pdf file in the final paragraph] -- but rather than updating it (and always remembering that we have two versions to update), should we just delete it?
