### 2021-11-03 TeamCOMPAS/COMPAS#682 @HeloiseS
[DOCS] 1_Slicing_Data.ipynb  | KeyError: "Unable to open object 
_This issue is submitted as part of the JOSS review [#3838](https://github.com/openjournals/joss-reviews/issues/3838)_

### :bug:  description
I got a `KeyError: "Unable to open object (object 'DoubleCompactObjects' doesn't exist)"` on cell 5, evidently because the COMPAS_Ouptut file i generated with one of my test runs does not have double compact objects.

### Recommendation
This notebook has some untold expectations of the output to process - I recommend the developers choose one output example and provide a link to it for the user to run this notebook with. I did see the link to the general compas website to download data, but as a new user  (and not knowing what the notebook expects) I wouldn't know what to chose to ensure proper running.

**Note**
@ilyamandel @jeffriley, If i encounter other issues in the jupyter notebooks/tutorials would you like me to collate them in this issue or make new ones for each?
