### 2021-11-03 TeamCOMPAS/COMPAS#682 @HeloiseS
I looked a little further down the notebook and I don't think it'll run with outputs form COMPAS v2.25
I'm seeing `Data['SystemParameters']['Mass@ZAMS_1'][()]` and I my output data has the key `Mass@ZAMS(1)`.

When did the keys changes? It's worth thinking about having a warning somewhere in the guide in case someone learns with new compas and then downloads and tries to use data generated a while back.

----

**Wait a minute my jupyter notebooks don't match the read the docs!** - I think I have some old files. Looking [here](https://github.com/TeamCOMPAS/COMPAS/tree/dev/postProcessing/Folders/BasicsPython) I can see the ones i downloaded from the release branch were made 2 years ago! There it is ;)

Please update the release branch with the up to date notebooks, adding a link to the data you want me to use :) I'll give them another try then.
