### 2021-11-04 TeamCOMPAS/COMPAS#680 @jeffriley
> do we still need the documentation in docs (rather than docs/online-docs)? For example, https://github.com/TeamCOMPAS/COMPAS/blob/dev/docs/getting_started.md is out of date

I'd be happy to remove them, otherwise, as you say, we'd have to keep them in sync with the online documentation - they were created by @reinhold-willcox and @themikelau I think, so they might have an opinion.
