### 2021-11-04 TeamCOMPAS/COMPAS#680 @themikelau
Happy to get rid of the documents in the `docs` directory as long as we have made sure we don't refer to them anywhere.
