### 2021-11-04 TeamCOMPAS/COMPAS#680 @jeffriley
> Where is COMPAS_Output_Definitions.txt meant to go?
> goes and looks in $COMPAS_INPUT_DIR_PATH
> Where is that?

There are two separate concepts here.

First, COMPAS needs to know where to look for its input files (grid file, logfile-definitions file), and it needs to know where to put its output (log) files.

The default for both of these is CWD.

COMPAS provides a program option (command-line argument) to specify the path to be used for output files: --ouput-path

    ./compas --mode BSE -n 1000 --output-path path-to-where-user-wants-logfiles

There is no program option to specify the directory in which the input files reside (maybe there should be - never thought about it until now), so if the user wants the input files in a specific directory, they need to use a fully-qualified filename when the specify the input files:

    ./compas --mode BSE -n 1000 --output-path path-to-where-user-wants-logfiles --grid fully-qualified-path-to-grid-file  --logfile-definitions fully-qualified-path-to-logfile-definitions-file

So that's how COMPAS manages input and output files.

pythonSubmit.py will use the following environment variables if they are set:

    COMPAS_INPUT_DIR_PATH
    COMPAS_LOGS_OUTPUT_DIR_PATH

If COMPAS_INPUT_DIR_PATH is set, pythonSubmit.py will prepend it to any input filenames specified (grid file, logfile-definitions file).
If COMPAS_INPUT_DIR_PATH is NOT set, pythonSubmit.py will prepend the CWD to any input filenames specified (grid file, logfile-definitions file) - not strictly necessary, because the COMPAS executable will default to the CWD anyway.

If COMPAS_LOGS_OUTPUT_DIR_PATH is set, pythonSubmit.py use $COMPAS_LOGS_OUTPUT_DIR_PATH as the value for the '--output-path' program option.
If COMPAS_LOGS_OUTPUT_DIR_PATH is NOT set, pythonSubmit.py will use the CWD as the value for the '--output-path' program option - not strictly necessary, because the COMPAS executable will default to the CWD anyway.

Second, COMPAS can be run on a physical machine, or in a VM using the docker image.

The user can't put the input files in a directory in the VM, nor can they access the output (log) files in a directory in the VM (directly, anyway).  We have to map a directory on a physical machine to a directory in the VM.  That's what the "-v" options to the 'docker run' command do.

    -v $COMPAS_ROOT_DIR/compas-logs:/app/COMPAS/logs

maps the directory $COMPAS_ROOT_DIR/compas-logs on a physical system to the /app/COMPAS/logs directory in the VM.

Putting both of those together should help explain the 'docker run' options, and where the files should be.
