### 2021-11-04 TeamCOMPAS/COMPAS#682 @jeffriley
I'll defer here to @reinhold-willcox, @themikelau, @cneijssel etc. (I don't know who else...).  I don't do Jupyter notebooks.  It could very well be my meddling trying to get them into the online docs that caused the problems Heloise is seeing...
