### 2021-11-04 TeamCOMPAS/COMPAS#683 @HeloiseS
h5view.py -- Error accessing HDF5 file -- 'Resource temporarily unavailable'
_This issue is submitted as part of the JOSS review [#3838](https://github.com/openjournals/joss-reviews/issues/3838)_

### :bug: description
I was trying to run the `h5view.py` utility on a test output and i got the following error. It happened both when i called the script from a separate directory and when after I copied the script to my working directory.

```
Error accessing HDF5 file /home/fste075/Documents/reviews/COMPAS/COMPAS_Output/COMPAS_Output.h5 : Unable to open file (unable to lock file, errno = 11, error message = 'Resource temporarily unavailable')
```
**Versioning (please complete the following information):**
 - OS: [e.g. Ubuntu 20.02]
 - COMPAS [e.g. v02.25.03]
