### 2021-11-04 dc25e0d
Changed GiantBranch::CalculateRemnantMassByMullerMandel() to ensure that the remnant mass is no great than the He core mass
