### 2021-11-04 TeamCOMPAS/COMPAS#683 @jeffriley
It could be that the file is currently open and being accessed by another process (even though you should be able to open it for read access, I think h5py complains about it).

Can you access it with h5dump (comes with hdf5 if you have that installed)?
