### 2021-11-04 TeamCOMPAS/COMPAS#683 @HeloiseS
oops yes - i had it open in jupyter, forgot about that.

Changed the title of the issue so if someone gets the same error not knowing what it means they can find this :)
