### 2021-11-04 #3838 @HeloiseS
Update! I'm going through each page of the user manual and tutorial and it's going very smoothly for the most part. I'll make a few general recommendations about the user manual and tutorial at the end when I've digested the whole thing and I've successfully ran all the notebooks. That being said, I have a couple of requests regarding the command list descriptions that i know i'm going to make so I might as well tell you now :smile:

## Program option list and default values

- **Topic-based command listing**: The alphabetical list with all the details on the commands is really good and exhaustive. Problem is if you don't know what options do exist in compas it's a tough way to figure it out. Could we also have a summary page (or table or else) of just the command names grouped by topic? That can help people find the commands they might need and then they can look up the exact descriptions in the alphabetical list. To be fair, the effective naming conventions of the commands do a lot of that job (e.g. the kicks are all together), but it doesn't do everything, for example LBV options are at the end of a very long list of logs and might be overlooked, or --detailed-output is after a bunch of debug commands.

- Some of the command options are basically references, but many people won't know what they actually mean. It would be good to at least provide a link to the references (and so commands do already have references), and if you think it's relevant an equation or section number. To make your life easier I have made a note as I read of all the commands that need a ref(s), they are listed below.
```
                --common-envelope-lambda-prescription
                --common-envelope-mass-accretion-prescription
                --common-envelope-slope-kruckow
                --cool-wind-mass-loss-multiplier
                --eccentricity-distribution
                --envelope-state-prescription
                --initial-mass-function
                --kick-magnitude-distribution
                --kick-phi-1 -> typo "for the" repeated
                --luminous-blue-variable-multiplier
                --luminous-blue-variable-prescription
                --mass-ratio-distribution
                --mass-transfer-rejuvenation-prescription
                --metallicity-distribution
                --pulsational-pair-instability-prescription
                --remnant-mass-prescription
                --rotational-velocity-distribution
                --semi-major-axis-distribution
                --stellar-zeta-prescription
```

- **Other questions:**
    * `--case-bb-stability-prescription` has an option TREAT_AS_OTHER_MT ,  what does that mean?
    *  `--common-envelope-allow-main-sequence-survive`:  does it mean they always do or they're just not constrained to merge?
    *  `--common-envelope-lambda-multiplier` What does that do? Why not just have a larger lambda param value?
   *  `--debug-classes` description has vactor instead of vector. Also what are the options?
   *  `--envelope-state-prescription` What is legacy? (if that's taken care of by adding references ignore me)
   *  `--help` state that -h and --help don't do the same thing on this page too
   *  ` --kick-direction`  what's the wedge? is poles just the poles or an opening angle around the pole (from the ApJ I'm guessing that's the cone you mention on p18)? perpendicular to the orbit I guess? what's in-plane?
   *  ` --kick-magnitude-random` Is the random distribution taken from the `--kick-magnitude-distribution` speficied or defaulted? Manual says the float is used to draw the kick magnitude but not sure what that means in that context since distributions need more numbers than that right?
   *  `--kick-magnitude-sigma-CCSN-BH` is that for a maxwellian kick distribution (and so assuming that `--kick-magnitude-distribution` goes to default)? What if the kick distribution has been set to something that doesn't take a sigma value?
   *  `--kick-mean-anomaly-1` What's this ?:D
   *  `--mass-ratio-distribution`  No Moe and Di stefano :(?
   *  `--neutrino-mass-loss-BH-formation-value` _"Amount of mass lost in neutrinos during BH formation (either as fraction or in solar masses, depending on the value of --neutrino-mass-loss-bh-formation)"_ Is the default in msol or as a fraction? does it turn into solar masses above 1 or when's that threshold?
