### 2021-11-04 461a487
Avoid SN remnants more massive than preSN stars in Mandel&Mueller prescription (#684)
* Changed GiantBranch::CalculateRemnantMassByMullerMandel() to ensure that the remnant mass is no great than the He core mass

* Fixed typo

Changed

    to ensure that the remnant mass is no great than the He core mass

to

    to ensure that the remnant mass is no greater than the He core mass

Co-authored-by: jeffriley <jeffrey.riley@monash.edu>
