### 2021-11-04 TeamCOMPAS/COMPAS#680 @jeffriley
> Why are the permissions different on those directories than the ones created via not-docker

I hadn't noticed.  It's probably something docker does (I haven't checked), but certainly after the docker run is finished you can freely change the permission to whatever suits you

> What's the time recorded in the log files?

The time of the run in Run_Details?  That's the time on the system used to run COMPAS.  We set the timezone in our docker images to UTC/GMT, which would explain what you're seeing (NZDT is UTC+13 I think).  We did that because our docker image can be run anywhere - I've done a number of runs on AWS with 32 nodes, and those nodes were in the US somewhere (may have even been across timezones), so recording local time wouldn't make sense.
