### 2021-11-04 #3838 @jeffriley
> --debug-classes description has vactor instead of vector. Also what are the options?

I'll fix the "vactor".  These are user (actually, more correctly, developer) defined.  When a developer wants to do some debugging, they can add debug statements that have an associated "class" string - so they can dot many debug statements around the code with different class strings, then when they run the code they can specify for which of those class strings they want the debug statements printed - they can print those that match just one class string, or an array of class strings.  It's just a way of making life a bit easier for a developer looking for a problem - they can turn on/off different debug statements in the code without having to change the code.

Log classes are similar, but they aren't really used any more - though they could be.  Similar concept to debug classes, but instead of debug statement with debug class strings, we have log statements with log class strings.  At the moment, all log statements have as an associated class string the name of the file they are logging to - so users could turn logging to particular files on and off using the '--log-classes' option, but there are also other options to control which log files we produce, so  that's a bit redundant.  Log classes were implemented in the first iteration of the logging class, but over time their need lessened.

> --help state that -h and --help don't do the same thing on this page too

Sure, good idea.

> --mass-ratio-distribution No Moe and Di stefano :(?

I thought I saw or heard discussion about that being in the plans.

I'll leave explanations of the other options to others who can probably provide more detail than me.  Reinhold is a good person to ask about the kick-related options (but Ilya and others will be able to answer those questions too).
