### 2021-11-04 TeamCOMPAS/COMPAS#682 @reinhold-willcox
Thanks @HeloiseS for pointing this out. Updating the notebooks has been one of the "back-burner" priorities for a while now. As has adding them to a CI check to make sure we catch it sooner if they go out of date again.

I'll go through them now and try to clean them up.
