### 2021-11-04 TeamCOMPAS/COMPAS#686 @ilyamandel
Clean documentation in docs/
There is some documentation in docs (rather than docs/online-docs) that is outdated and no longer needed. For example, https://github.com/TeamCOMPAS/COMPAS/blob/dev/docs/getting_started.md is out of date (see, e.g., reference to non-existing .pdf file in the final paragraph).  Rather than updating it, and always remembering that we have two versions to update, we should check that it's not referenced elsewhere and delete it.
