### 2021-11-04 TeamCOMPAS/COMPAS#680 @ilyamandel
Since there are now multiple threads here, I opened up docs/ cleaning as a separate issue, #686 .
