### 2021-11-04 TeamCOMPAS/COMPAS#680 @HeloiseS
Thanks for the detailed explenation @jeffriley now the docker command makes complete sense - it still doesn't work :laughing: but i can start asking questions more confidently:

> If COMPAS_INPUT_DIR_PATH is NOT set, pythonSubmit.py will prepend the CWD to any input filenames specified

Okay that's what i thought, and I did make sure the COMPAS_Output_definitions.txt was in my CWD. I haven't set COMPAS_INPUT_DIR_PATH, I put nothing in my .bashrc,  and i checked that doing`$COMPAS_INPUT_DIR_PATH` in my shell gives me nothing, so I think it's fair to assume it will default.

So why isn't it seeing the `COMPAS_Output_definitions.txt` that's in my cwd?
I don't know if it'll help but here is the directory i'm running things in
```
fste075@kvin:~/Documents/reviews/COMPAS$ ls -al
total 204
drwxr-xr-x 4 fste075 62215   4096 Nov  4 17:45 .
drwxr-xr-x 8 fste075 62215   4096 Oct 30 13:40 ..
-rw-r--r-- 1 fste075 62215 166138 Oct 30 13:56 10.21105.joss.03838.pdf
drwxr-xr-x 3 fste075 62215   4096 Nov  4 15:54 COMPAS_Output
drwxr-xr-x 2 fste075 62215   4096 Nov  3 16:51 COMPAS_Output_2
-rw-r--r-- 1 fste075 62215   1319 Oct 30 14:07 COMPAS_Output_Definitions.txt
-rw-r--r-- 1 fste075 62215  16526 Nov  4 17:45 review.log

```
----
> but certainly after the docker run is finished you can freely change the permission to whatever suits you

- [ ] Maybe worth mentioning in a sentence in the manual? Confident unix users would just do it instinctively but baby-PhD-me would have been scared to mess with a locked directory ^^"

> The time of the run in Run_Details? That's the time on the system used to run COMPAS. We set the timezone in our docker images to UTC/GMT, which would explain what you're seeing (NZDT is UTC+13 I think). We did that because our docker image can be run anywhere - I've done a number of runs on AWS with 32 nodes, and those nodes were in the US somewhere (may have even been across timezones), so recording local time wouldn't make sense.

- [ ] That makes sense! Probably worth mentioning in the the docker page manual that the time recording is different :)
