### 2021-11-04 #3838 @HeloiseS
Thanks!!

> I'll leave explanations of the other options to others who can probably provide more detail than me. Reinhold is a good person to ask about the kick-related options (but Ilya and others will be able to answer those questions too).

@ilyamandel  do you want me to add this previous post about command clarification to the issue [#686](https://github.com/TeamCOMPAS/COMPAS/issues/686) so that Reinhold et al. can join in?
