### 2021-11-04 TeamCOMPAS/COMPAS#680 @jeffriley
> So why isn't it seeing the COMPAS_Output_definitions.txt that's in my cwd?

Are you running on your physical system, or using a docker image?  Can you share the command you are using to run?
