### 2021-11-04 TeamCOMPAS/COMPAS#682 @jeffriley
> I'll go through them now and try to clean them up.

@reinhold-willcox I've copied the notebooks to the online documentation - so in docs/online-docs/notebooks

I tidied them up a little - mainly just rewording and/or reformatting some of the text(I tried to make headings uniform, etc.), but also some of the paths to files.  I'm not a notebooks user, so I might have inadvertently broken something (though they all seem to run if you go through the online docs versions).
