### 2021-11-04 TeamCOMPAS/COMPAS#680 @HeloiseS
Docker: I'm using the command you gave me the other day :)

```
sudo docker run   --rm  -it  -v $COMPAS_ROOT_DIR/compas-logs:/app/COMPAS/logs  -v  $COMPAS_ROOT_DIR/examples/methods_paper_plots/chirpmass_distribution/pythonSubmit.py:/app/starts/pythonSubmit.py -e COMPAS_EXECUTABLE_PATH=/app/COMPAS/bin/COMPAS -e COMPAS_INPUT_DIR_PATH=/app/COMPAS/config -e COMPAS_
```
