### 2021-11-04 TeamCOMPAS/COMPAS#680 @jeffriley
This is my fault - sorry.

Your command is this:

    sudo docker run
    --rm
    -it
    -v $COMPAS_ROOT_DIR/compas-logs:/app/COMPAS/logs
    -v
      $COMPAS_ROOT_DIR/examples/methods_paper_plots/chirpmass_distribution/pythonSubmit.py:/app/starts/pythonSubmit.py
    -e COMPAS_EXECUTABLE_PATH=/app/COMPAS/bin/COMPAS
    -e COMPAS_INPUT_DIR_PATH=/app/COMPAS/config
    -e COMPAS_LOGS_OUTPUT_DIR_PATH=/app/COMPAS/logs
    teamcompas/compas
    python3 /app/starts/pythonSubmit.py

The  command-line argument:

    -v $COMPAS_ROOT_DIR/compas-logs:/app/COMPAS/logs

maps the VM directory /app/COMPAS/logs to the directory $COMPAS_ROOT_DIR/compas-logs on your physical system.  And the argument:

    -e COMPAS_LOGS_OUTPUT_DIR_PATH=/app/COMPAS/logs

sets the environment variable COMPAS_LOGS_OUTPUT_DIR_PATH in the VM to /app/COMPAS/logs, so COMPAS will write its log files to /app/COMPAS/logs in the VM, which is mapped to $COMPAS_ROOT_DIR/compas-logs, so the COMPAS logs will be in the directory $COMPAS_ROOT_DIR/compas-logs on your physical system

This command-line argument:

    -e COMPAS_INPUT_DIR_PATH=/app/COMPAS/config

sets the environment variable COMPAS_INPUT_DIR_PATH in the VM to /app/COMPAS/config, so COMPAS (via pythonSubmit.py) will look in /app/COMPAS/config in the VM for any grid file or logfile-definitions file, but we haven't mapped that directory anywhere!

Sorry, this is my fault - I didn't look carefully enough when I added the command-line argument:

    -e COMPAS_INPUT_DIR_PATH=/app/COMPAS/config

I should have added this mapping (I assumed it was there - I don't know why):

    -v directory-on-physical-system-for-input-files:/app/COMPAS/config

where "directory-on-physical-system-for-input-files" should be (in this case) "~/Documents/reviews/COMPAS" (or the path to the directory on t he physical system in which the logfile-definitions file resides)

So the "input" directory and the "output" directory both need the respective environment variable set, and a mapping physical->virtual.

Sorry - I should have paid closer attention when I answered the question initially.
