### 2021-11-05 TeamCOMPAS/COMPAS#680 @HeloiseS
lmao :rofl:  I'm going insane it's still not working.

- I tried what you said - to set the /app/COMPAS/config to the path of my CWD -> didn't work
- I then went "fuck it I'll just define COMPAS_INPUT_DIR_PATH in my bashrc"  and ran the first command you sent me -> didn't work, although i know i did it right - I can `more` the file with `$COMPAS_INPUT_DIR_PATH/COMPAS_Output_Definitions.txt`.

#### I think i had an epiphany!!!
=> the error is `File does not exist: /app/COMPAS/COMPAS_Output_Definitions.txt` so from what i understand **it's not even looking in /app/COMPAS/config** it's looking in /app/COMPAS/... so is there a problem in the pythonSubmit where it's not making the command properly to look in the **config** directory?

It would make sense if the python file was the problem actually because i ran docker no problem with the command line version
