### 2021-11-05 TeamCOMPAS/COMPAS#680 @jeffriley
> so is there a problem in the pythonSubmit where it's not making the command properly to look in the config directory?

It would seem so...

So, the version of the pythonSubmit.py file I'm looking at has:

    compas_input_path_override = os.environ.get('COMPAS_INPUT_DIR_PATH')

    ...

    if logfile_definitions != None:
        # if the grid filename supplied is already fully-qualified, leave it as is
        head, tail = ntpath.split(logfile_definitions)          # split into pathname and base filename

        if head == '' or head == '.':                           # no path (or CWD) - add path as required
            logfile_definitions = tail or ntpath.basename(head)
            if compas_input_path_override == None:
                logfile_definitions = os.getcwd() + '/' + logfile_definitions.strip("'\"")
            else:
                logfile_definitions = compas_input_path_override + '/' + logfile_definitions.strip("'\"")

Leaving aside the typo (# if the grid filename supplied), that should prepend whatever is in the environment variable COMPAS_INPUT_DIR_PATH to the logfile-definitions file name.

Oh wait...  that will fail if the logfile-definitions file already has a path prepended to it.  Hmmm.  Might have to rethink that code.  So, if the logfile-definitions file (and  the grid file if you use one), is not just a bare filename, then you should make it one (i.e. no path prepended), and use the environment variable to prepend the path.

So what does the logfile-definitions filename look like in your pythonSubmit file?  (Also grid file name)
