### 2021-11-05 TeamCOMPAS/COMPAS#680 @HeloiseS
**TL;DR: logfile_definitions is None and pythonSubmit broken in 2.25?**

I have the following bit of code in the pythonSubmit file
```
    if logfile_definitions != None:
        # if the grid filename supplied is already fully-qualified, leave it as is
        head, tail = ntpath.split(logfile_definitions)          # split into pathname and base filename

        if head == '' or head == '.':                           # no path (or CWD) - add path as required
            logfile_definitions = tail or ntpath.basename(head)
            if compas_input_path_override == None:
                logfile_definitions = os.getcwd() + '/' + logfile_definitions.strip("'\"")
            else:
                logfile_definitions = compas_input_path_override + '/' + logfile_definitions.strip("'\"")
```
#### The python error
I tried running the pythonSubmit file alone (it's the one in chirpmassDistributions btw) to try and print you the `logfile_definitions` but :musical_note: it faiiiiled :musical_note: did y'all update the commands again? I hadn't rerun the files since 2.24 but it seems there's extra stuff in v2.25
`Commandline Options error: unrecognised option '--common-envelope-allow-radiative-envelope-survive'`

This command doens't exist in your user manual, so (without looking at the PR) I'm assuming it came form the recent update - both my pythonSubmit and pythonSubmitDemo have it and don't run.

I tried removing the problematic command and there is _another one_ that's causing me grief `--common-envelope-allow-immediate-rlof-post-ce-survive`. I rebuilt the c++ because altho I pulled the new PR I hadn't run the makefile again... stil broken. When i git pull, does it know if the c++ has changed? I'm wondering if maybe it jsut didn't get the updated src

### The  logfile_definitions comes out as `None`!
(_fire_elmo.gif_)

here's some more detail:
- The environment variable **is correct**: /home/fste075/COMPAS/examples/methods_paper_plots/chirpmass_distribution
- The `self.logfile_definitions` is initially `None` which is alright since we don't set it in this file
- BUT the final `self.logfile_definitions` at the end of the if statments is also `None` when it should be `compas_input_path_override`.
