### 2021-11-05 TeamCOMPAS/COMPAS#680 @HeloiseS
SORRY FOR THE SPAM - realised i misread something so just reposting the correct information

**TL;DR: logfile_definitions is okay I think and pythonSubmit broken in 2.25?**

[----see below----]

#### The python error
I tried running the pythonSubmit file alone (it's the one in chirpmassDistributions btw) to try and print you the `logfile_definitions` but :musical_note: it faiiiiled :musical_note: did y'all update the commands again? I hadn't rerun the files since 2.24 but it seems there's extra stuff in v2.25
`Commandline Options error: unrecognised option '--common-envelope-allow-radiative-envelope-survive'`

This command doens't exist in your user manual, so (without looking at the PR) I'm assuming it came form the recent update - both my pythonSubmit and pythonSubmitDemo have it and don't run.

I tried removing the problematic command and there is _another one_ that's causing me grief `--common-envelope-allow-immediate-rlof-post-ce-survive`. I rebuilt the c++ because altho I pulled the new PR I hadn't run the makefile again... stil broken. When i git pull, does it know if the c++ has changed? I'm wondering if maybe it jsut didn't get the updated src
