### 2021-11-05 TeamCOMPAS/COMPAS#680 @HeloiseS
#### UPDATE: logfile_definitions = /home/fste075/COMPAS/examples/methods_paper_plots/detailed_evolution/COMPAS_Output_Definitions.txt

My bad I copied stuff form the wrong pythonSubmit. Above was pythonSubmitDemo, the pythonSubmit in chirpmassDistribution DOES have `logfile_definitions = "COMPAS_Output_Definitions.txt" ` and therefore the logfile_definitions come out as i expected. Huzza!

Still docker doesn't work hahah.

here is the code you asked for from the **correct file** this time
```
    grid_filename = None                                        # grid file name (e.g. 'mygrid.txt')

    if grid_filename != None:
        # if the grid filename supplied is already fully-qualified, leave it as is
        head, tail = ntpath.split(grid_filename)                # split into pathname and base filename

        if head == '' or head == '.':                           # no path (or CWD) - add path as required
            grid_filename = tail or ntpath.basename(head)
            if compas_input_path_override == None:
                grid_filename = os.getcwd() + '/' + grid_filename.strip("'\"")
            else:
                grid_filename = compas_input_path_override + '/' + grid_filename.strip("'\"")

    logfile_definitions = "COMPAS_Output_Definitions.txt"       # logfile record definitions file name (e.g. 'logdefs.txt')

    if logfile_definitions != None:
        # if the grid filename supplied is already fully-qualified, leave it as is
        head, tail = ntpath.split(logfile_definitions)          # split into pathname and base filename

        if head == '' or head == '.':                           # no path (or CWD) - add path as required
            logfile_definitions = tail or ntpath.basename(head)
            if compas_input_path_override == None:
                logfile_definitions = os.getcwd() + '/' + logfile_definitions.strip("'\"")
                print('blaaa')
            else:
                logfile_definitions = compas_input_path_override + '/' + logfile_definitions.strip("'\"")
```
