### 2021-11-05 TeamCOMPAS/COMPAS#680 @jeffriley
--common-envelope-allow-radiative-envelope-survive and --common-envelope-allow-immediate-rlof-post-ce-survive are  in the  online documentation:

https://compas.readthedocs.io/en/latest/pages/User%20guide/Program%20options/program-options-list-defaults.html#options-props-c

You may need to reload the page - you could still have the old version in your browser cache (we updated it a couple of days ago).

Ok, the pythonSubmit looks ok.  Does your docker command include the mapping?  I know you have already put the docker command here, but can you do it again (because you should have added to it since the last time).  Then I'll have everything I need to try to reproduce what you're seeing.
