### 2021-11-05 TeamCOMPAS/COMPAS#680 @HeloiseS
> You may need to reload the page - you could still have the old version in your browser cache (we updated it a couple of days ago).

you're right :)

Still why does it error? Do I need to manually redownload the c++ code before I remake it?

> I know you have already put the docker command here, but can you do it again (because you should have added to it since the last time).

No problem here it is.
```
sudo docker run   --rm  -it  -v $COMPAS_ROOT_DIR/compas-logs:/app/COMPAS/logs  -v  $COMPAS_ROOT_DIR/examples/methods_paper_plots/chirpmass_distribution/pythonSubmit.py:/app/starts/pythonSubmit.py -e COMPAS_EXECUTABLE_PATH=/app/COMPAS/bin/COMPAS -e COMPAS_INPUT_DIR_PATH=/app/COMPAS/config -v ~/Documents/reviews/COMPAS=/app/COMPAS/config -e COMPAS_LOGS_OUTPUT_DIR_PATH=/app/COMPAS/logs teamcompas/compas python3 /app/starts/pythonSubmit.py
```
