### 2021-11-05 TeamCOMPAS/COMPAS#680 @jeffriley
> Still why does it error? Do I need to manually redownload the c++ code before I remake it?

You need to do something like 'git pull upstream dev' (depending upon how you've set things up.  'git remote -v' will tell you what upstream and origin etc are pointing to).

I'll look at the docker problem later.  Do you get the same error?
