### 2021-11-05 TeamCOMPAS/COMPAS#680 @HeloiseS
*facepalm*  it's not missing.... I typed a = instead of the :  in `-v ~/Documents/reviews/COMPAS=/app/COMPAS/config`
I'm so sorry!!!!

anyways it worked!!

okay i have completely lost track of what needs updating in the manual - I trust you haven't?
What I can do is that when the review is pretty much complete i can delete compas off my computer and then rerun everything clean just to double check.

But for now it's done! woohoo thanks a bunch :)

---

> You need to do something like 'git pull upstream dev' (depending upon how you've set things up. 'git remote -v' will tell you what upstream and origin etc are pointing to).

I'll try to fix that later.
