### 2021-11-05 TeamCOMPAS/COMPAS#680 @jeffriley
> anyways it worked!!

Great!

> okay i have completely lost track of what needs updating in the manual - I trust you haven't?

Yeah, I think I know what needs updating!
