### 2021-11-06 #3838 @ilyamandel
Sorry for the silence, @HeloiseS .  I'll handle the options clarification over the next 24 hours -- it should be sufficiently straightforward.
