+++ paper.md
@@ -109,7 +109,7 @@ Binary population synthesis generates population models of isolated stellar bina
 
 In particular, rapid binary population synthesis is needed in order to efficiently explore a broad parameter space of uncertain assumptions about the physics of stellar and binary evolution, including supernova remnant masses and natal kicks, mass transfer efficiency and stability, and the outcome of common-envelope events.  
 
-A range of binary population synthesis codes have been developed over the last three decades.  These include the Scenario Machine [@1996A&A...310..489L], IBiS [@]1996MNRAS.280.1035T, SeBa [@1996A&A...309..179P],  BSE [@]Hurley:2002rf], StarTrack [@Belczynski:2008],  binary$\_$c [@2004MNRAS.350..407I],  MOBSE=[@2018MNRAS.474.2959G]  and COSMIC [@2019arXiv191100903B].  These codes range from private to semi-public to fully public, and differ in the range of available tools, computational complexity, and speed of execution.
+A range of binary population synthesis codes have been developed over the last three decades.  These include the Scenario Machine [@Scenario], IBiS [@IBiS], SeBa [@Seba],  BSE [@Hurley:2002rf], StarTrack [@Belczynski:2008],  binary$\_$c [@BinaryC],  MOBSE [@2018MNRAS.474.2959G]  and COSMIC [@2019arXiv191100903B].  These codes range from private to semi-public to fully public, and differ in the range of available tools, computational complexity, and speed of execution.
 
 [COMPAS](https://compas.science) is a rapid binary population synthesis suite. It parametrizes complex astrophysical processes with prescriptions calibrated to detailed models.  COMPAS is designed to allow for flexible modifications as evolutionary models improve.  All code is fully public and, including pre-processing and post-processing tools.  COMPAS is computationally efficient, with a focus on the statistical analysis of large populations, particularly but not exclusively in the context of gravitational-wave astronomy.  
 
