+++ paper.bib
@@ -2013,7 +2013,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{2004MNRAS.350..407I,
+@ARTICLE{BinaryC,
    author = {{Izzard}, R.~G. and {Tout}, C.~A. and {Karakas}, A.~I. and {Pols}, O.~R.
   },
     title = "{A new synthetic model for asymptotic giant branch stars}",
@@ -6708,7 +6708,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{1996MNRAS.280.1035T,
+@ARTICLE{IBiS,
        author = {{Tutukov}, A. and {Yungelson}, L.},
         title = "{Double-degenerate semidetached binaries with helium secondaries: cataclysmic variables, supersoft X-ray sources, supernovae and accretion-induced collapses}",
       journal = {\mnras},
@@ -6747,7 +6747,7 @@ archivePrefix = {arXiv},
 
 
 
-@ARTICLE{1996A&A...310..489L,
+@ARTICLE{Scenario,
        author = {{Lipunov}, V.~M. and {Postnov}, K.~A. and {Prokhorov}, M.~E.},
         title = "{The Scenario Machine: restrictions on key parameters of binary evolution.}",
       journal = {\aap},
@@ -8440,7 +8440,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{1996A&A...309..179P,
+@ARTICLE{SeBa,
        author = {{Portegies Zwart}, S.~F. and {Verbunt}, F.},
         title = "{Population synthesis of high-mass binaries.}",
       journal = {\aap},
