### 2021-11-06 #3838 @ilyamandel
@HeloiseS , @christinahedges , @katiebreivik -- I updated the text to include a brief context of other available rapid binary population synthesis tools, as suggested.  [Please ignore the one broken reference in the last compiled version -- it's fixed now, just didn't want to generate another version.]
