### 2021-11-07 43eaa98
Updating user option documentation, minor tweaks
