+++ docs/online-docs/pages/User guide/Program options/program-options-list-defaults.rst
@@ -84,6 +84,7 @@ Default = FALLBACK
 **--case-bb-stability-prescription** |br|
 Prescription for the stability of case BB/BC mass transfer. |br|
 Options: { ALWAYS_STABLE, ALWAYS_STABLE_ONTO_NSBH, TREAT_AS_OTHER_MT, ALWAYS_UNSTABLE } |br|
+Case BB mass transfer is treated as always stable, always stable only for mass transfer onto neutron stars or black holes, with stability as determined for all other mass transfer, or always unstable, respectively |br|
 Default = ALWAYS_STABLE
 
 **--check-photon-tiring-limit** |br|
@@ -105,7 +106,7 @@ Allow binaries that experience Roche lobe overflow immediately at the end of the
 Default = FALSE
 
 **--common-envelope-allow-main-sequence-survive** |br|
-Allow main sequence donors to survive common envelope evolution. |br|
+Allow main sequence accretors to survive common envelope evolution if other criteria point to survival. |br|
 Default = TRUE
 
 **--common-envelope-allow-radiative-envelope-survive** |br| 
@@ -123,15 +124,17 @@ Default = 1.0
 
 **--common-envelope-lambda** |br|
 Common Envelope lambda. |br|
+Only used when ``--common-envelope-lambda-prescription = LAMBDA_FIXED``. |br|
 Default = 0.1
 
 **--common-envelope-lambda-multiplier** |br|
-Multiplication constant to be applied to the common envelope lambda parameter. |br|
+Multiplicative constant to be applied to the common envelope lambda parameter for any prescription. |br|
 Default = 1.0
 
 **--common-envelope-lambda-prescription** |br|
-CE lambda prescription. |br|
+CE lambda (envelope binding energy) prescription. |br|
 Options: { LAMBDA_FIXED, LAMBDA_LOVERIDGE, LAMBDA_NANJING, LAMBDA_KRUCKOW, LAMBDA_DEWI } |br|
+``LAMBDA_FIXED`` is a constant; ``LAMBDA_LOVERIDGE`` is the prescription from Loveridge et al., 2011; ``LAMBDA_NANJING`` is from Xu & Li, 2010; ``LAMBDA_KRUCKOW`` is from Kruckow et al., 2016; and ``LAMBDA_DEWI`` is the fit from Appendix A in Claeys et al. 2014, based on Dewi & Tauris 2000 |br|
 Default = LAMBDA_NANJING
 
 **--common-envelope-mass-accretion-constant** |br|
@@ -149,6 +152,7 @@ Default = 0.04
 
 **--common-envelope-mass-accretion-prescription** |br|
 Assumption about whether NS/BHs can accrete mass during common envelope evolution. |br|
+``ZERO`` is no accretion; ``CONSTANT`` means a fixed amount of accretion determined by ``--common-envelope-mass-accretion-constant''; ``UNIFORM`` means a uniform random draw between ``--common-envelope-mass-accretion-min`` and ``--common-envelope-mass-accretion-max`` (Oslowski et al., 2011); and ``MACLEOD`` follows the prescription of MacLeod et al., 2015 |br|
 Options: { ZERO, CONSTANT, UNIFORM, MACLEOD } |br|
 Default = ZERO
 
@@ -157,7 +161,7 @@ Recombination energy density (erg g−1). |br|
 Default = :math:`1.5 \times 10^{13}`
 
 **--common-envelope-slope-kruckow** |br|
-Common Envelope slope for Kruckow lambda. |br|
+Slope for the Kruckow lambda (see Kruckow et al. 2016 as implemented by Vigna-Gomez et al. 2018). |br|
 Default = −0.833333
 
 **--cool-wind-mass-loss-multiplier** |br|
[…]
