### 2021-11-07 d3a148b
Updating user option documentation, minor tweaks
