### 2021-11-07 #3838 @ilyamandel
@HeloiseS -- I updated the user options documentation as requested (just waiting for pull request https://github.com/TeamCOMPAS/COMPAS/pull/688 to be checked/approved).  Thank you for the excellent suggestions!
