+++ docs/online-docs/pages/User guide/Program options/temp.txt
@@ -415,25 +415,23 @@ Must be a floating-point number in the range :math:`[0.0, 2\pi)`. |br|
 Default = Random number drawn uniformly from :math:`[0.0, 2\pi)`
 
 **--kick-phi-1** |br|
-The angle between ’x’ and ’y’, both in the orbital plane of the supernova vector, for the primary star of a binary system when evolving in BSE mode, should it undergo a supernova event (radians). |br|
-Default = Drawn according to specified ``--kick-direction`` distribution
+The angle between ’x’ and ’y’, both in the orbital plane of the supernova vector, for the primary star
+of a binary system when evolving in BSE mode, should it undergo a supernova event (radians). |br|
+Default = Random number drawn uniformly from :math:`[0.0, 2\pi)`
 
 **--kick-phi-2** |br|
 The angle between ’x’ and ’y’, both in the orbital plane of the supernova vector, for the secondary
 star of a binary system when evolving in BSE mode, should it undergo a supernova event (radians). |br|
-Default = Drawn according to specified ``--kick-direction`` distribution
+Default = Random number drawn uniformly from :math:`[0.0, 2\pi)`
 
 **--kick-scaling-factor** |br|
 Arbitrary factor used to scale kicks. |br|
 Default = 1.0
 
 **--kick-theta-1** |br|
-The angle between the orbital plane and the ’z’ axis of the supernova vector for the primary star of a binary system when evolving in BSE mode, should it undergo a supernova event (radians). |br|
-Default = Drawn according to specified ``--kick-direction`` distribution
-
-**--kick-theta-2** |br|
-The angle between the orbital plane and the ’z’ axis of the supernova vector for the secondary star of a binary system when evolving in BSE mode, should it undergo a supernova event (radians). |br|
-Default = Drawn according to specified ``--kick-direction`` distribution
+The angle between the orbital plane and the ’z’ axis of the supernova vector for the for the primary star of a
+binary system when evolving in BSE mode, should it undergo a supernova event (radians). |br|
+Default = Random number drawn uniformly from :math:`[0.0, 2\pi)`
 
 .. _options-props-L:
 
@@ -924,27 +922,11 @@ Go to :ref:`the top of this page <options-props-top>` for the full alphabetical 
 
 **Initial conditions**
 
---initial-mass-function, --initial-mass, --initial-mass-1, --initial-mass-2, --initial-mass-min, --initial-mass-max, --initial-mass-power
-
---mass-ratio-distribution, --mass-ratio, --mass-ratio-min, --mass-ratio-max, --minimum-secondary-mass
-
---eccentricity-distribution, --eccentricity, --eccentricity-min, --eccentricity-max
-
---metallicity-distribution, --metallicity, --metallicity-min, --metallicity-max
-
---orbital-period-distribution, --orbital-period, --orbital-period-min, --orbital-period-max, --semi-major-axis-distribution, --semi-major-axis, --semi-major-axis-min, --semi-major-axis-max, --allow-rlof-at-birth, --allow-touching-at-birth
-
---rotational-velocity-distribution, --rotational-frequency, --rotational-frequency-1, --rotational-frequency-2
-
 :ref:`Back to Top <options-props-top>`
 
 .. _options-stellar-evolution
 
 **Stellar evolution and winds**
-
[…]
