### 2021-11-08 aec8063
Added list of options by category
