+++ docs/online-docs/pages/User guide/Program options/temp.txt
@@ -1,951 +1,0 @@
-Program option list and default values
-======================================
-
-Any program options that are not specified take default values.
-
-- On the command line, program options that are not explicitly specified default to the COMPAS default value for the option (as specified in the COMPAS code - may be sampled from a distribution).
-
-- On a :doc:`grid file <../grid-files>` line, program options that are not explicitly specified default to the value specified for that option on the command line. If the program option was not explicitly specified on the command line, it will default to the COMPAS default value for the option, as described above. That is, the value for any option not specified on a grid file line option falls back to the value specified on the command line, which falls back to the COMPAS default if it was not specified on the command line.
-
-
-.. _options-props-top:
-
-The full list of program options with brief explanations and their default values is shown below.  We also include a listing of options (this time, by name only) grouped by category.
-
-**Jump to**
-:ref:`A <options-props-A>` :ref:`B <options-props-B>` :ref:`C <options-props-C>` :ref:`D <options-props-D>`
-:ref:`E <options-props-E>` :ref:`F <options-props-F>` :ref:`G <options-props-G>` :ref:`H <options-props-H>`
-:ref:`I <options-props-I>` :ref:`J <options-props-J>` :ref:`K <options-props-K>` :ref:`L <options-props-L>`
-:ref:`M <options-props-M>` :ref:`N <options-props-N>` :ref:`O <options-props-O>` :ref:`P <options-props-P>`
-:ref:`Q <options-props-Q>` :ref:`R <options-props-R>` :ref:`S <options-props-S>` :ref:`T <options-props-T>`
-:ref:`U <options-props-U>` :ref:`V <options-props-V>` :ref:`W <options-props-W>` :ref:`X <options-props-X>`
-:ref:`Y <options-props-Y>` :ref:`Z <options-props-Z>`
-
-**Jump to**
-:ref:`Initial conditions <options-initial-conditions>`
-:ref:`Stellar evolution and winds <options-stellar-evolution>`
-:ref:`Mass transfer physics <options-mass-transfer>`
-:ref:`Supernovae <options-supernovae>`
-:ref:`Administrative <options-admin>`
-
-COMPAS information
-------------------
-
-**--help [ -h ]** |br|
-Prints COMPAS help.
-
-**--version [ -v ]** |br|
-Prints COMPAS version string.
-
-
-Alphabetical listing
---------------------
-
-.. _options-props-A:
-
-**--add-options-to-sysparms** |br|
-Add columns for program options to SSE System Parameters/BSE System Parameters file (mode dependent). |br|
-Options: { ALWAYS, GRID, NEVER } |br|
-Default = GRID
-
-.. list-table::
-   :widths: 11 80 
-   :header-rows: 0
-   :class: aligned-text
-
-   * - ALWAYS
-     - indicates that the program options should be added to the sysparms file
-   * - GRID
-     - indicates that the program options should be added to the sysparms file `only if`
-   * -  
-     - a GRID file is specified, or RANGEs or SETs are specified for options
-   * - NEVER
-     - indicates that the program options should `not` be added to the sysparms file
-
-**--allow-rlof-at-birth** |br|
-Allow binaries that have one or both stars in RLOF at birth to evolve as over-contact systems. |br|
-Default = TRUE
-
-**--allow-touching-at-birth** |br|
[…]
