### 2021-11-08 #3838 @ilyamandel
@HeloiseS -- that's been pulled in.  I've now added a list of program options categorised by what they do as you suggested, to be included once https://github.com/TeamCOMPAS/COMPAS/pull/689 is approved.  I _think_ this addresses all of your must-implement comments so far -- hope I haven't missed anything!
