### 2021-11-08 6fc8b2a
updated chirpmass plot and made naming consistent with other notebooks
