### 2021-11-09 #3838 @HeloiseS
@jeffriley - finished reading the whole manual including the dev guide and there wasn't much more about debug classes so i jsut wanna make sure i udnerstand what you said ealier: Are the 'class strings' names of the classes within COMPAS, say I know my troubles are in BaseBinaryStar, then i give it the name of that class so i only get the debug statements from that class?
