### 2021-11-09 #3838 @jeffriley
@HeloiseS You're right, there's not a lot.  Did you see this:

https://compas.readthedocs.io/en/latest/pages/Developer%20guide/Services/Logging%20and%20debugging/Base-level%20logging/services-base-level-logging.html?highlight=class#log-debug-record-filtering

The guide is a work in progress - particularly the developer guide...
