### 2021-11-09 #3838 @HeloiseS
# Initial Review

Alright I finished reading the manual and running (or trying to) your tutorials and utility codes.
It's no surprise to me that COMPAS runs like clockwork and I'm very excited about all the data it records on the systems - so much to play with!

The manual is very exhaustive, the C++ code is a thing of absolute beauty - I can't code in C++ but the lay out, the comments and everything meant that I could navigate it easily as I was going through the Dev guide, locating the classes and functions mentioned. The jupyter notebooks have loads of info and some stupendous graphs to illustrate some of the tougher concepts, I love it.

Overall it's really amazing.

That's for the "good cop" part of this review, let's move on to the "bad cop" part.

## Bring user facing code to the same standard as the rest of the code

As much as the core of compas is refined and polished, a lot of the python utilities feel like an afterthought or something that's not been revised with the same level of care. My main request for this section is the following:
- [ ] Python utilities need to be brought to the same standard as the core of the simulation (more details below)

Now @ilyamandel  I know that you're going to say  _"that it's not the most urgent development priority at the moment"_ but my question would then be: if not priority in a JOSS review when your referee is asking you, then when will it be?  If you don't do it now you're going to be stuck with code that's on the side-lines of good practice, so I'm sorry but **this is non-negotiable, you gotta to fix the python**. Now for more details:

### Your Python should be Pythonic
I am not  being a PEP8 pedant,  this is a case where  The Principle of Least Astonishment really comes to mind - I have been genuinely confused by some of the constructions, and I don't believe I'm all that special. If I was confused someone else will be too ;)

- [ ]  **Refactor scripts and modules to ensure OO doesn't add complexity without adding functionality**
For example take the `pythonSubmit.py` (I told you it'd be back), pythonProgramOptions has no need to be class: the code runs top to bottom, there is no inheritance or actual need for encapsulation since we're not importing it in any other code... **it adds complexity without adding functionality** - plus when I see a class I unconsciously expect to be making objects with it in other places, in short  the construction is not coherent with the use case in the manual. The Dev guide has some _fantastic_ sections about the principles of Object Oriented Programming. Just make sure those are followed throughout COMPAS.

- [ ] **Ensure names of modules, classes don't break conventions and are sufficiently descriptive.**
Some pretty ubiquitous naming conventions were ignored: for example ClassCOMPAS is a **module**  but it looks like a class name in python, and then the class within it is `COMPASData`. An easy fix would be to call your class COMPASData and your module compas_data.py or compasdata.py. But  even then, you want to rethink that name because that's confusing in itself: it's NOT the COMPASData, its a processing pipeline. The name needs to reflect that.
(by the way I've noticed a lot of C like camel case naming conventions, even though that's not very python it's consistent with the rest of the code and so it'd be fine to adopt across the board).

- [ ] **Add all necessary Docstrings**
Docstrings need to be ubiquitous and also they need to show some consistency. Some functions have beautifully detailed docstrings (e.g. in `FastComsicIntegration.py`  or `selection_effects.py`, **the latter being a perfect example of what I mean by pythonic code**), some have pretty rudimentary ones, some have just plain comments, some have none. For example in the tutorial on COMPASData you have a cell with a markdown description of the parameters but the class itself doesn't have the info in docstrings.

     - [ ] Also if you have optional parameters give me the default in the description.

- [ ] **Be consistent with your docstring conventions**
You should pick a docstring style and stick to it. `selections_effects.py` uses numpy (my fave :relaxed:) which is excellent, but [there are plenty](https://stackoverflow.com/questions/3898572/what-is-the-standard-python-docstring-format). The Dev Guide has a whole section about style conventions but it's about the C++. Again, same level of care should go into your python, it could be a good time for you all to establish solid standards for your post processing / pre processing utilities.

### Other things

- [ ] **Group your utilities under one directory at the same level as the `src` directory**
Your utilities need to be grouped together under an easy to find (and easy to test, see below) directory. I would strongly recommend making it a package that can be installed, but even if you don't want to do that, at least have all the python utilities in one place.

- [ ] **tests**
not every line of python needs testing but have some tests for the main utilities that i can run on my machine on install to make sure it works. It is a requirement for the JOSS publication to have tests and all the workflows I see on the github at the moment only check that  COMPAS _compiles_ (it's possible I'm missing something, please let me know).  Now, running a stellar evolution simulation in continuous integration sounds pretty unhinged  so I'm not asking you to do that, but you can give us some thing to run locally - good thing is you already have a lot of that stuff in the tutorials! (For more about the tutos see below.) I'm happy to let you use the tutorials as test, by just telling us that explicitly in the install or the Get started section, that if we want to make sure all the utilities run we can just run the notebooks. That is unless @christinahedges says there must be unitests, in which case do that.

- [ ] **no bare exceptions**
Errors aren't like pokemon, you don't want to catch them all ;) after the `except` keywords make sure you explicitly state which error you're catching, otherwise you might end up with silent bugs. It's genuinely bitten me in the arse before so I strongly recommend you abide by best practices on this one.

## User Manual

- [ ] **Move the examples and the `Tutorial: Simple COMPAS Run`**
Streamline the example directory to only contain the necessary things for the tutorial and call it accordingly (something like tutorial_somethingsomething). Also move that tutorial earlier in the manual (maybe even at the end of the getting started) and you could also make it so it gives data we can use with the other tutorials (see below).

- [ ] **Tutorials could be much more locatable**
There are plenty of tutorials and they are a little scattered and also sometimes hidden. Now I totally get why they are where they are and maybe moving them would take to much restructuring of the manual, but it would be helpful to find a way to bring them up in the levels or flagging them clearly somewhere that people will see much sooner when they're reading.

- [ ] **Tutorials need to come with their data**
**I have not been able to run the majority of your tutorials (I might have additional comments late on that once it's sorted)**, because i don't know what data to use and the one I made with the examples is doesn't have the right stuff in it (see earlier messages). It would be nice if it was though :) it would make everything fit together quite nicely but you might say it's not possible or realistic (e.g. if the simulation takes more computational resources than is reasonable). In that case give me a direct link to where I can download the data.

- [ ] **Can you tell me early in the tutorial that the `logs` are where the _data_ go**  :)
It took me way too long to realise that when you were talking about logfiles and logs you were talking about data files, and once again I'm not special, I'm sure a fair few astronomers will not find that intuitive. No need to change the whole naming convention but really sign post very explicitly and very early on that that the data files are the log files. In my mind logs are for like administrative info like errors etc, so quite distinct from data files. See for illustration this epiphany I found in my notes "OMG THE LOG FILE RECORD SPECIFICATION PROPERTIES ARE THE DAMN COLUMN IN THE DATA SET". I wrote that nearly a week after starting the review, it really took me that long.

- [ ] **Some fundamental info should be flagged earlier**
"Since the seed is the unique identifier of each binary." I had to get to the middle of rewrite_5.py to see it written and that tutorial is quite buried but that seems like the kind of stuff you might want your user to know and understand quite earlier on. You do say in the Random Seed page that "each binary star will evolve using a unique random seed" but that didn't convey the notion that you use those as IDs, so maybe add a sentence there ;)

- [ ] **The random seeds are not always unique - the caveat needs to be flagged harder and best practice given**
So the random seeds are unique identifiers BUT at the bottom of the page on the random seeds you say:
> "Unfortunately there is no reasonable way to protect the user against specifying duplicate random seeds – especially since the random seed increments for each star or binary star."

Sounds like maybe it should be flagged harder and bolder (and maybe repeated in the tutos where you say the random seeds are used as unique IDs, give us the caveat).
How do you avoid this? It sounds like not mixing up the ways in which you call compas (like in the random seed example) would avoid most issues. If yes give us that directive at the top of the page, the rest of the page does a great job of explaining what's going on, but some people don't need the detail, they just need to know what's best practice.

### Rapid Fire:
- [ ] (docker from cl) "--number-of-binaries=5"  is not a viable option for me, I think changed to --number-of-systems
- [ ] still don't know what "--kick-mean-anomaly" means ^^"
- [ ] h5copy.py page: can i have a use case at the start of the page before going into the nitty gritty so I know the context in which I'm reading this? (maybe link to the tuto as well).
- [ ] step 4 in rewrite_H5.py has the comment 'DOUBLE CHECK ORDER :)' - does the order of these seeds matter?
- [ ] in the COMPAS HDF5 tools is `--filter` a regex filter or what?
- [ ] preProcessing is not mentioned in the user manual and it seems to have some old stuff in there - clean up needed.
- [ ] tuto 2 in Cosmic Integration: `redshift  = np.array([2]) \ agesBirth = redshift`. That makes `redshift` and `agesBirth` the same variable (it's a pointer, they'll have the same memory address) not a copy. Is that on purpose or where you just trying to give `agesBirth` the same kind of dummy value?
- [ ] tuto 3 in Cosmic Integration reference should be flagged earlier?
"If you use this pipeline we would appreciate it if you cite Selection effects https://arxiv.org/pdf/1711.06287" | that might be missed and is worth repeating  - if you have a few applications like that which require different referencing, have a section of your read me and your readthedocs telling users what they need to reference depending on what they end up using. Get those citations yeah!
- [ ] some page of the readthedocs should really be subsections and put together on the same page. That happens both in the user guide and the dev guide but the clearest examples are in the later, e.g. "Random number service" and other such pages where the functions and macros have individual pages when they really don't need to (it's not just easthetics, it makes that stuff harder to find when we need to dig through 4 or 5 levels), similarly the OOP section can be just one big page.
- [ ] Rooky question, what's the right arrow in the commands like LOGGING→Enabled(); is it a tab?
- [ ] SQUWAK :laughing: what's the difference in use btw SQUAWK and SAY ( also I see it's on the todo int he Log.cpp: clean up use of Squawk() vs SAY())
- [ ] Error handling service: "error marcros are in ErrorsMacros.h"  -> i see many many more things in that .h than the 4 things on that page i'm a little confused but maybe it's because i dont c++
- [ ] MAJOR RELEASES VS MINOR RELEASES. Do you also consider backwards compatibility? In some software minor releases ensure backwards compatibility .e.g. my pythonSubmit for v2.25 should run the same on 2.26 (even if the pythonSubmit for v2.26 might not run for v2.25 because the older version is missing a feature), whereas major releases indicate it might break your code. That's not always the case at all but i was just wondering.

### Debug classes
Thanks for the reply Jeff, and for clarification i deleted the small post because I was going to post the question within this larger one, I'm putting it back here so things make sense

> @jeffriley - finished reading the whole manual including the dev guide and there wasn't much more about debug classes so i jsut wanna make sure i udnerstand what you said ealier: Are the 'class strings' names of the classes within COMPAS, say I know my troubles are in BaseBinaryStar, then i give it the name of that class so i only get the debug statements from that class?

And to answer your question, yes I did, and since you're asking I'm guessing my interpretation is not correct lmao. And the dev guide is actually super good, I'm just a noob i that domain, so I'm afraid you're going to have to explain like I'm 5 :laughing: because it clearly went over my head.

----
