### 2021-11-09 #3838 @jeffriley
> some page of the readthedocs should really be subsections and put together on the same page.

That's probably true (i a number of places).  As I said, a work in progress... The online documentation was put together fairly recently, and we still have to go through and rationalise it.  You review will help us do that :-)

> Rooky question, what's the right arrow in the commands like LOGGING→Enabled(); is it a tab?

In C/C++ (and other languages), dot-notation indicates object membership (i.e. 'a.x' means 'x' in object 'a'). The arrow notation (->) is used for pointers and derefences first: 'a->x' means x in the object pointed to by 'a').  See https://www.cplusplus.com/doc/tutorial/pointers/ for a discussion of pointers (a pointer variable doesn't contain the value of the variable, it contains the memory address of the value of the variable (it "points to" the variable)).

> what's the difference in use btw SQUAWK and SAY

SQUAWK() writes to stderr; SAY() writes to stdout.  Generally, the idea is that SAY() is for information; SQUAWK() is to complain or report an error condition.

> Error handling service: "error marcros are in ErrorsMacros.h" -> i see many many more things in that .h than the 4 things on that page

There are 4 macros: SHOW_ERROR, SHOW_ERROR_IF, SHOW_WARN, and SHOW_WARN_IF - and then there are static versions of each.

Each of the 4 macros has 7 lines in ErrorsMacros.h - the multiple lines support calling the macros with a variable number of arguments (the '_#' indicates the number of arguments - the code to the right shows how the arguments are treated).
