### 2021-11-09 #3838 @ilyamandel
@HeloiseS :

> What data am I expected to use for the Cosmic Integration tutorials?

> Reason I'm asking is that i don't have a COMPAS output with 'BSE_Double_Compact_Objects' it seems. I used the output from the chirpmass distribution example (expecting that I'd defo have DCOs in there) and that didn't work.

Not sure about the chirp mass distribution example (sorry, I haven't tried it, though I agree with your surprise -- will check).  In the meantime, you can use one of the recent COMPAS data sets from zenodo, http://zenodo.org/communities/compas/ , e.g., https://zenodo.org/record/5544170#.YYoihb1Bxp8 .
