### 2021-11-09 #3838 @ilyamandel
Ahh, I see that Lieke has a pending PR on the chirp mass distribution. ;-)  https://github.com/TeamCOMPAS/COMPAS/pull/692/files
