### 2021-11-09 TeamCOMPAS/COMPAS#686 @ilyamandel
@jeffriley , @reinhold-willcox --
It seems that all of the information in https://github.com/TeamCOMPAS/COMPAS/blob/dev/docs/getting_started.md is indeed repeated in the new readthedocs documentation at https://compas.science/docs/.  However, there are several other .md files in https://github.com/TeamCOMPAS/COMPAS/tree/dev/docs which contain information not available at readthedocs as far as I can tell.  I think we should have documentation in one place, though, and they are too hidden here.  Can you please check whether any of this information should be transferred to readthedocs before these are deleted?
