+++ postProcessing/Tutorial/DataAnalysis/2_visualisingData.py
@@ -1,0 +1,192 @@
+# ---
+# jupyter:
+#   jupytext:
+#     formats: ipynb,py:light
+#     text_representation:
+#       extension: .py
+#       format_name: light
+#       format_version: '1.5'
+#       jupytext_version: 1.12.0
+#   kernelspec:
+#     display_name: Python 3
+#     language: python
+#     name: python3
+# ---
+
+# # Introduction
+#
+# Although math is the fundamental basis of physics and astrophysics, we cannot always easily convert numbers and equations into a coherent picture. Plotting is therefore a vital tool in bridging the gap between raw data and a deeper scientific understanding. 
+#
+# *Disclaimer:*
+#
+# There are many ways to make the same plot in matplotlib and there are many ways to bin your data. Often, there is no "best" way to display data in a plot, and the message conveyed can be heavily dependent on the context of the data as well as asthetic plotting decisions.
+#
+# For example, in histograms, as we discuss below, the relatively subjective choice of bin size can significantly affect the interpretation of the results. It is important to be aware of when and how we make these choices and to try to reduce any unintended bias.
+#
+# ---
+#
+# ** Example: inspect the component masses of Double Compact Objects**
+#
+# In the example below, we use the following conventions:
+#
+# 1 - We deliberately choose to use the matplotlib.pyplot.subplots routine even when creating a single figure (as opposed to using pyplot.plot). This is because many online forums (e.g Stackoverflow) use this syntax. Furthermore, this means you do not have to learn two different types of syntax when creating either a single or multiple panel figure.
+#
+# 2 - We choose to do the binning within the numpy/array environment instead of with inbuilt functions such as plt.hist / axes.hist. The reason is that you have more control over what you do, such as custom normalization (using rates, weights, pdf, etc.). It also forces you to have a deeper understanding of what you are calculating, and allows you to check intermediate steps with print statements.  Once you know how to bin your data this way you can also easily expand these routines for more complicated plots (2D binning).
+
+# # Path to be set by user
+#
+
+pathToData = '/home/cneijssel/Desktop/Test/COMPAS_output.h5'
+
+# # Imports
+
+# Python libraries
+import numpy as np               #for handling arrays
+import h5py as h5                #for reading the COMPAS data
+import matplotlib.pyplot as plt  #for plotting
+# %matplotlib inline
+
+# # Get some data to plot
+
+# +
+Data  = h5.File(pathToData)
+
+print(list(Data.keys()))
+# DCOs = double compact objects
+
+
+DCOs = Data['DoubleCompactObjects']
+
+M1   = DCOs['Mass_1'][()]
+M2   = DCOs['Mass_2'][()]
+Mtot = np.add(M1, M2)
+# -
+
+Data.close()
+
+# # Histogram
+
+# +
+# You can use numpy to create an array with specific min, max and interval values
+minMtot = 0
+maxMtot = max(Mtot)
+nBins   = 50
+
+# Number of bin edges is one more than number of bins
+binEdges = np.linspace(minMtot, maxMtot, nBins+1)
+
+# What is the value at the center of the bin?
+# add each edge of the side of the bin and divide by 2
+xvaluesHist  = (binEdges[:-1] + binEdges[1:])/2.
+
+# What is the width of each bin? (an array in general, if the spacing is non-uniform)
[…]
