+++ postProcessing/Tutorial/WorkingWithHDF5/WorkingWithHDF5.py
@@ -1,0 +1,240 @@
+# ---
+# jupyter:
+#   jupytext:
+#     formats: ipynb,py:light
+#     text_representation:
+#       extension: .py
+#       format_name: light
+#       format_version: '1.5'
+#       jupytext_version: 1.12.0
+#   kernelspec:
+#     display_name: Python 3
+#     language: python
+#     name: python3
+# ---
+
+# # Introduction
+#
+# COMPAS simulations produce all output by default in the form of an [HDF5 file](https://www.hdfgroup.org/solutions/hdf5/), which is a compact and memory efficient, but non-human-readable data file format. In order to interact with these output files, we use the python module `h5py`. 
+#
+# Within each file are a variety of HDF5 Groups, representing a specific common event in binary evolution, e.g Roche-Lobe Overflow or Supernovae. These are described throughout the post-processing jupyter notebooks.
+
+# ## Material
+#
+# ### [1. Producing HDF5 output:](#1.-Producing-HDF5-output)
+# How to run default COMPAS and produce a simple output file.
+#         
+#         
+# ### [2. Reading HDF5 files:](#2.-Reading-HDF5-files)
+# The basics and syntax of loading an HDF5 file.
+#         
+#         
+# ### [3. Rewriting HDF5 files:](#3.-Rewriting-HDF5-files)
+# How to rewrite/reduce the HDF5 data.
+
+
+
+# +
+import os, sys    # for handling paths
+import h5py as h5  #for handling data format
+
+# Import COMPAS root and python script directories
+compasRootDir = os.environ['COMPAS_ROOT_DIR'] 
+sys.path.append(compasRootDir + 'postProcessing/PythonScripts')
+
+# Compas custom scripts used here
+from printCompasDetails import printCompasDetails
+import rewrite_H5
+
+
+# -
+
+# # 1. Producing HDF5 output
+#
+# ## Here we show an example of how to run a basic COMPAS simulation in order to produce output in an HDF5 format. To run COMPAS, you need to set the environment variable $COMPAS_ROOT_DIR to the top level COMPAS directory. Ensure that you have compiled COMPAS, and that there is a COMPAS executable. We will store the output in the Tutorial directory for easy access later on.
+#
+# ## For simplicity, we will run COMPAS on all default settings except for the number of binaries produced. Here, we will run 1k binaries, which is relatively small. Later, we will look at production of double compact objects (DCOs), but these objects are sufficiently rare that in order to study them, we will need to run a simulation of ~1M binaries, but this can take some time depending on your hardware resources.
+#
+# ### *Note:* It is required to have some COMPAS output to complete the rest of the post-processing notebooks, but if you already have an output file, you can skip this section.
+
+# To run terminal commands in a jupyter notebook, prepend the command with an exclamation mark
+# !COMPAS -n 1000 -o {outputDir}
+
+# ### *Note:* This will produce output in the COMPAS_Output directory only the first time it is run. If you run this multiple times, new directories will be created as COMPAS_Output_X. Ensure that you are using the data in the desired directory.
+
+# ## Non-default settings can be seen by running the following in the command line
+
+# !COMPAS --help
+
+
+
+# # 2. Reading HDF5 files
+#
+# ## Here we show the basic h5 file syntax for how to load and close the file, and extract/inspect the data.
+#
+# ### For this and the following sections, you will need to have the python package h5py installed. 
+#
+# ### Running `pip install h5py` on the command line is one way to do this.
+
+# ## Load the h5 file
+
+# +
+# Set the appropriate path to the data file
+
[…]
