+++ postProcessing/Tutorial/DataAnalysis/0_BasicsOverview.py
@@ -1,0 +1,32 @@
+# ---
+# jupyter:
+#   jupytext:
+#     formats: ipynb,py:light
+#     text_representation:
+#       extension: .py
+#       format_name: light
+#       format_version: '1.5'
+#       jupytext_version: 1.12.0
+#   kernelspec:
+#     display_name: Python 3
+#     language: python
+#     name: python3
+# ---
+
+# # Introduction
+#
+#
+# Most of the post-Processing material here uses python code.
+# Here we show two important basics of python code in the context
+# of investigating your COMPAS simulation.
+#
+#
+# ## Material
+#
+# ### [Slicing the Data](./1_slicingData.ipynb)
+# How to select specific systems and their parameters using seeds.
+#
+# ### [Visualising the Data](./2_visualisingData.ipynb)
+# The basics of binning and visualising your data.
+
+
