### 2021-11-12 #3838 @HeloiseS
@ilyamandel  update -- my vaccine failed to properly install my 5G I think and I can't download the data set you recommended (I'm looking at DL speeds of <1GB/hour). If you have a recommendation for a smaller dataset I'll take it otherwise I'll just try to make a little simulation with all the things I need to run the tutos and play with them :)
