### 2021-11-12 f153401
EAGB luminosity and radius fix
