### 2021-11-12 782b3cc
EAGB luminosity and radius fix
