### 2021-11-12 5f87fe5
Update changelog.h
Changed "section 4 of Hurley+, 2000" to "section 5.4 of Hurley+, 2000"
