### 2021-11-12 #3838 @katiebreivik
Hi Everyone -- apologies for my radio silence so far! I've been swamped with other tasks but wanted to give a heads up that I will be focusing on finishing up my review next week. One quick thing I wanted to bring up for discussion is the need for 100% consistent docstrings in all the code. In the review criteria (https://joss.readthedocs.io/en/latest/review_criteria.html#api-documentation) the absolute requirement is that the API be documented, which I think is satisfied at the OK level at present (though I must confess I haven't gone 100% through the docs yet; next week though!). I agree with @HeloiseS that this is important, though, and would advocate for standardized docstrings. At the same time, I want to make sure that we aren't pushing too far past the JOSS requirements. Very happy to have a convo about it here though!

Apologies again for my slowness!
