### 2021-11-12 #3838 @HeloiseS
Agreed, the formatting standadarsation is a looser request - I do strongly recommend it because I can see COMPAS growing a lot in the future as more users get on it and the standards you establish now (and I am not just talking about docstrings) are things you're going to be carrying with you long term - the C++ has a solid and explicit set of standards and good practices followed throughout, and tbh these are of a higher standard than would be required with JOSS, but I still think it's worth the devs use this opportunity to think along similar lines for their post processing python pipelines. At the end of the day it's up to the dev team to decide whether they want to engage with that at the current time, and if they choose not to I won't make a fuss ;)

Going back to docstrings I do think there is a need to have complete docstrings in the core pipelines (and that is something I would consider a requirement, rather than a request) such as COMPASData though. Besides it's quick to copy what's in the notebook already into a dosctring, so it's a low "cost" / high impact improvement :)
