### 2021-11-13 #3838 @ilyamandel
> update -- my vaccine failed to properly install my 5G I think and I can't download the data set you recommended (I'm looking at DL speeds of <1GB/hour 😢 and it keeps giving me an error and failing ). If you have a recommendation for a smaller dataset I'll take it otherwise I'll just try to make a little simulation with all the things I need to run the tutos and play with them :)

@HeloiseS: If I change
number_of_systems = 5000
in the standard pythonSubmit file (say, the one in examples/methods_paper_plots/chirpmass_distribution), I get ~20 DCOs in the output file in a run that takes under 1 minute.  That's not enough for pretty plots, but enough to get  examples to run without failing.

The output file is still large enough (129 MB) that it will probably be faster for you to run it yourself than to download it, until your vaccine booster gives you superpowers or New Zealand broadband enters the 21st century. ;-)
