### 2021-11-13 #3838 @ilyamandel
@HeloiseS, @katiebreivik -- a lot of Heloise's comments [thank you for the very nice list!] are about the python scripts.

You are absolutely correct that they are not at a consistent level of code quality, maintenance, and documentation.  Part of the problem is that many of the developers don't actually use these. :)  Because the processing tools are very problem-specific, so we tend to develop our own on an as-needed basis (sometimes not even in python -- e.g., personally, I do most of my processing in Matlab ;-) ).

We thought that making public some of the python scripts behind, say, the figures in the ApJS-submitted methods paper would be useful.  However, these are being provided on an as-available basis.  Realistically, we do not currently have the human resources to guarantee their quality and documentation at the requested level.  And I certainly understand that providing bad/confusing/dysfunctional examples may be worse than providing none at all.

So, my question is, what is the best approach?  I would be quite happy writing in big red letters that the only processing files we are providing as part of the core COMPAS distribution and take full responsibility for are h5copy.py, selection_effects.py, and FastCosmicIntegration.py.  Everything else (including pythonSubmit.py) are just examples that the user may find helpful, but if not, they are not unnecessary for single or binary evolution or for processing the results.  Would that be better?  Would you prefer if we moved all other .py files into a different repository to avoid confusion?
