### 2021-11-13 #3838 @HeloiseS
Let's not throw the baby out with the bath water ;)

> Because the processing tools are very problem-specific, so we tend to develop our own on an as-needed basis

So like you suggested it might be worth differentiating between the python processing utilities that you anticipate users will need ont he regular and have those in tutorials, and then the stuff that reproduces paper results and are very specific could go in a separate section where it's not flagged as a utility but as a "here is the code we used for this"?
