### 2021-11-15 #3838 @HeloiseS
@ilyamandel @jeffriley - is there an native option to make COMPAS use X cores that I missed  or do i need to do that by calling it in a bash script or something (or maybe does it do it automatically but i'm just not running enough models for that to trigger).
