### 2021-11-15 #3838 @jeffriley
@HeloiseS COMPAS is not threaded, but it doesn't need to be.  You can just run X standalone runs on X cores, and provided you used appropriate starting random seeds (so there's no overlap between runs), you can just join all the output files together after the runs have completed.  If you use HDF5 output files, the use h5copy.py to join the output files into a single HDF5 files.

I think we're testing a script now that will split a big run into several smaller runs - but I don't think it automatically joins the output files.
