### 2021-11-15 TeamCOMPAS/COMPAS#695 @HeloiseS
Commandline Options error: Missing ']' for option 'metallicity'
_This issue is submitted as part of the JOSS review [#3838](https://github.com/openjournals/joss-reviews/issues/3838)_

**:bug:  Description**
So I'm trying to make models for a range of metallicities following the instructions on [this page](https://compas-docs-readthedocs.readthedocs.io/en/latest/program-options-ranges.html).

The command I'm trying to parse is as follows (note i have aliased the location of COMPAS to `compas`):
`compas --metallicity [0.0012, 3, 0.01] --number-of-systems 5000
`
I tried with having the `r`and `range` in front of the square brackets to no avail.
I tried within the pythonSubmit.py file with `metallicity = [0.0112, 3, 0.01]` and got the same error.

EDIT: your example works though, so I clearly did something wrong, just not sure what :)
