### 2021-11-15 TeamCOMPAS/COMPAS#695 @jeffriley
@HeloiseS Leave the spaces out inside the [ ].  If you have spaces there the shell parser thinks it's a new command-line argument (command-line arguments are space-delimited).  COMPAS can't do anything about that - the shell parser does it's work before COMPAS gets to see the options.  I thought I documented that in the online documentation, but I can't find it.  I'm sure it was in the PDF documentation (the predecessor to the online documentation), so I must have inadvertently left it out.  I'll update the documentation.  Please leave this issue open, and I'll close it once I have updated the documentation.

I note that I do document the spaces issue for vector options here:

https://compas.readthedocs.io/en/latest/pages/User%20guide/Program%20options/program-options-vector-options.html

Edit:  Also, if you want to specify ranges and sets in pythonSubmit, you need to put quotes around the range/set value.  i.e.

This will fail in pythonSubmit:

    metallicity = [0.0112,3,0.01]

but this will work:

    metallicity = '[0.0112,3,0.01]'

Python will try to parse the first one, and fail, but for the second it will pass the quoted parameter through to the command line.

This is documented here:

https://compas.readthedocs.io/en/latest/pages/User%20guide/Running%20COMPAS/running-via-python.html

(maybe it should also be on t he ranges/sets pages).

(Note to self: also fix "correspodning" in program-options-vector-options.rst)
