### 2021-11-15 TeamCOMPAS/COMPAS#695 @HeloiseS
Cheers!

It now runs with the spces removed but when i run `compas --metallicity [0.012,3,0.001] -n 5000` i only get 3 models. I thought it  would give me 5000 models per mets but it seems to ignore the -n? switching order doesn't help is there priorities in commnds i need to consider?
