### 2021-11-15 TeamCOMPAS/COMPAS#695 @jeffriley
I did that because the ranges and sets were meant to make it easy to specify a range (and set) of initial conditions. and if the user specified ranges and sets for multiple options (e.g. metallicity, initial-mass-1, initial-mass-2, etc. - the ones that otherwise would be sampled by COMPAS), then also using -n/--number-of-systems would just produce multiple copies of the same binaries - so I ignore -n/--number-of-systems.  I probably didn't document that, did I...? I'll fix that.

However, if you use a grid file, then the ranges and sets on the command-line will be played out for each grid file line.

So this:

./compas --metallicity [0.012,3,0.001] --eccentricity s[0.1,0.2,0.3] --grid mygrid.txt

and 5 lines in mygrid.txt

will result in 45 binaries (3 x 3 x 5)
