### 2021-11-15 TeamCOMPAS/COMPAS#698 @HeloiseS
[h5copy.py] Dataset count mismatch between source and destination for group ....
_This issue is submitted as part of the JOSS review [#3838](https://github.com/openjournals/joss-reviews/issues/3838)_

I'm trying to combine two datasets of 10,000 systems each (the only difference is the metallicity) and I get an error I can't really diagnose from `h5copy.py`. Here is what I run and the trace.

```
fste075@kvin:~/Documents/reviews/COMPAS$ python h5copy.py . -r
Processing directory /home/fste075/Documents/reviews/COMPAS
Processing directory /home/fste075/Documents/reviews/COMPAS/COMPAS_Output_1
Copying file /home/fste075/Documents/reviews/COMPAS/COMPAS_Output_1/COMPAS_Output.h5
Processing directory /home/fste075/Documents/reviews/COMPAS/COMPAS_Output
Copying file /home/fste075/Documents/reviews/COMPAS/COMPAS_Output/COMPAS_Output.h5
Dataset count mismatch between source and destination for group BSE_Double_Compact_Objects in files /home/fste075/Documents/reviews/COMPAS/COMPAS_Output/COMPAS_Output.h5 and /home/fste075/Documents/reviews/COMPAS/h5out.h5
WARNING: Group BSE_Double_Compact_Objects from file /home/fste075/Documents/reviews/COMPAS/COMPAS_Output/COMPAS_Output.h5 not copied to /home/fste075/Documents/reviews/COMPAS/h5out.h5
Dataset count mismatch between source and destination for group BSE_System_Parameters in files /home/fste075/Documents/reviews/COMPAS/COMPAS_Output/COMPAS_Output.h5 and /home/fste075/Documents/reviews/COMPAS/h5out.h5
WARNING: Group BSE_System_Parameters from file /home/fste075/Documents/reviews/COMPAS/COMPAS_Output/COMPAS_Output.h5 not copied to /home/fste075/Documents/reviews/COMPAS/h5out.h5
^CTraceback (most recent call last):
  File "h5copy.py", line 634, in <module>
    main()
  File "h5copy.py", line 610, in main
    ok = processDirectory(thisFullPath,
  File "h5copy.py", line 498, in processDirectory
    processDirectory(absDirpath + '/' + dirname,
  File "h5copy.py", line 484, in processDirectory
    ok = copyHDF5File(absDirpath + '/' + filename,
  File "h5copy.py", line 381, in copyHDF5File
    srcDataset       = srcGroup[srcDatasetName]                                                 # source dataset
  File "h5py/_objects.pyx", line 54, in h5py._objects.with_phil.wrapper
  File "h5py/_objects.pyx", line 55, in h5py._objects.with_phil.wrapper
  File "/home/fste075/.local/lib/python3.8/site-packages/h5py/_hl/group.py", line 264, in __getitem__
    oid = h5o.open(self.id, self._e(name), lapl=self._lapl)
  File "h5py/_objects.pyx", line 54, in h5py._objects.with_phil.wrapper
  File "h5py/_objects.pyx", line 55, in h5py._objects.with_phil.wrapper
  File "h5py/h5o.pyx", line 190, in h5py.h5o.open
  File "h5py/h5i.pyx", line 40, in h5py.h5i.wrap_identifier
  File "<frozen importlib._bootstrap>", line 389, in parent
KeyboardInterrupt

```
The keybooard interupt is because after this my computer proceeds to have a bit of a stroke :)
I get hundreds of messages like the ones below. Maybe it'd be a good idea to limit how many of these messages can pop up on the terminal? The first time I keyboard interrupted after maybe 15 seconds and couldn't scroll back up to the top of the Traceback ^^"
```
Error accessing group Age(1) in file /home/fste075/Documents/reviews/COMPAS/h5out.h5 : 'Dataset' object has no attribute 'keys'
Error accessing group Age(2) in file /home/fste075/Documents/reviews/COMPAS/h5out.h5 : 'Dataset' object has no attribute 'keys'
Error accessing group Ang_Momentum(1) in file /home/fste075/Documents/reviews/COMPAS/h5out.h5 : 'Dataset' object has no attribute 'keys'
Error accessing group Ang_Momentum(2) in file /home/fste075/Documents/reviews/COMPAS/h5out.h5 : 'Dataset' object has no attribute 'keys'
Error accessing group Ang_Momentum_Total in file /home/fste075/Documents/reviews/COMPAS/h5out.h5 : 'Dataset' object has no attribute 'keys'
Error accessing group Dominant_Mass_Loss_Rate(1) in file /home/fste075/Documents/reviews/COMPAS/h5out.h5 : 'Dataset' object has no attribute 'keys'
Error accessing group Dominant_Mass_Loss_Rate(2) in file /home/fste075/Documents/reviews/COMPAS/h5out.h5 : 'Dataset' object has no attribute 'keys'
Error accessing group Eccentricity in file /home/fste075/Documents/reviews/COMPAS/h5out.h5 : 'Dataset' object has no attribute 'keys'
Error accessing group Energy_Total in file /home/fste075/Documents/reviews/COMPAS/h5out.h5 : 'Dataset' object has no attribute 'keys'
Error accessing group Luminosity(1) in file /home/fste075/Documents/reviews/COMPAS/h5out.h5 : 'Dataset' object has no attribute 'keys'
Error accessing group Luminosity(2) in file /home/fste075/Documents/reviews/COMPAS/h5out.h5 : 'Dataset' object has no attribute 'keys'
```
