### 2021-11-16 4e88648
merge main and add wrapper function for calculating Nanjing lambda depending on options
