+++ src/HeWD.h
@@ -60,7 +60,8 @@ protected:
     double          CalculateHeCoreMassOnPhase()                                                        { return m_HeCoreMass; }                                                        // NO-OP
 
     double          CalculateLambdaDewi()                                                               { return BaseStar::CalculateLambdaDewi(); }                                     // Not supported - use BaseStar
-    double          CalculateLambdaNanjing(const int p_MassInd, const int p_Zind)                       { return BaseStar::CalculateLambdaNanjing(p_MassInd, p_Zind); }                 // Not supported - use BaseStar     JR: todo: check this (type 10 not mentioned as not supported in original code)
+    double          CalculateLambdaNanjing()                                                            { return BaseStar::CalculateLambdaNanjing; }                 // Not supported - use BaseStar     JR: todo: check this (type 10 not mentioned as not supported in original code)
+    double          CalculateLambdaNanjingEnhanced()                                                    { return CalculateLambdaNanjing; }
 
     double          CalculateLuminosityOnPhase(const double p_Mass,
                                                const double p_Time,
