+++ src/HG.cpp
@@ -69,27 +69,17 @@ double HG::CalculateLambdaDewi() {
 
 
 /*
- * Calculate the common envelope lambda parameter using the "Nanjing" prescription
- * from X.-J. Xu and X.-D. Li arXiv:1004.4957 (v1, 28Apr2010) as implemented in STARTRACK
- *
- * This implementation adapted from the STARTRACK implementation (STARTRACK courtesy Chris Belczynski)
- *
- *
- * JR: todo: the coefficients and factors here are hard-coded until I figure out an
- * efficient way of putting them in constants.h.  Because they are indexed by a few
- * things: stellar type, metallicity and ZAMS mass the easiet thing would be to  put
- * them in a map - but because they can be re-calculated at every timestep the hashing
- * overhead becomes a performance concern.  Vectors are a good alternative, but I need
- * to figure out how best to structure them for reasonable (and inuitive) access.
+ * Calculate the common envelope lambda parameter using the enhanced "Nanjing" prescription
+ * from X.-J. Xu and X.-D. Li arXiv:1004.4957 (v1, 28Apr2010)
  *
  * This function good for HG and FGB stars.
  *
  *
- * double CalculateLambdaNanjing(const int p_MassInd, const int p_Zind)
+ * double CalculateLambdaNanjingEnhanced(const int p_MassInd, const int p_Zind)
  *
  * @return                                      Nanjing lambda for use in common envelope
  */
-double HG::CalculateLambdaNanjing(const int p_MassInd, const int p_Zind) {
+double HG::CalculateLambdaNanjingEnhanced(const int p_MassInd, const int p_Zind) {
 
 	DBL_VECTOR maxBG    = {};                                                           // [0] = maxB, [1] = maxG
 	DBL_VECTOR lambdaBG = {};                                                           // [0] = lambdaB, [1] = lambdaG
@@ -368,6 +358,331 @@ double HG::CalculateLambdaNanjing(const int p_MassInd, const int p_Zind) {
     // Calculate lambda as some combination of lambda_b and lambda_g by
     // lambda = alpha_th • lambda_b    +  (1-alpha_th) • lambda_g
     // STARTRACK uses alpha_th = 1/2
+    return (OPTIONS->CommonEnvelopeAlphaThermal() * lambdaBG[0]) + ((1.0 - OPTIONS->CommonEnvelopeAlphaThermal()) * lambdaBG[1]);
+}
+
+
+/*
+ * Calculate the common envelope lambda parameter using the "Nanjing" prescription
+ * from X.-J. Xu and X.-D. Li arXiv:1004.4957 (v1, 28Apr2010) as implemented in STARTRACK
+ *
+ * This implementation adapted from the STARTRACK implementation (STARTRACK courtesy Chris Belczynski)
+ *
+ *
+ * JR: todo: the coefficients and factors here are hard-coded until I figure out an
+ * efficient way of putting them in constants.h.  Because they are indexed by a few
+ * things: stellar type, metallicity and ZAMS mass the easiet thing would be to  put
+ * them in a map - but because they can be re-calculated at every timestep the hashing
+ * overhead becomes a performance concern.  Vectors are a good alternative, but I need
[…]
