### 2021-11-16 501a84b
merge main and add wrapper function for calculating Nanjing lambda depending on options
