### 2021-11-16 TeamCOMPAS/COMPAS#698 @jeffriley
@HeloiseS Can you upload the two HDF5 files somewhere that I can grab them?
