### 2021-11-16 926ea53
Defect repair: Revert EAGB treatment to 02.25.06 until a proper fix is introduced
