### 2021-11-16 #3838 @christinahedges
Hello everyone, just wanted to let you know that I am reading with interest, I appreciate all the effort going into these discussions. If anyone has any questions specifically about JOSS standards please do let me know.
