### 2021-11-16 TeamCOMPAS/COMPAS#698 @jeffriley
@HeloiseS Thanks for providing the hdf5 files.  I took a look at them with h5view.py:

    $ python3 /compas/dev/my_fork/compas/postProcessing/Folders/H5/PythonScripts/h5view.py ./COMPAS_Output_Heloise_1.h5

    Summary of HDF5 file /d/compas/heloise/COMPAS_Output_Heloise_1.h5
    =================================================================

    File size    : 0.1149 GB
    Last modified: 2021-11-17 06:53:11.033530

    COMPAS Filename              Columns   Entries   Unique SEEDs
    --------------------------   -------   -------   ------------
    Run_Details                      360         1
    BSE_Common_Envelopes              73      4660           4457
    BSE_Double_Compact_Objects        12        77             77
    BSE_RLOF                          33     24910           5525
    BSE_Supernovae                    32      4270           4107
    BSE_System_Parameters             29     10000          10000

    $ python3 /compas/dev/my_fork/compas/postProcessing/Folders/H5/PythonScripts/h5view.py ./COMPAS_Output_Heloise_2.h5

    Summary of HDF5 file /d/compas/heloise/COMPAS_Output_Heloise_2.h5
    =================================================================

    File size    : 0.1201 GB
    Last modified: 2021-11-17 06:54:59.209393

    COMPAS Filename              Columns   Entries   Unique SEEDs
    --------------------------   -------   -------   ------------
    Run_Details                      360         1
    BSE_Common_Envelopes              73      3314           3217
    BSE_Double_Compact_Objects        18        65             65
    BSE_RLOF                          33     21697           5365
    BSE_Supernovae                    32      7494           7327
    BSE_System_Parameters             32     10000          10000

You can see that the number of columns (datasets for HDF5 files) differs for both BSE_Double_Compact_Objects and BSE_System_Parameters.  h5copy.py won't copy groups (BSE_Double_Compact_Objects is a group in HDF5 nomenclature) to a file that has a different number of datasets for that group.  There are reasons for that - if we're copying to an existing file (which we will be at some stage when we're joining multiple of files), and that existing file has fewer datasets in a group (aka columns in a file) than the file being copied to it, there is no place to put the new datasets (columns), and even if we create them in the existing file on-the-fly, how would we populate them for the existing file?  Similarly, if the existing file has datasets in a group (columns in a file) that the file being copied to it doesn't have, what values should be given the the dataset values missind from the file being copied?  So we just issue a warning and don't copy those groups...

This behaviour is documented in the online documentation (under the heading "Appending data":

https://compas.readthedocs.io/en/latest/pages/User%20guide/Post-processing/post-processing-h5copy.html

Did you create the files?  If so, did you use a logfile-definitions file to change what was written to the output files?

Edit: Yes, I see you did for one of the files - logfile-definitions file "/home/fste075/Documents/reviews/COMPAS/COMPAS_Output_Definitions.txt"
