### 2021-11-16 TeamCOMPAS/COMPAS#698 @jeffriley
Having said that (^^^) though, you should only see that message once per group in any affected HDF5 - so in your case you should have only seen the error twice for the second file.

And... I just re-read the issue - you're getting lots of errors about missing keys.  Let me look into that - I suspect it's a ricochet from t he mismatch problem (but it shouldn't be...).
