### 2021-11-16 TeamCOMPAS/COMPAS#698 @jeffriley
Ok, this is what I get when I do the copy:

    $ python3 /compas/postProcessing/Folders/H5/PythonScripts/h5copy.py ./files
    Processing directory /d/compas/heloise/files
    Copying file /d/compas/heloise/files/COMPAS_Output_Heloise_1.h5
    Copying file /d/compas/heloise/files/COMPAS_Output_Heloise_2.h5
    Dataset count mismatch between source and destination for group BSE_Double_Compact_Objects in files /d/compas/heloise/files/COMPAS_Output_Heloise_2.h5 and /d/compas/heloise/h5out.h5
    WARNING: Group BSE_Double_Compact_Objects from file /d/compas/heloise/files/COMPAS_Output_Heloise_2.h5 not copied to /d/compas/heloise/h5out.h5
    Dataset count mismatch between source and destination for group BSE_System_Parameters in files /d/compas/heloise/files/COMPAS_Output_Heloise_2.h5 and /d/compas/heloise/h5out.h5
    WARNING: Group BSE_System_Parameters from file /d/compas/heloise/files/COMPAS_Output_Heloise_2.h5 not copied to /d/compas/heloise/h5out.h5

It doesn't matter what order I do the copy - same result.  Can you try to reproduce and let me know if you can.  Are the files you provided the exact files you tried to join with h5copy.py?  I guess they are - maybe your version of h5py is different?  Can you try updating h5py?  And you're using python3, right?
