### 2021-11-17 TeamCOMPAS/COMPAS#698 @HeloiseS
Also - I'm guessing i can't ensure that the BSE_Compact_objects entries will be the same number when I do two separate runs... so how can i put together my hdf5 files from two different metallicities?
