### 2021-11-17 TeamCOMPAS/COMPAS#698 @HeloiseS
well silly me - i should have done h5view.py from the start... It's because I've got a detailed output and I forgot :eyes: I have thousands of files such as `COMPAS_Output/Detailed_Output/BSE_Detailed_Output_4625.h5` - I should have known when it took 1200 seconds to run my simulation when ilya said it'd take 2 minutes... I just though kvn was slow!!
