### 2021-11-17 #3838 @HeloiseS
### "FINAL" UPDATE - tutorials ran perfectly

Alright! thanks @ilyamandel  for the pointers, I've played with making my own models and ran the tutorials on Cosmic Integration: they ran like clockwork!

I made some detailed models and some not-detailed models: the run time you mentioned was spot on :+1: Super fast!
Suggestion: give  instructions to make the models to run those tutorials  as an alternative to having to DL a big data set? Plus to run the 6th tutorial you need several metallicities which gives an opportunity to demonstrate the use of the h5view.py and h5copy.py scripts in practice :) That would tie all of the examples and tutorials you have together to train your new users in all your utilities.

---
I have one more (it's more of a comment than a) question and then I'm done: i looked at the ZAMS of some of my model (just for fun :) ) compared it to BPASS  and it all looks spotless, but there is that one dot a little lower than the rest ....
![image](https://user-images.githubusercontent.com/15654896/142140712-827f508d-2d95-4ab2-8106-30ea0c5c4014.png)

So I looked at the IMF and if you plot it weird (not loging the masses) you see ONE model at 8 Msol... which is weird because my minimum mass in my run is 10 Msol (which I double checked in the run log file). I can't imagine this would have any impact anywhere but since it's unexpected behaviour I thought I'd let you know :)

![image](https://user-images.githubusercontent.com/15654896/142140991-e356fe11-c68c-46bd-9eca-48dae7aefdd8.png)
