### 2021-11-17 #3838 @jeffriley
>  see ONE model at 8 Msol... which is weird because my minimum mass in my run is 10 Msol

minimum-secondary-mass is 10?  I assume yes from the IMF plot.  If you have allow-rlof-at-birth enabled and one of the stars is overflowing its Roche lobe at birth, then COMPAS will equilibrate the masses of the stars and circularise the orbit - so the equilbrated masses could be lower than expected - but that shouldn't happen if your minimum secondary mass is 10.  There is a column in the BSE_System_Parameters file that records if the masses were equlibrated at birth (the "Equilibrated_At_Birth" column).
