### 2021-11-18 5cb05ea
added printing for CO, He cores during CEE (or stellar mergers)
