### 2021-11-19 c37b83f
Defect repairs: (a) clamp timestep returned in BaseStar::CalculateTimestep() to NUCLEAR_MINIMUM_TIMESTEP (b) change NUCLEAR_MINIMUM_TIMESTEP to 1 year (from 100 years) in constants.h
