+++ src/BaseStar.cpp
@@ -2903,7 +2903,7 @@ double BaseStar::CalculateTimestep() {
 
     double dt = ChooseTimestep(m_Age);
 
-    return LimitTimestep(dt);
+    return max(LimitTimestep(dt), NUCLEAR_MINIMUM_TIMESTEP);                            // clamp timestep to minimum NUCLEAR_MINIMUM_TIMESTEP
 }
 
 
@@ -3056,7 +3056,6 @@ STELLAR_TYPE BaseStar::UpdateAttributesAndAgeOneTimestep(const double p_DeltaMas
  * @param   [IN]    p_DeltaTime                 Timestep in Myr
  */
 void BaseStar::AgeOneTimestepPreamble(const double p_DeltaTime) {
-
     if (p_DeltaTime > 0.0) {                        // if dt > 0    (don't use utils::Compare() here)
         m_Time += p_DeltaTime;                      // advance physical simulation time
         m_Age  += p_DeltaTime;                      // advance age of star
