### 2021-11-19 99d7db6
Update BaseBinaryStar.cpp
