### 2021-11-19 ab1c371
Update BaseStar.cpp
