### 2021-11-19 199b9ef
Update Star.cpp
