### 2021-11-19 22c50a5
Update BaseBinaryStar.cpp
