### 2021-11-19 dd3510f
Update BaseStar.cpp
