### 2021-11-19 b9e8e47
Update BaseBinaryStar.cpp
