### 2021-11-22 6f35ffb
cleaning up postproc and related python files
