+++ postProcessing/plot_detailed_evolution.py
@@ -1,0 +1,593 @@
+###################################################################
+#                                                                 #                                                               
+#  Example of plotting detailed output COMPAS with python         #
+#                                                                 #
+# ##################################################################
+
+import os, sys
+import math
+import numpy as np
+import h5py as h5
+import pandas as pd
+from scipy import ndimage
+import matplotlib.pyplot as plt
+import matplotlib.ticker as ticker  
+from matplotlib.legend import Legend
+from matplotlib import rcParams, transforms, patches
+import matplotlib.gridspec as gridspec
+
+def main():
+
+    ### Read file and create dataframe.
+    #data_path = './COMPAS_Output/Detailed_Output/BSE_Detailed_Output_{}.h5'.format(sys.argv[1])
+    #data_path = '/home/rwillcox/astro/compas/COMPAS/examples/methods_paper_plots/detailed_evolution/COMPAS_Output/Detailed_Output/BSE_Detailed_Output_0.h5'
+    #data_path = '/home/rwillcox/astro/compas/COMPAS/examples/methods_paper_plots/detailed_evolution/COMPAS_Output_2/Detailed_Output/BSE_Detailed_Output_0.h5'
+    data_path = '/home/rwillcox/astro/compas/COMPAS/output/detailed_evol_vanDenHeuval_plots/COMPAS_Output_2/Detailed_Output/BSE_Detailed_Output_0.h5'
+
+    Data = h5.File(data_path, 'r')
+
+    #testing(Data)
+
+    ### Collect the important events in the detailed evolution
+    events = getAllEvents(Data) # Calculate the events here, for use in plot sizing parameters
+    printEvolutionaryHistory(events=events)
+    events = [event for event in events if event.eventClass != 'Stype'] # want to ignore simple stellar type changes
+
+    ### Produce the two plots
+    makeDetailedPlots(Data, events)
+    makeVanDenHeuvalPlot(Data, events)
+
+
+fontparams = {
+    "font.serif": "Times New Roman",
+    "text.usetex": "True",
+    "axes.grid": "True",
+    "grid.color": "gray",
+    "grid.linestyle": ":",
+    "axes.titlesize": "18",
+    "axes.labelsize": "12",
+    "xtick.labelsize": "10",
+    "ytick.labelsize": "10",
+    "xtick.labelbottom": "True", 
+    "legend.framealpha": "1",
+    #"legend.fontsize": "None",
+}
+
+
+
+####### Functions to organize and call the plotting functions 
+
+def makeDetailedPlots(Data=None, events=None):
+
+    #listOfPlots = [ plotMassAttributes, plotLengthAttributes, plotStellarTypeAttributesAndEccentricity ]
+    listOfPlots = [ plotMassAttributes, plotLengthAttributes, plotStellarTypeAttributes, plotEccentricity]
+
+    events = [event for event in events if event.eventClass != 'Stype'] # want to ignore simple stellar type changes
+    num_events = len(events)
+    event_times = [event.time for event in events]
+
+
+    rcParams.update(fontparams) # Set configurations for uniform plot output
+    fig, axes = plt.subplots(nrows=len(listOfPlots), figsize=(10, 20)) # W, H
+    #gs = gridspec.GridSpec(ncols=nEventsColumns, nrows=nRows, figure=fig, height_ratios=height_ratios)
+
[…]
