### 2021-11-22 TeamCOMPAS/COMPAS#686 @reinhold-willcox
Sorry, I haven't kept up with all the activity on github! I am making changes to the postProcessing folder & notebooks, so I will add on to clean up the docs/ folder as well (there are a lot of hyperlinks back and forth, so it makes sense to do them in parallel).
