### 2021-11-23 f495faa
Changed RLOF(X)>MT to just RLOF(X) in the BSE_RLOF output. This is much clearer, since it refers to the state immediately during the MT, not after it.
