### 2021-11-23 #3838 @avigna
> If you're curious to see what i did with this class (i refactored it for my own use) i can fork a suggestion for the class and yaml file (unless Jeff prefers JSON 😉 ) for you to look at and see if it would work with the general workflow of compas.
@HeloiseS if the offer still holds, I'd be keen to check it out. I just got into thinking/working on it, and haven't decided which way to proceed, so your suggestions might be elucidating. Thanks in advance.
