### 2021-11-23 #3838 @HeloiseS
i think this is worth discussing viva voce in a call (becuse it'll clog the thread which is already lengthy) - i don't even think you need  something like what I'd refactored it as initially (but it depends on what you want from it later!) and it's easier to explain with the document under our eyes. email me at hfstevance@gmail.com and we can zom or something. then i can write a summary on here for record keeping. sounds good?
