### 2021-11-23 24853f5
changed stellar type numbers to names
