+++ postProcessing/plot_detailed_evolution.py
@@ -29,7 +29,7 @@ def main():
     #testing(Data)
 
     ### Collect the important events in the detailed evolution
-    events = getAllEvents(Data) # Calculate the events here, for use in plot sizing parameters
+    events = allEvents(Data).allEvents # Calculate the events here, for use in plot sizing parameters
     printEvolutionaryHistory(events=events)
     events = [event for event in events if event.eventClass != 'Stype'] # want to ignore simple stellar type changes
 
@@ -272,17 +272,14 @@ def plotVanDenHeuval(fig=None, axes=None, events=None):
         ax.set_xlim([0, imgH*imgAspectWH])
         ax.set_ylim([0, imgH])
 
-        #nImg = [2, 10, 26, 28][ii%4]
-        #fname = 'drawings/'+str(nImg)+'.png'
         img = events[ii].eventImage
-        #img = plt.imread(fname) # import image
     
         x0, xF = ax.get_xlim()
         y0, yF = ax.get_ylim()
         ax.imshow(img, extent=(x0, xF, y0, yF)) # l, r, b, t 
 
         ax.text(x=-0.1*(xF-x0), y=0.9*(yF-y0), s=chr(ord('@')+1+ii), fontweight='bold') # The unicode representation of the capital letters - works as long as there are less than 26 images to show
-        ax.set_xlabel(events[ii].eventString) #, pad=-100)
+        ax.set_xlabel(events[ii].eventString) 
 
 
 
@@ -319,28 +316,14 @@ def getStellarTypes(Data):
 ###########################################################
 
 
-def testing(Data):
-    event = Event(Data, 0, 'Beg')
-    img = event.getEventImage(2, True)
-
-    fig, ax = plt.subplots()
-
-    imgAspectWH = 1.7786666666666666 # inverse of above
-    ax.set_xlim([0, imgAspectWH])
-    ax.set_ylim([0, 1])
-    x0, xF = ax.get_xlim()
-    y0, yF = ax.get_ylim()
-    ax.imshow(img, extent=(x0, xF, y0, yF)) # l, r, b, t 
-    plt.show()
-
-
 class Event(object):
 
-    def __init__(self, Data, index, eventClass, **kwargs):
+    def __init__(self, Data, index, eventClass, stellarTypeMap, **kwargs):
 
         self.Data   = Data
         self.index  = index
         self.eventClass = eventClass # Can be any of 'Beg', 'End', 'MT', 'SN', 'Stype'
+        self.stellarTypeMap = stellarTypeMap
 
         ii = index
         self.time   = Data['Time'][ii] 
@@ -348,13 +331,12 @@ class Event(object):
         self.m2     = Data['Mass(2)'][ii] 
         self.stype1 = Data['Stellar_Type(1)'][ii] 
         self.stype2 = Data['Stellar_Type(2)'][ii] 
+        self.stypeName1 = stellarTypeMap[self.stype1]
+        self.stypeName2 = stellarTypeMap[self.stype2]
         self.a      = Data['SemiMajorAxis'][ii]
         self.e      = Data['Eccentricity'][ii]
 
-
         self.eventString = self.getEventDetails(**kwargs)
-        #self.eventImage = getEventImage(image_num)
-        #self.eventImage = self.getEventImage()
 
 
     def getEventDetails(self, **kwargs):
@@ -416,10 +398,11 @@ class Event(object):
 
         elif eventClass == 'SN':
             whichStar = kwargs['whichStar']
-            remnantType = 'NS' if (Data['Stellar_Type({})'.format(whichStar)][ii] == 13) else 'BH'
+            remnantTypeName = self.stellarTypeMap[Data['Stellar_Type({})'.format(whichStar)][ii]] 
             compType = Data['Stellar_Type({})'.format(2 if whichStar==1 else 1)][ii]
[…]
