+++ postProcessing/plot_detailed_evolution.py
@@ -17,16 +17,11 @@ from matplotlib import rcParams, transforms, patches
 import matplotlib.gridspec as gridspec
 
 def main():
-
     ### Read file and create dataframe.
-    #data_path = './COMPAS_Output/Detailed_Output/BSE_Detailed_Output_{}.h5'.format(sys.argv[1])
-    #data_path = '/home/rwillcox/astro/compas/COMPAS/examples/methods_paper_plots/detailed_evolution/COMPAS_Output/Detailed_Output/BSE_Detailed_Output_0.h5'
-    #data_path = '/home/rwillcox/astro/compas/COMPAS/examples/methods_paper_plots/detailed_evolution/COMPAS_Output_2/Detailed_Output/BSE_Detailed_Output_0.h5'
-    data_path = '/home/rwillcox/astro/compas/COMPAS/output/detailed_evol_vanDenHeuval_plots/COMPAS_Output_2/Detailed_Output/BSE_Detailed_Output_0.h5'
+    # data_path = '/home/rwillcox/astro/compas/COMPAS/output/detailed_evol_vanDenHeuval_plots/COMPAS_Output_2/Detailed_Output/BSE_Detailed_Output_0.h5'
+    data_path = '/Users/13lauy1/git/compas2/COMPAS/postProcessing/BSE_Detailed_Output_0.h5'
 
     Data = h5.File(data_path, 'r')
-
-    #testing(Data)
 
     ### Collect the important events in the detailed evolution
     events = getAllEvents(Data) # Calculate the events here, for use in plot sizing parameters
@@ -34,8 +29,10 @@ def main():
     events = [event for event in events if event.eventClass != 'Stype'] # want to ignore simple stellar type changes
 
     ### Produce the two plots
-    makeDetailedPlots(Data, events)
-    makeVanDenHeuvalPlot(Data, events)
+    # makeDetailedPlots(Data, events)
+    plotVanDenHeuval(events=events)
+    plt.savefig('vanDenHeuvalPlot.eps', format='eps')
+    plt.show()
 
 
 fontparams = {
@@ -110,27 +107,7 @@ def makeDetailedPlots(Data=None, events=None):
     fig.suptitle('Detailed evolution for seed={}'.format(Data['SEED'][()][0]), fontsize=24) #, y=1)
     fig.tight_layout(h_pad=8, rect= (0, .08, 1, .95)) #, h_pad=8) # (left, bottom, right, top) 
     plt.savefig('detailedEvolutionPlot.eps', format='eps')
-    #plt.show()
-
-
-def makeVanDenHeuvalPlot(Data=None, events=None):
-
-    num_events = len(events)
-
-    # The figure size should increase to account for the number of events
-    #nPlots = len(listOfPlots) # do I need this?
-    #nEventsColumns = 3 # vary this as desired
-    #nRows = nPlots + math.ceil(num_events/nEventsColumns)
-    #height_ratios=[1]*nPlots + [0.8]*(nRows-nPlots)
-    figHeight = 3*num_events + .5*(num_events-1)
-    fig, axes = plt.subplots(nrows=num_events, figsize=(4, figHeight)) # W, H  # constrained_layout=True, 
-
-    plotVanDenHeuval(fig=fig, axes=axes, events=events) #, figHeight=20):
-
-    #fig.tight_layout()
-    fig.tight_layout(h_pad=8, rect= (0, .05, 1, .95) ) # (left, bottom, right, top)  
-    plt.savefig('vanDenHeuvalPlot.eps', format='eps')
-    plt.show()
+    #plt.show()    
 
 
 
[…]
