### 2021-11-24 fb8000a
added createMtHistString functions to compasUtils
