+++ postProcessing/PythonScripts/compasUtils.py
@@ -1,8 +1,14 @@
-# Function to print the data in a given COMPAS HDF5 group in a readable pandas template
-
 import h5py as h5
 import numpy as np
 import pandas as pd
+
+
+########################################################################
+### 
+### Function to print the data from a given COMPAS HDF5 group 
+### in a readable pandas template
+### 
+########################################################################
 
 def printCompasDetails(data, *seeds, mask=()):
     """
@@ -55,3 +61,269 @@ def printCompasDetails(data, *seeds, mask=()):
     df.insert(loc=0, column='(units)', value=pd.Series(units_dict))
     return df
 
+
+
+########################################################################
+### 
+### Get event histories of MT data, SN data, and combined MT, SN data
+### 
+########################################################################
+
+def getMTevents(MTs):                                     
+    """
+    This function takes in the `BSE_RLOF` output category from COMPAS, and returns the information
+    on the Mass Transfer (MT) events that happen for each seed. The events do not have to be in order, 
+    either chronologically or by seed, this function will reorder them as required.
+    
+    OUT:
+        tuple of (returnedSeeds, returnedEvents, returnedTimes)
+        returnedSeeds (list): ordered list of the unique seeds in the MTs file
+        returnedEvents (list): list of sublists, where each sublist contains all the MT events for a given seed.
+            MT event tuples take the form :
+            (stellarTypePrimary, stellarTypeSecondary, isRlof1, isRlof2, isCEE)
+        returnTimes (list): is a list of sublists of times of each of the MT events
+    """
+
+    mtSeeds = MTs['SEED'][()]
+    mtTimes = MTs['Time<MT'][()]
+    mtPrimaryStype = MTs['Stellar_Type(1)<MT'][()]
+    mtSecondaryStype = MTs['Stellar_Type(2)<MT'][()]
+    mtIsRlof1 = MTs['RLOF(1)>MT'][()] == 1
+    mtIsRlof2 = MTs['RLOF(2)>MT'][()] == 1
+    mtIsCEE = MTs['CEE>MT'][()] == 1
+
+    # We want the return arrays sorted by seed, so sort here.
+    mtSeedsInds = mtSeeds.argsort()
+    mtSeeds = mtSeeds[mtSeedsInds]  
+    mtTimes = mtTimes[mtSeedsInds]  
+    mtPrimaryStype = mtPrimaryStype[mtSeedsInds]
+    mtSecondaryStype = mtSecondaryStype[mtSeedsInds]
+    mtIsRlof1 = mtIsRlof1[mtSeedsInds]
+    mtIsRlof2 = mtIsRlof2[mtSeedsInds]
+    mtIsCEE = mtIsCEE[mtSeedsInds]
+
+    # Process the MT events
+
+    returnedSeeds = []                                      # array of seeds - will only contain seeds that have MT events
+    returnedEvents = []                                     # array of MT events for each seed in returnedSeeds
+    returnedTimes = []                                      # array of times for each event in returnedEvents (for each seed in returnedSeeds)
+
+    lastSeed = -1                                           # initialize most recently processed seed
+
+    for seedIndex, thisSeed in enumerate(mtSeeds):          # iterate over all RLOF file entries
+        thisTime = mtTimes[seedIndex]                       # time for this RLOF file entry
+        thisEvent = (mtPrimaryStype[seedIndex], mtSecondaryStype[seedIndex], 
+                     mtIsRlof1[seedIndex], mtIsRlof2[seedIndex], mtIsCEE[seedIndex])  # construct event tuple
+
+        # If this is an entirely new seed:
+        if thisSeed != lastSeed:                            # same seed as last seed processed?
+            returnedSeeds.append(thisSeed)                  # no - new seed, record it
[…]
