### 2021-11-24 3d15df8
fixed references to local files, now images just call COMPAS_ROOT_DIR env var
