### 2021-11-24 8e7244f
added function to space out event letters when they would overlap
