+++ postProcessing/plot_detailed_evolution.py
@@ -2,7 +2,7 @@
 #                                                                 #                                                               
 #  Example of plotting detailed output COMPAS with python         #
 #                                                                 #
-# ##################################################################
+###################################################################
 
 import os, sys
 import math
@@ -19,14 +19,9 @@ import matplotlib.gridspec as gridspec
 def main():
 
     ### Read file and create dataframe.
-    #data_path = './COMPAS_Output/Detailed_Output/BSE_Detailed_Output_{}.h5'.format(sys.argv[1])
-    #data_path = '/home/rwillcox/astro/compas/COMPAS/examples/methods_paper_plots/detailed_evolution/COMPAS_Output/Detailed_Output/BSE_Detailed_Output_0.h5'
-    #data_path = '/home/rwillcox/astro/compas/COMPAS/examples/methods_paper_plots/detailed_evolution/COMPAS_Output_2/Detailed_Output/BSE_Detailed_Output_0.h5'
-    data_path = '/home/rwillcox/astro/compas/COMPAS/output/detailed_evol_vanDenHeuval_plots/COMPAS_Output_2/Detailed_Output/BSE_Detailed_Output_0.h5'
+    data_path = 'COMPAS_Output/Detailed_Output/BSE_Detailed_Output_0.h5'
 
     Data = h5.File(data_path, 'r')
-
-    #testing(Data)
 
     ### Collect the important events in the detailed evolution
     events = allEvents(Data).allEvents # Calculate the events here, for use in plot sizing parameters
@@ -50,7 +45,6 @@ fontparams = {
     "ytick.labelsize": "10",
     "xtick.labelbottom": "True", 
     "legend.framealpha": "1",
-    #"legend.fontsize": "None",
 }
 
 
@@ -69,11 +63,9 @@ def makeDetailedPlots(Data=None, events=None):
 
     rcParams.update(fontparams) # Set configurations for uniform plot output
     fig, axes = plt.subplots(nrows=len(listOfPlots), figsize=(10, 20)) # W, H
-    #gs = gridspec.GridSpec(ncols=nEventsColumns, nrows=nRows, figure=fig, height_ratios=height_ratios)
 
     for ii, specificPlot in enumerate(listOfPlots): # exclude the last one
 
-        #ax = fig.add_subplot(gs[ii,:])     
         ax = axes[ii]
         if ii == 0:
             ax0 = ax
@@ -93,10 +85,10 @@ def makeDetailedPlots(Data=None, events=None):
         [ax.axvline(time, ymin=0.975, zorder=0) for time in event_times]
 
         # On all plots, add the event letters
-        # TODO: find a way to not let them overlap
-        #if ii == 0:
+        spaced_out_event_times = space_out(event_times, min_separation=ax.get_xlim()[1]/75) # min_separation of xmax/50 was found to fit the letter sizes well
         for jj in range(num_events):
-            ax.text(x=event_times[jj], y=ax.get_ylim()[1]*1.05, s=chr(ord('@')+1+jj)) # The unicode representation of the capital letters - works as long as there are less than 26 images to show
+            yOffsetFactor = 3 if (ax.get_yscale() == 'log') else 1.1
+            ax.text(x=spaced_out_event_times[jj], y=ax.get_ylim()[1]*yOffsetFactor, s=chr(ord('@')+1+jj)) # The unicode representation of the capital letters - works as long as there are less than 26 images to show
 
         if handles is not None:
[…]
