### 2021-11-24 10e0015
more tweaks to van den Heuval plot
