### 2021-11-24 TeamCOMPAS/COMPAS#713 @katiebreivik
Docker command line and documentation typos
This issue is submitted as part of the JOSS review [#3838](https://github.com/openjournals/joss-reviews/issues/3838)

The Docker documentation has a typo in the options for specifying the number of systems to evolve when running via the command line

**To Reproduce**
Steps to reproduce the behavior:
Running in the Docker via the command line following the docs (https://compas.readthedocs.io/en/latest/pages/User%20guide/Running%20COMPAS/running-via-docker-running.html#via-the-command-line):
```
docker run                                  \
    --rm                                    \
    -it                                     \
    -v $(pwd)/compas-logs:/app/COMPAS/logs  \
    teamcompas/compas                       \
    bin/COMPAS                              \
    --number-of-binaries=5                  \
    --output-path=/app/COMPAS/logs
```

I expect this to run seamlessly, but get an error due to a typo in the option specifier: `--number-of-binaries=5 `

Based on output from a `--help` run, I think this should be `--number-of-systems=5 ` which runs as expected.
