### 2021-11-24 TeamCOMPAS/COMPAS#713 @katiebreivik
One more quick typo: In [compile-time configuration](https://compas.readthedocs.io/en/latest/pages/User%20guide/configuration.html): constanst.h --> constants.h
