### 2021-11-24 TeamCOMPAS/COMPAS#714 @katiebreivik
plotting script typo in examples/methods_paper_plots/detailed_evolution
This issue is submitted as part of the JOSS review [#3838](https://github.com/openjournals/joss-reviews/issues/3838)

**Describe the bug**
The example use case of the detailed evolution of GW15226 plotting script fails because of a typo

**Label the issue**
`urgency_moderate` - This is a moderately urgent issue
`severity_moderate` - This is a moderately severe bug
urgency/severity are moderate only because it breaks a test that confirms a correct installation!

**To Reproduce**
Steps to reproduce the behavior:
when running the example given in the [documentation](https://compas.readthedocs.io/en/latest/pages/User%20guide/Tutorial/example-compas-run.html#tutorial-simple-compas-run), the plotting script fails because of a non-existing file. The automatically generated COMPAS output is placed in the correct directory, but since a single system is run, the `_0.h5`  in the [plotting script](https://github.com/TeamCOMPAS/COMPAS/blob/99cd751377f2a6d0d856124e6e516259386a0523/examples/methods_paper_plots/detailed_evolution/detailed_evol_plotter.py#L18) causes an error. I think this can be fixed by removing the `_0` in the plotting script.
