### 2021-11-24 #3838 @katiebreivik
Hi All -- here is my long awaited (APOLOGIES) review:

The paper contains all the necessary information and is clear and concise! I was also able to follow all of the documentation for the installation/running of COMPAS as expected with the exception of a couple of typos I found after the updates based on @HeloiseS's suggestions above so I dropped an issue: [#713](https://github.com/TeamCOMPAS/COMPAS/issues/713).

Next, here are some suggestions that I won't require for approval but think are very much worth considering:
Similar to @HeloiseS, I do find the pythonSubmit.py files to be cumbersome. I absolutely understand the reasons for this based on the COMPAS c++ code-base and also understand that it is unlikely that the pythonSubmit.py files are used in paper-quality population runs on a cluster by the team. But! I would advocate (similar to @HeloiseS again!) that the team consider a near-future refactor of the pythonSubmit.py-like files or add some extra documentation that walks a user through the process of editing them (e.g. which lines should and shouldn't be modified). Since most of the tutorials seem to rely on them, making them as clear as possible will pay many dividends!

Similarly, the tutorials for the post-processing and paper figures are on the right track but I also found it frustrating to try to track down the data for each of them. I understand that they are a work in progress, but I think it is worth considering whether there are some simple tutorials that could be quickly developed which show a user how to run/work with moderate amounts of data that they generate themselves (a la the detailed evolution figure!). I am a *huge* fan of reproducibility and open data access, but when somebody is just starting out with population synthesis, downloading a huge file from the paper zenodo to run a tutorial might be the barrier that sends them elsewhere.

The other outstanding issues I have that are required for approval but are quick fixes:

- [ ] the documentation doesn't provide a clear statement of need right on the docs [splash page](https://compas.readthedocs.io/en/latest/). The quickest/easiest fix is probably to copy the COMPAS README.md text (which is great!) over from the repository.
- [ ] The docs don't give an explicit way to test whether the installation was successful in the [Getting started ](https://compas.readthedocs.io/en/latest/pages/Getting%20started/getting-started.html#getting-started) tab. I did find some examples for testing whether COMPAS is properly installed in the configuration tab (e.g. `./compas -v` or `./compas -h`). These would be helpful for a user to have as a 'testing your installation' tab under `Getting started` so they don't have to dig.
- [ ] I'm assuming that the detailed evolution demo in the examples folder is supposed to be used as the quick way for a user to test whether the evolution is working properly. When I run the pythonSubmit.py file, it produces output as expected but the plotting script has a typo based on the automatic output for the [BSE_Detailed_Output.h5](https://github.com/TeamCOMPAS/COMPAS/blob/dev/examples/methods_paper_plots/detailed_evolution/detailed_evol_plotter.py#L18) file. Since this is the only very quick way for a user to verify that they are producing a proper stellar evolution output, I think this needs to be fixed for this review and have opened up an issue here: [#714](https://github.com/TeamCOMPAS/COMPAS/issues/714)

Apologies again for my delay -- and special apologies to @christinahedges for accidentally clicking "make an issue" on one of my checkboxes! I am cursed to accidentally do this on every JOSS review I participate in. Sorry about that!
