### 2021-11-24 #3838 @christinahedges
@HeloiseS, @ilyamandel and @avigna, I'd like to ask you to discuss these suggestions as much as possible in the thread here, or on the repo. These reviews are open, and taking the review into a different format (zoom) excludes other participants in the review. If @HeloiseS has suggestions for code and you want to keep this thread shorter, feel free to open an issue or PR on the main repo and then cross-reference it here. Especially if this is a suggested change to the code, please open a PR and discuss in the open. Thanks!
