### 2021-11-24 #3838 @christinahedges
Hello all, thank you to both @HeloiseS and @katiebreivik  for your excellent feedback, and thorough review. I think @HeloiseS may have further review comments and suggestions after @katiebreivik's final review, please do post these here. I want to check in at this point and sum up that
- There are a few issues pointed out by both reviewers which need to be fixed before the submission can be accepted by JOSS
- There are some larger pieces of feedback about the usability/API of the tool and the tutorials.

At this point, it's clear some of this feedback is necessary to pass the review. We do encourage authors to "go the extra mile", and I think the feedback from the reviewers about improving the code usability is extremely valuable. I do understand it is also work for the submitting authors.  I'd like to give the authors a chance to act on the feedback they've received and they are able to respond to, and then we can discuss how much resource the authors have to respond to higher level, go-the-extra-mile feedback, and whether this feedback prevents the submission being accepted to JOSS.
