### 2021-11-24 TeamCOMPAS/COMPAS#715 @HeloiseS
Refactoring pythonSubmit.py
_This issue is submitted as part of the JOSS review [#3838](https://github.com/openjournals/joss-reviews/issues/3838)_

When it comes to the pythonSumbit.py file refactor you can go one of two routes - the one you choose will depend on whether you *ever* use the pythonProgramOption class anywhere else.

If you do not use that class anywhere else, there it is not a use case for a class. Remove the class, just have the  giant list of options listed as it is, your comments are detailed already, and maybe instead of having two separate lists for the commands and the variable names, you can put those in a dictionary directly. You do not need a whole bunch of functions and methods (e.g. the ones whose sole purpose is to make and return the list) if they are only called one time in the __main__ portion of the file.

If you DO want this to be a class to be imported in other scripts and feel OO is necessary, then the main thing is to remove the methods without purpose (e.g. creating lists doesn't need to be done in a method) and have an __init__ method. importantly this method would then be where you parse the configuration for COMPAS so that you don't have to go in the class everytime you want a different simulation. **if you are going to keep it a class, your user shouldn't have to modify the class itself to change the COMPAS options**, that is behaviour you expect from a script, not a class. You can put your options in a setup file in a format like yaml. This makes the actual code in pythonSubmit much shorter. Here is a quick refactor i did early on (note github won't let me have .py or .yaml so  you'll need to amend the extension after downloading) that you can use as inspiration or to build on. Note I have not tested this refactor extensively.

[compas_config.txt](https://github.com/TeamCOMPAS/COMPAS/files/7598005/compas_config.txt)
[pythonSubmitDemo.txt](https://github.com/TeamCOMPAS/COMPAS/files/7598006/pythonSubmitDemo.txt)
