### 2021-11-24 a1fe1fd
space out van den heuval event letters (#712)
* added function to space out event letters when they would overlap

* cleaned up merge with dev

* removed local file reference

* more tweaks to van den Heuval plot

Co-authored-by: Mike Lau <mike.lau@monash.edu>
