+++ postProcessing/plot_detailed_evolution.py
@@ -2,7 +2,7 @@
 #                                                                 #                                                               
 #  Example of plotting detailed output COMPAS with python         #
 #                                                                 #
-# ##################################################################
+###################################################################
 
 import os, sys
 import math
@@ -16,7 +16,7 @@ from matplotlib.legend import Legend
 from matplotlib import rcParams, transforms, patches
 import matplotlib.gridspec as gridspec
 
-compasRootDir = os.environ['COMPAS_ROOT_DIR']
+compasRootDir = os.path.expandvars(os.environ['COMPAS_ROOT_DIR'])
 
 def main():
     ### Read file and create dataframe.
@@ -32,7 +32,7 @@ def main():
     ### Produce the two plots
     makeDetailedPlots(Data, events)
     plotVanDenHeuval(events=events)
-    plt.savefig('vanDenHeuvalPlot.eps', format='eps')
+    plt.savefig('vanDenHeuvalPlot.eps', bbox_inches='tight',pad_inches = 0, format='eps')
     plt.show()
 
 
@@ -48,7 +48,6 @@ fontparams = {
     "ytick.labelsize": "10",
     "xtick.labelbottom": "True", 
     "legend.framealpha": "1",
-    #"legend.fontsize": "None",
 }
 
 
@@ -57,7 +56,6 @@ fontparams = {
 
 def makeDetailedPlots(Data=None, events=None):
 
-    #listOfPlots = [ plotMassAttributes, plotLengthAttributes, plotStellarTypeAttributesAndEccentricity ]
     listOfPlots = [ plotMassAttributes, plotLengthAttributes, plotStellarTypeAttributes, plotEccentricity]
 
     events = [event for event in events if event.eventClass != 'Stype'] # want to ignore simple stellar type changes
@@ -67,16 +65,10 @@ def makeDetailedPlots(Data=None, events=None):
 
     rcParams.update(fontparams) # Set configurations for uniform plot output
     fig, axes = plt.subplots(nrows=len(listOfPlots), figsize=(10, 20)) # W, H
-    #gs = gridspec.GridSpec(ncols=nEventsColumns, nrows=nRows, figure=fig, height_ratios=height_ratios)
 
     for ii, specificPlot in enumerate(listOfPlots): # exclude the last one
 
-        #ax = fig.add_subplot(gs[ii,:])     
         ax = axes[ii]
-        if ii == 0:
-            ax0 = ax
-        else:
-            ax.sharex(ax0)
 
         # TODO: Set the reverse log scale for time
 
@@ -90,11 +82,15 @@ def makeDetailedPlots(Data=None, events=None):
         # Add vertical lines for specific event times
         [ax.axvline(time, ymin=0.975, zorder=0) for time in event_times]
 
-        # On all plots, add the event letters
-        # TODO: find a way to not let them overlap
-        #if ii == 0:
-        for jj in range(num_events):
-            ax.text(x=event_times[jj], y=ax.get_ylim()[1]*1.05, s=chr(ord('@')+1+jj)) # The unicode representation of the capital letters - works as long as there are less than 26 images to show
+        # Add the event letters to the first plot
+        if ii == 0:
+            spaced_out_event_times = space_out(event_times, min_separation=ax.get_xlim()[1]/75) # min_separation of xmax/50 was found to fit the letter sizes well
+            for jj in range(num_events):
+                yOffsetFactor = 1.5 if (ax.get_yscale() == 'log') else 1.02
[…]
