### 2021-11-24 TeamCOMPAS/COMPAS#714 @reinhold-willcox
Hi @katiebreivik , thanks for catching this! I'm not sure I follow though. The `_0.h5` extension should be appended to the detailed output file even if there is only a single system. If you delete all the COMPAS_Output* directories and rerun the pythonSubmitDemo.py script, what is the path to your detailed output file?

Somewhat related, we've been heavily refactoring this plotting script recently. For the purposes of the review, would you like us to swap out the old script for the new one, or simply ensure that the old one is bug free? The new script does not reproduce the plot exactly as seen in the methods paper, but it is very similar.
