### 2021-11-25 TeamCOMPAS/COMPAS#715 @reinhold-willcox
If I can add my two cents to this, since we need a major refactor of the pythonSubmit code anyway, I propose we work with the stroopwafel folks to put together a single script where users can specify all COMPAS parameters, as well as whether to use stroopwafel and/or HPC (and the related parameters there), instead of having many disparate scripts that will sometimes need to be combined in some convoluted way.
