### 2021-11-25 TeamCOMPAS/COMPAS#715 @HeloiseS
yes please - was there a stroopwafel tutorial i missed? i haven't used that yet
