### 2021-11-25 TeamCOMPAS/COMPAS#714 @ilyamandel
@katiebreivik -- I confirmed that this is not an issue, as the latest COMPAS version does produce the file BSE_Detailed_Output_0.h5, which is what the plotting script expects.  What COMPAS version are you running?
