### 2021-11-25 TeamCOMPAS/COMPAS#715 @reinhold-willcox
You haven't missed anything! Some of the compas people / affiliates have been working toward improving stroopwafel, and there currently exists a stroopwafel interface file for compas which is external to the compas code. But it's been on ongoing question how best to bring stroopwafel, or just the interface file, into compas.
