### 2021-11-25 TeamCOMPAS/COMPAS#695 @ilyamandel
@jeffriley  -- do I understand correctly that the only action item on this is to document the existing behaviour?
