### 2021-11-25 TeamCOMPAS/COMPAS#714 @katiebreivik
@ilyamandel You are 100% correct I was on an old version; I hadn't added the repo as an upstream remote branch which means when I fetch upstream and got nothing, I figured I had cloned directly and just did a git pull.

On the latest dev branch (v02.25.10) I have verified that everything works!
