### 2021-11-25 TeamCOMPAS/COMPAS#715 @katiebreivik
@reinhold-willcox +++1 to your suggestion! Keeping a consistent, single interface to the c++ code is a great idea. And I agree with @HeloiseS that the options should not be baked into a class. Ideally, I think a user should be able to specify options through a command line interface or an ini/text file since modifying submit scripts introduces many opportunities for a user to break the code.
