### 2021-11-26 cfab18d
many cheanges to clean up the tutorial
