+++ postProcessing/Tutorial/DataAnalysis.py
@@ -22,7 +22,7 @@
 # ## Material
 #
 # ### [1. Inspecting the data ](#1.-Inspecting-the-data)
-# TODO:! Look at the raw hdf5 data to see which parameters are available and check that it matches expectations.
+# Look at the data to see which parameters are available and check that it matches expectations.
 #
 # ### [2. Slicing the data ](#2.-Slicing-the-data)
 # Select specific systems and their parameters using seeds.
@@ -45,15 +45,16 @@ import time                      # for finding computation time
 import matplotlib.pyplot as plt  #for plotting
 
 # Import COMPAS specific scripts
+compasRootDir = os.environ['COMPAS_ROOT_DIR'] 
 sys.path.append(compasRootDir + 'postProcessing/PythonScripts')
-from compasUtils import printCompasDetails
+from compasUtils import printCompasDetails, getEventHistory, getEventStrings
+
+# Choose an output hdf5 file to work with
+pathToData = 'COMPAS_Output_tutorial.h5'
 
 # This is known as an ipython magic command, and allows plots to be produced within the notebook
 # %matplotlib inline
 
-# Choose an output hdf5 file to work with
-compasRootDir = os.environ['COMPAS_ROOT_DIR'] 
-pathToData = compasRootDir + 'postProcessing/Tutorial/COMPAS_Output/COMPAS_Output.h5'
 
 # -
 
@@ -94,37 +95,70 @@ DCs = Data['BSE_Double_Compact_Objects']
 
 print(SPs.keys())
 
+# One of the most important parameters in the COMPAS output is the system seed. The seed represents the unique identifier to a specific system in a simulation. It is also used as the seed value in random number generation, which is useful when trying to reproduce a given system identically. 
+#
 # If we want to view, say, the random seeds in the system parameters file, we run
 
-spSeeds = SPs['SEED'][()]
-print(spSeeds)
+seedsSP = SPs['SEED'][()]
+print(seedsSP)
 
+# ### printCompasDetails
+#
 # This is useful for extracting the arrays of single parameters, but for a more convenient view of the whole system parameters file, we can use the `printCompasDetails` function.
 
 printCompasDetails(SPs) # Note - the output of this is a pandas dataframe
 
-# PrintCompasDetails optionally also takes seeds and/or a boolean mask as arguments, to help filter the data.
+# `PrintCompasDetails` optionally also takes seeds as arguments, to focus on specific systems.
+
+seedsMT = MTs['SEED'][()]
+firstThreeUniqueSeeds = np.unique(seedsMT)[0:3]
+print("Look at these seeds: ", firstThreeUniqueSeeds)
+printCompasDetails(MTs, firstThreeUniqueSeeds)
+
+# ### getEventHistory
 #
-# Some of the output categories have multiple events for a single seed (e.g `BSE_RLOF` if there are multiple mass transfer events). Using both seeds and mask inputs can help to extract a specifc event from several for that seed. 
+# Often, it is useful to quickly retrieve an overview of the event history of a binary, including all mass transfer, common envelope, and supernova events. For this, we use `getEventHistory`
 
 # +
-# Example: out of the first 20 seeds, only show systems which end as a merger
+seeds, events = getEventHistory(Data)
 
-mtSeeds = MTs['SEED'][()]
-sds, cts = np.unique(mtSeeds, return_counts=True)
-sds[cts>5]
-printCompasDetails(MTs, 1636090318)
+for ii, seed in enumerate(seeds):
+    print(seed, events[ii])
 # -
 
[…]
