### 2021-11-26 c02e9c8
added event and string functions to compasUtils
