### 2021-11-26 TeamCOMPAS/COMPAS#715 @jeffriley
Users can specify options via the COMPAS command-line interface (only those that the user wants to differ from the default values need to be specified).  All pythonSubmit sought to do was provide an .ini-like file because in the past specifying the command-line arguments was a bit cumbersome, but we've made that a lot better.  The only time I ever use pythonSubmit to run COMPAS is when somebody sends me a pythonSubmit file to reproduce a problem, otherwise I just run it from the command line.  I'm all for bringing stroopwafel into the COMPAS suite, but the core COMPAS executable should remain capable of being run standalone.

The grid file functionality effectively provides an enhanced .ini-like capability.
