### 2021-11-29 1f247ac
combined cosmic integration folder into single script, began clean up
