+++ postProcessing/Tutorial/CosmicIntegration/6_FastCosmicIntegrator.py
@@ -1,0 +1,434 @@
+# ---
+# jupyter:
+#   jupytext:
+#     formats: ipynb,py:light
+#     text_representation:
+#       extension: .py
+#       format_name: light
+#       format_version: '1.5'
+#       jupytext_version: 1.12.0
+#   kernelspec:
+#     display_name: Python 3
+#     language: python
+#     name: python3
+# ---
+
+# Magic function to set the backend of matplotlib to the 'inline' backend
+# %matplotlib inline
+
+# # Introduction
+#
+# This last notebook is meant to introduce the "FastCosmicIntegration.py"
+#
+# This script works in a very similar way as the other cosmic integrators, but it is optimised to be run from the terminal, and thus can be used to easily run on HPC and as part of grid runs. 
+#
+# For now this only includes one variation of dP/dZ, namely a skewed log-normal distribution for which the default values result in the log-normal distribution from Neijssel et al. 2019.
+# Following Neijssel et al. 2019, the SFR(z) for this function follows the functional form from Madau & Dickinson 2014, but allows the user to manually adjust the parameters of this function
+#
+#
+
+# # Path definitions
+
+# +
+import os
+
+pathNoteBook    = os.getcwd()
+pathScripts     = pathNoteBook + '/PythonScripts/'
+
+print(pathScripts)
+
+pathData        = '/Users/lieke/surfdrive/Documents/test_CI/COMPAS_Output/'
+# -
+
+# # Imports
+
+# +
+import numpy as np
+import sys
+import matplotlib.pyplot as plt
+import astropy.units as u
+
+#custom scripts
+sys.path.append(pathScripts)
+import FastCosmicIntegration as CI #this imports other routines by itself
+import ClassCOMPAS
+import selection_effects
+import warnings
+
+# To make all your plots look nice
+from matplotlib import rc
+import matplotlib
+matplotlib.rcParams['mathtext.fontset'] = 'stix'
+matplotlib.rcParams['font.family'] = 'STIXGeneral'
+rc('font',**{'family':'sans-serif','sans-serif':['Helvetica']})
+rc('text', usetex=True)
+
+plt.rc('font', family='serif')
+
+# -
+
+# # The magic happens in find_detection_rate()
+#
+# The most important function in FasCosmicIntegration.py is find_detection_rate
+# We will spend most of this notebook breaking down this function.
+#
+#
+# When you run FasCosmicIntegration.py from your terminal, it is run with a lot of default parameters. 
+# These parameters can be changed from your terminal run. 
+#
+# E.g. if you would like to know the rate for BHNS in stead of for BBHs (which is default) you would run
+#
+#  ``` python  FasCosmicIntegration.py --dco_type 'BHNS' ```
+
+# +
+# First define the parameters
+path            = pathData
+filename        ="COMPAS_Output.h5"
+
+# For what DCO would you like the rate?  options: ALL, BHBH, BHNS NSNS
+dco_type        ="BBH"
+weight_column   =None
+                        
+merges_hubble_time     = True
+pessimistic_CEE        = True
+no_RLOF_after_CEE      = True
+
+# Options for the redshift evolution 
+max_redshift           = 10.0
+max_redshift_detection = 2.0
+redshift_step          = 0.001
+z_first_SF             = 10
+
+# Metallicity of the Universe
+min_logZ               =-12.0 
+max_logZ               =0.0 
+step_logZ              =0.01
+
[…]
