### 2021-11-29 5f8a2eb
combined cosmic integration folder into single script, began clean up
