### 2021-11-30 6f6ec62
removed old directories, completely removed development section, further updates to Cosmic Integration notebook
