+++ postProcessing/Tutorial/CosmicIntegration/1_ClassCOMPAS.py
@@ -1,217 +1,0 @@
-# ---
-# jupyter:
-#   jupytext:
-#     formats: ipynb,py:light
-#     text_representation:
-#       extension: .py
-#       format_name: light
-#       format_version: '1.5'
-#       jupytext_version: 1.12.0
-#   kernelspec:
-#     display_name: Python 3
-#     language: python
-#     name: python3
-# ---
-
-# # Introduction
-#
-# In order to calculate the formation rate of a type of double compact object at a given redshift, we need to know:
-#
-# 1 - The grid of metallicities we assume for the integral
-#
-# 2 - The amount of solar mass evolved per metallicity per system $\frac{d}{dMsun}(Z)$
-#
-# 3 - The type of double compact object (DCO) we are interested in 
-#
-# 4 - The metallicity  $Z$ at which each DCO formed and the
-#     delay time (time from formation till merger) $t_c$ for each DCO
-#
-#
-# Given a time at which it merges we can then calculate the time at which it formed to recover the MSSFR ($\frac{dMsun}{dt})$ 
-#
-#
-# In addition we need to know the component masses of the system in order to calculate any selection effects
-#
-#
-# The ClassCOMPAS.py is there to store the information that we need such that we can access it quickly when calculating the rates.
-
-# # Paths
-
-import os
-pathNoteBook     = os.getcwd()
-pathClassCOMPAS  = pathNoteBook + '/PythonScripts/'
-pathData        = '/Users/lieke/surfdrive/Documents/test_CI/COMPAS_Output/'#"/home/cneijssel/Desktop/Test/"
-
-# # Imports
-
-import numpy as np
-import sys
-sys.path.append(pathClassCOMPAS)
-import ClassCOMPAS
-
-# # Creating an instance of the COMPAS data class
-
-# To create an instance of the COMPAS output class, we need to specify the following attributes:
-#
-#     path      = None
-#     fileName  = 'COMPAS_output.h5'
-#     
-#     lazyData  = True
-#
-#     Mlower    = None
-#     Mupper    = None
-#     binaryFraction=None
-
-# Path: 
-#     
-#     path to the h5-data. 
-#     
-# fileName:
-#
-#     name of the h5 data
-#
-# lazyData:
-#
-#     means we store additional info like the mass ratios, chirpmasses of 
-#     each DCO system. In principle this could be done externally but this
-#     is slightly easier when plotting/combingin info later on, 
-#     but it does use more memory
-#     
-#     
-# Mlower:
-#
-#     lower limit used for M1 in the pythonSubmit of the simulation. 
-#     Needed to recover `true' amount of Msun evolved (see step 4)
-#     
-# Mupper:
-#
-#     upper limit used for M1 in the pythonSubmit of the simulation. 
-#     Needed to recover `true' amount of Msun evolved (see step 4)
-#
-# binaryFraction:
-#
-#
-#     assumed fraction of stars in binaries. 
-#     Needed to recover `true' amount of Msun evolved (see step 4)
-#
-
-#I assume all the defaults and just set the path
-COMPASData = ClassCOMPAS.COMPASData(path=pathData)
-
-# The output are  reminders which will be explained in next steps
-
-# # 1 Total mass Evolved
-#
-# In the COMPAS simulation we often only evolve massive stars.
-# This means that the total mass in our simulation does not represent the total mass evolved
-# in all stars. Here we recover an estimate of what that total mass is using the
-# lower and upper mass for the primary from your python submit and assuming a binary fraction.
-#
-# The code will then check your data and test per metallicity how much mass is evolved.
-# It assumes the metallicities are subject to the same pytonSubmit, but maybe due to sampling had a different number of systems. It also recovers from the data what metallicity grid is used
-
-# +
-COMPASData.Mlower = 15
-COMPASData.Mupper = 150
[…]
