### 2021-11-30 481bed9
removed old directories, completely removed development section, further updates to Cosmic Integration notebook
