+++ postProcessing/Tutorial/CosmicIntegration/2_MSSFR-or-SFRD-prescriptions.py
@@ -1,753 +1,0 @@
-# ---
-# jupyter:
-#   jupytext:
-#     formats: ipynb,py:light
-#     text_representation:
-#       extension: .py
-#       format_name: light
-#       format_version: '1.5'
-#       jupytext_version: 1.12.0
-#   kernelspec:
-#     display_name: Python 3
-#     language: python
-#     name: python3
-# ---
-
-# Magic function to set the backend of matplotlib to the 'inline' backend
-# %matplotlib inline
-
-# # Introduction
-#
-#
-# The universe evolved over time and as such the star formation rate and the metallicity of the starforming gas change. The ClassMSSFR is a script which determines the amount of star formation that went into forming a system born at redshift z, in a metallicity bin dZ. In shorthand we will use the term metallicity-specific star formation, either indicated with MSSFR or SFRD(Z,z), which is a function of metallicity Z and redshift z. We will both use MSSFR and SFRD(Z,z) throughout this notebook and they are identical (just two different ways of naming the concept). 
-#
-# A schematic picture of how the SFRD(Z,z) is constructed is given in Figure 10 of the COMPAS methods paper and added below. 
-#
-# ![SFRD_cartoon-1.png](attachment:SFRD_cartoon-1.png)
-#
-#
-# Typically, the SFRD(Z,z) (MSSFR) is created by combining a Star formation rate density (SFRD) with a metallicity distribution function dP/dZ. The latter is typically constructed by combining a galaxy stellar-mass function (GSMF) with a mass-metallicity relationship
-#
-# For this script we assume that the SFR(z) and the metallicity distribution are independent of each other.
-# Hence the SFR is simply a normalization to the PDF of the metallicity distribution. Note that these distributions are averages over all galaxies at a specific redshift z and hence do not apply to specific galaxies.
-#
-# The script allows for several combinations of distributions.
-# To this end we have to define the following steps
-#
-#
-# 1 - Instance of the MSSFR class where we define constants
-#
-# 2 - Pick a SFR distribution
-#
-# 3 - Pick a metallicity distribution.
-#
-# Note that although this is meant to be used in the cosmic integration pipeline, 
-# this class can be independently used to plot distributions or as input for other calculations
-#
-# 4 - Example of plotting
-#
-# 5 - Examples of plotting the SFRD, the GSMF and the MZR relations (Figures 11, 13, and 14 from the COMPAS methods paper) 
-#
-#
-
-# # Paths needed
-
-# +
-import os
-
-pathNoteBook     = os.getcwd()
-pathClassMSSFR  = pathNoteBook + '/PythonScripts/'
-
-
-# -
-
-# # Imports
-
-import numpy as np
-import matplotlib.pyplot as plt
-import sys
-#custom scripts
-sys.path.append(pathClassMSSFR)
-import ClassMSSFR
-
-# # 1 - Instance MSSFR class
-
-# To create an instance we need to define the following parameters;
-#
-# Here are the possible parameters and their default values
-#
-#     verbose             =False,
-#     
-#     metallicityGrid     =None,
-#     
-#     binInLogSpace        =True,
-#     metallicityLowerLimit=1E-90, 
[…]
