+++ postProcessing/Tutorial/CosmicIntegration/4_MergersAtSingleRedshift.py
@@ -1,216 +1,0 @@
-# ---
-# jupyter:
-#   jupytext:
-#     formats: ipynb,py:light
-#     text_representation:
-#       extension: .py
-#       format_name: light
-#       format_version: '1.5'
-#       jupytext_version: 1.12.0
-#   kernelspec:
-#     display_name: Python 3
-#     language: python
-#     name: python3
-# ---
-
-# Magic function to set the backend of matplotlib to the 'inline' backend
-# %matplotlib inline
-
-# # Introduction
-#
-# Here we show how to calculate the merger rate density of systems
-# merging at a single redshift z. By now we assume you understand the indiviual pipelines of;
-#
-# ClassCOMPAS:
-#
-#     -- (handling the (mock) data needed)
-#     
-# ClassMSSFR:
-#
-#     -- defining the model for metallicity specific SFR
-#     
-# selection_effects:
-#
-#     -- module to calculate probility detecting a system
-#
-#
-#
-# Here we show which additional steps are needed for the calculation.
-# All these steps are done by default in the ClassCosmicIntegrator.
-# We highlight the steps here outside the function for clarity
-# since the ClassCosmicIntegrator merely acts as a giant for loop over multiple redshifts and a way to conveniently store the results
-
-# # Paths
-
-# +
-import os
-
-pathNoteBook    = os.getcwd()
-pathScripts     = pathNoteBook + '/PythonScripts/'
-
-
-pathData        = '/Users/lieke/surfdrive/Documents/test_CI/COMPAS_Output/'#"/home/cneijssel/Desktop/Test/"
-
-# -
-
-# # Imports
-
-# +
-import numpy as np
-import sys
-import matplotlib.pyplot as plt
-
-from   astropy.cosmology import WMAP9 as cosmology
-from   scipy.optimize import newton
-#custom scripts
-sys.path.append(pathScripts)
-import ClassCOMPAS 
-import ClassMSSFR  
-
-
-
-# -
-
-# # 1- Set up data and MSSFR model
-
-# +
-# Create instance COMPAS data class
-
-COMPAS = ClassCOMPAS.COMPASData(path=pathData, fileName='COMPAS_Output.h5')
-
-
-# +
-#Set the type of DCO of interest and recover their parameters
-COMPAS.Mlower = 15
-COMPAS.Mupper = 150
-COMPAS.binaryFraction =0.7
-
-COMPAS.setGridAndMassEvolved()
-
-COMPAS.setCOMPASDCOmask() #Pessimistic BBHs
-COMPAS.setCOMPASData()
-
-# +
-# The MSSFR model
-
-#use the metallicityGrid of the Data
-metallicityGrid = COMPAS.metallicityGrid
-
-#Create instance Class
-MSSFR = ClassMSSFR.MSSFR(metallicityGrid=metallicityGrid)
-
-# -
-
-#Set the MSSFR model
-MSSFR.SFRprescription = 'Neijssel et al. (2019)'
-MSSFR.Zprescription = 'logNormal'
-MSSFR.logNormalPrescription ='Neijssel Phenomenological'
-
-# # 2 - Define the redshifts
-#
-# The entire calculation depends on defining a redshift
-# at which the DCOs merge. Then using the delaytimes and astropy
-# you need to recover when the systems were born
-#
-# First define our cosmology
-
-# +
-# see top notebook
-
-# +
-mergerRedshift = 0.2
-
-#Define an age when the first stars formed based on redshift
-firstSFR = cosmology.age(10).value
-#single value in units of Gyr
-ageUniverseAtMergerGyr = cosmology.age(mergerRedshift)
-#Recover the array delaytimes in Units of Gyr
[…]
