+++ postProcessing/Tutorial/README.txt
@@ -1,0 +1,3 @@
+This folder contains jupyter notebooks that function as a tutorial for new COMPAS users in how to 
+do basic post-processing tasks. To run through the tutorial linearly, start with the Overview.ipynb file
+and follow the links therein.
