### 2021-11-30 09fe456
moved joss-logo.png file to docs/media/, cleaned up scripts that are now outdated in postprocessing
