### 2021-11-30 17792b7
moved online-docs and media to the root dir, and added web pages for the former markdown files for sampling, docker, and dev-git-workflow
