+++ docs/online-docs/pages/Getting started/COMPAS-dependencies.rst
@@ -1,13 +1,0 @@
-Installing dependencies
-=======================
-
-COMPAS requires a ``C++`` compiler, and the ``gsl``, ``boost``, and ``hdf5`` libraries.  ``Python`` is required for the COMPAS post-processing tools.
-
-
-.. toctree::
-   :maxdepth: 1
-
-   ./COMPAS-dependencies-linux
-   ./COMPAS-dependencies-macOS
-   ./COMPAS-dependencies-python
-
