+++ docs/online-docs/pages/Developer guide/Services/Random numbers/services-random-numbers-randomInt2.rst
@@ -1,10 +1,0 @@
-Rand::RandomInt(p_Upper)
-========================
-
-::
-
-    INT Rand::RandomInt(const INT p_Upper)
-    
-Returns a random integer number uniformly distributed in the range [0, ``p_Upper``), where 0 :math:`\small \leq` ``p_Upper``.
-    
-Returns 0 if ``p_Upper`` :math:`\small \lt` 0.
