+++ requirements.txt
@@ -1,0 +1,42 @@
+# File: docs/requirements.txt
+
+# Defining the exact version will make sure things don't break
+
+# sphinx is required.  readthedocs does that automatically, so it doesn't need
+# to be here but you will need it if you want to build the html files locally
+
+# seems obvious, but python is required.  readthedocs does that automatically, 
+# so it doesn't need to be here but you will need it if you want to build the 
+# html files locally
+
+sphinx_rtd_theme==0.5.1
+sphinx-math-dollar>=1.2
+
+sphinxcontrib.bibtex
+
+linuxdoc==20210324
+
+nbsphinx
+
+ipykernel
+IPython
+
+numpy >= 1.16'
+numpydoc
+astropy >= 4.0'
+numba >= 0.50'
+scipy >= 1.5.0'
+h5py
+matplotlib >= 3.3.2'
+
+# The following are listed in the 'extensions []' block in conf.py
+# Readthedocs doesn't want them here, but they are required if you want
+# to build the html files locally
+#
+# matplotlib.sphinxext.plot_directive
+# sphinx.ext.mathjax
+# sphinx.ext.intersphinx
+# matplotlib.sphinxext.plot_directive
+# IPython.sphinxext.ipython_console_highlighting
+# IPython.sphinxext.ipython_directive
+
