+++ docs/online-docs/pages/User guide/COMPAS output/standard-logfiles-default-record-specifications-BSE-ce.rst
@@ -1,81 +1,0 @@
-BSE common envelopes
-====================
-
-Default record definition for the BSE Common Envelopes log file::
-
-    const ANY_PROPERTY_VECTOR BSE_COMMON_ENVELOPES_REC = {
-        BINARY_PROPERTY::RANDOM_SEED,
-        BINARY_PROPERTY::TIME,
-        STAR_1_PROPERTY::LAMBDA_AT_COMMON_ENVELOPE,
-        STAR_2_PROPERTY::LAMBDA_AT_COMMON_ENVELOPE,
-        STAR_1_PROPERTY::BINDING_ENERGY_PRE_COMMON_ENVELOPE,
-        STAR_2_PROPERTY::BINDING_ENERGY_PRE_COMMON_ENVELOPE,
-        BINARY_PROPERTY::ECCENTRICITY_PRE_COMMON_ENVELOPE,
-        BINARY_PROPERTY::ECCENTRICITY_POST_COMMON_ENVELOPE,
-        BINARY_PROPERTY::SEMI_MAJOR_AXIS_PRE_COMMON_ENVELOPE,
-        BINARY_PROPERTY::SEMI_MAJOR_AXIS_POST_COMMON_ENVELOPE,
-        BINARY_PROPERTY::ROCHE_LOBE_RADIUS_1_PRE_COMMON_ENVELOPE,
-        BINARY_PROPERTY::ROCHE_LOBE_RADIUS_1_POST_COMMON_ENVELOPE,
-        BINARY_PROPERTY::ROCHE_LOBE_RADIUS_2_PRE_COMMON_ENVELOPE,
-        BINARY_PROPERTY::ROCHE_LOBE_RADIUS_2_POST_COMMON_ENVELOPE,
-        BINARY_PROPERTY::MASS_1_PRE_COMMON_ENVELOPE,
-        BINARY_PROPERTY::MASS_1_POST_COMMON_ENVELOPE,
-        BINARY_PROPERTY::MASS_ENV_1,
-        BINARY_PROPERTY::RADIUS_1_PRE_COMMON_ENVELOPE,
-        BINARY_PROPERTY::RADIUS_1_POST_COMMON_ENVELOPE,
-        BINARY_PROPERTY::STELLAR_TYPE_1_PRE_COMMON_ENVELOPE,
-        STAR_1_PROPERTY::STELLAR_TYPE,
-        STAR_1_PROPERTY::LAMBDA_FIXED,
-        STAR_1_PROPERTY::LAMBDA_NANJING,
-        STAR_1_PROPERTY::LAMBDA_LOVERIDGE,
-        STAR_1_PROPERTY::LAMBDA_LOVERIDGE_WINDS,
-        STAR_1_PROPERTY::LAMBDA_KRUCKOW,
-        STAR_1_PROPERTY::BINDING_ENERGY_FIXED,
-        STAR_1_PROPERTY::BINDING_ENERGY_NANJING,
-        STAR_1_PROPERTY::BINDING_ENERGY_LOVERIDGE,
-        STAR_1_PROPERTY::BINDING_ENERGY_LOVERIDGE_WINDS,
-        STAR_1_PROPERTY::BINDING_ENERGY_KRUCKOW,
-        BINARY_PROPERTY::MASS_2_PRE_COMMON_ENVELOPE,
-        BINARY_PROPERTY::MASS_2_POST_COMMON_ENVELOPE,
-        BINARY_PROPERTY::MASS_ENV_2,
-        BINARY_PROPERTY::RADIUS_2_PRE_COMMON_ENVELOPE,
[…]
