+++ docs/online-docs/pages/Developer guide/Services/Random numbers/services-random-numbers-randomGaussian.rst
@@ -1,8 +1,0 @@
-Rand::RandomGaussian(p_Sigma)
-=============================
-
-::
-
-    DOUBLE Rand::RandomGaussian(const DOUBLE p_Sigma)
-
-Returns a Gaussian random variate, with mean 0.0 and standard deviation ``p_Sigma``.

[…]
