### 2021-11-30 5bc745d
moved online-docs and media to the root dir, and added web pages for the former markdown files for sampling, docker, and dev-git-workflow
