+++ online-docs/pages/Getting started/getting-started.rst
@@ -14,6 +14,7 @@ To start using COMPAS, get a copy of the code, and install the libraries and too
    ./git-details
    ./COMPAS-dependencies
    ./building-COMPAS
+   ./dev-git-workflow
 
 Once you have completed the steps shown above, you're ready to run COMPAS. The :doc:`COMPAS User Guide <../User guide/user-guide>`
 explains in detail how to run COMPAS, but to check that COMPAS is installed correctly, and to get a taste of what running COMPAS looks
