+++ online-docs/pages/User guide/docker.rst
@@ -1,0 +1,351 @@
+COMPAS and Docker
+=================
+
+Docker has been added to COMPAS to reduce time and effort required to
+set up the COMPAS deployment environment.
+
+Instead of having to install and configure several libraries and tools
+(e.g. python/pip, numpy, g++, boost) which can vary considerably beween
+operating systems and existing toolchains, users can instead opt to
+install Docker and run COMPAS with a single command.
+
+This also gives users the ability to run COMPAS on cloud solutions like
+`AWS EC2 <https://aws.amazon.com/ec2/>`__ or `Google Compute
+Engine <https://cloud.google.com/compute>`__ where hundreds of cores can
+be provisioned without having to manually configure the environment.
+
+| Docker works by creating an isolated and standalone environment known
+as a `container <https://www.docker.com/resources/what-container>`__.
+| Containers can be created or destroyed without affecting the host
+machine or other containers\*.
+
+Containers are instances of images. An image is a pre-defined
+setup/environment that is instantiated when started as a container
+(containers are to images what objects are to classes). More
+`here <https://stackoverflow.com/questions/23735149/what-is-the-difference-between-a-docker-image-and-a-container#:~:text=An%20instance%20of%20an%20image,of%20layers%20as%20you%20describe.&text=You%20can%20see%20all%20your,an%20image%20is%20a%20container.>`__
+on the relationship between images and container.
+
+Containers are (almost) always run as a Linux environment. A major
+benefit of this is the ability to run Linux applications in a Windows or
+MacOS environment without having to jump through hoops or have a
+diminished experience.
+
+Image definitions can be defined by users (e.g. Dockerfiles); there are
+also standard images publicly available on `Docker
+Hub <https://hub.docker.com/>`__
+
+| All that is required to start using COMPAS with Docker is the "Usage"
+section (the "CI/CD" section is also highly recommended).
+| The other sections are provided for extra info.
+
+\* Containers can still interact with each other and the host machine
+through mounted directories/files or exposed ports.
+
+--------------
+
+Usage
+-----
+
+| N.B. This section assumes `Docker <https://www.docker.com/>`__ has
+been installed and is running.
+| For Windows and MacOS users, see
+`here <https://www.docker.com/products/docker-desktop>`__.
+
+Installing
+~~~~~~~~~~
+
+| The latest compiled version of COMPAS (dev branch) can be retrieved by
+running
+| ``docker pull teamcompas/compas``
+
+| Other versions can be used by adding a version
+`tag <https://docs.docker.com/engine/reference/commandline/tag/>`__.
+| For example, COMPAS version 2.12.0 would be
+``teamcompas/compas:2.12.0``.
+| To see all available versions, go to the TeamCOMPAS docker hub page
+`here <https://hub.docker.com/u/teamcompas>`__.
+
+Running
+~~~~~~~
+
+COMPAS can still be configured via command line arguments passed to the
+COMPAS executable or via a ``pythonSubmit.py`` file.
+
+Run pythonSubmit.py
+^^^^^^^^^^^^^^^^^^^
+
+To run COMPAS via a ``pythonSubmit.py`` file, the command is a little
+more complex.
+
+::
+
+    docker run                                                  \
+        --rm                                                    \
+        -it                                                     \
+        -v $(pwd)/compas-logs:/app/COMPAS/logs                  \
[…]

[…]
