### 2021-11-30 972ab61
cleaned up notebooks so that they only reference ipynb, not py files
