### 2021-11-30 fe6bea9
cleaned up notebooks and hdf5 section of the online-docs (post-processing), and added in sampling, docker, and dev-git-workflow pages - TODO, still need to make sure dev-git-workflow does not overlap with other git related web pages
