+++ online-docs/pages/User guide/docker.rst
@@ -14,9 +14,9 @@ This also gives users the ability to run COMPAS on cloud solutions like
 Engine <https://cloud.google.com/compute>`__ where hundreds of cores can
 be provisioned without having to manually configure the environment.
 
-| Docker works by creating an isolated and standalone environment known
+Docker works by creating an isolated and standalone environment known
 as a `container <https://www.docker.com/resources/what-container>`__.
-| Containers can be created or destroyed without affecting the host
+Containers can be created or destroyed without affecting the host
 machine or other containers\*.
 
 Containers are instances of images. An image is a pre-defined
@@ -34,45 +34,45 @@ Image definitions can be defined by users (e.g. Dockerfiles); there are
 also standard images publicly available on `Docker
 Hub <https://hub.docker.com/>`__
 
-| All that is required to start using COMPAS with Docker is the "Usage"
+All that is required to start using COMPAS with Docker is the "Usage"
 section (the "CI/CD" section is also highly recommended).
-| The other sections are provided for extra info.
+The other sections are provided for extra info.
 
 \* Containers can still interact with each other and the host machine
 through mounted directories/files or exposed ports.
 
---------------
+
 
 Usage
------
 
-| N.B. This section assumes `Docker <https://www.docker.com/>`__ has
+
+N.B. This section assumes `Docker <https://www.docker.com/>`__ has
 been installed and is running.
-| For Windows and MacOS users, see
+For Windows and MacOS users, see
 `here <https://www.docker.com/products/docker-desktop>`__.
 
 Installing
-~~~~~~~~~~
 
-| The latest compiled version of COMPAS (dev branch) can be retrieved by
+
+The latest compiled version of COMPAS (dev branch) can be retrieved by
 running
-| ``docker pull teamcompas/compas``
+``docker pull teamcompas/compas``
 
-| Other versions can be used by adding a version
+Other versions can be used by adding a version
 `tag <https://docs.docker.com/engine/reference/commandline/tag/>`__.
-| For example, COMPAS version 2.12.0 would be
+For example, COMPAS version 2.12.0 would be
 ``teamcompas/compas:2.12.0``.
-| To see all available versions, go to the TeamCOMPAS docker hub page
+To see all available versions, go to the TeamCOMPAS docker hub page
 `here <https://hub.docker.com/u/teamcompas>`__.
 
 Running
-~~~~~~~
+
 
 COMPAS can still be configured via command line arguments passed to the
 COMPAS executable or via a ``pythonSubmit.py`` file.
 
 Run pythonSubmit.py
-^^^^^^^^^^^^^^^^^^^
+
 
 To run COMPAS via a ``pythonSubmit.py`` file, the command is a little
 more complex.
@@ -80,98 +80,98 @@ more complex.
 ::
 
     docker run                                                  \
-        --rm                                                    \
-        -it                                                     \
-        -v $(pwd)/compas-logs:/app/COMPAS/logs                  \
-        -v $(pwd)/pythonSubmit.py:/app/starts/pythonSubmit.py   \
-        -e COMPAS_EXECUTABLE_PATH=/app/COMPAS/bin/COMPAS        \
-        -e COMPAS_LOGS_OUTPUT_DIR_PATH=/app/COMPAS/logs         \
-        teamcompas/compas                                       \
-        python3 /app/starts/pythonSubmit.py                     
+    --rm                                                    \
+    -it                                                     \
[…]
