### 2021-11-30 5c2cb85
cleaned up notebooks and hdf5 section of the online-docs (post-processing), and added in sampling, docker, and dev-git-workflow pages - TODO, still need to make sure dev-git-workflow does not overlap with other git related web pages
