+++ online-docs/pages/User guide/sampling.rst
@@ -1,7 +1,7 @@
 Sampling in COMPAS
 ==================
 
---------------
+
 
 Here are some basic instructions for efficient sampling of the COMPAS
 input parameters, using the python sampling package Stroopwafel.
@@ -11,18 +11,18 @@ Importance Sampling" is not yet implemented, but is currently in
 development.
 
 Requirements
-------------
 
---------------
+
+
 
 If you have not already, you will need to install Stroopwafel. If you
 have admin rights, Stroopwafel can be installed on your system with
 ``pip install stroopwafel``.
 
 Instructions
-------------
 
---------------
+
+
 
 To use Stroopwafel sampling, copy
 ``preProcessing/stroopwafelInterface.py`` into your working directory.
@@ -31,36 +31,33 @@ Settings
 ~~~~~~~~
 
 1. PythonSubmit
-^^^^^^^^^^^^^^^
+
 
 If you are running COMPAS on default settings, skip this section.
 
-| If you have many non-default COMPAS arguments, you may want to set
-them in the
-| ``pythonSubmit.py`` file in the same directory. For now, the file must
-be named this way
-| and placed in the same directory as the ``stroopwafelInterface.py``
+If you have many non-default COMPAS arguments, you may want to set
+them in the ``pythonSubmit.py`` file in the same directory. For now, the file must
+be named this way and placed in the same directory as the ``stroopwafelInterface.py``
 file.
 
 A configurable pythonSubmit file can be found in the ``preProcessing/``
 directory.
 
-| Set your desired options, then set the ``usePythonSubmit`` parameter
-to ``True``
-| in the ``stroopwafelInterface.py``.
+Set your desired options, then set the ``usePythonSubmit`` parameter to ``True``
+in the ``stroopwafelInterface.py``.
 
 2. Stroopwafel inputs
-^^^^^^^^^^^^^^^^^^^^^
 
-| The lines below ``usePythonSubmit`` represent stroopwafel inputs.
+
+The lines below ``usePythonSubmit`` represent stroopwafel inputs.
 These are treated as
-| defaults, but can be overriden by command-line arguments to
+defaults, but can be overriden by command-line arguments to
 stroopwafel.
-| See ``python3 stroopwafelInterface.py --help``.
+See ``python3 stroopwafelInterface.py --help``.
 
-| ``num_systems`` is the total number of binaries you would like to
+``num_systems`` is the total number of binaries you would like to
 evolve.
-| This value overrides the value set in the pythonSubmit.
+This value overrides the value set in the pythonSubmit.
 
 ``num_cores`` is the number of cores you would like to use to
 parallelize your run. More cores means your run will finish sooner, but
@@ -75,9 +72,9 @@ This is more relevent for adaptive importance sampling.
 ``mc_only`` specifies if you would like to do naive MC sampling only.
 Currently, this option must be set to True
 
-| ``run_on_hpc`` specifies if you are running on a High-Performance
+``run_on_hpc`` specifies if you are running on a High-Performance
 Computer (HPC).
-| If so, see `docs/compasHPC.md <compasHPC.md>`__ for assistance.
+If so, see `docs/compasHPC.md <compasHPC.md>`__ for assistance.
 
 ``output_folder`` a string specifying the output folder. Relative paths
 will be appended onto the current directory path.
@@ -88,28 +85,28 @@ file.
 ``debug`` whether to print the COMPAS output/error.
 
 3. Sampling parameters
-^^^^^^^^^^^^^^^^^^^^^^
 
-| Sampled parameters will be combined into grid files which COMPAS then
+
+Sampled parameters will be combined into grid files which COMPAS then
 reads in.
-| Users should choose which parameters they would like to be sampled
[…]
