### 2021-11-30 a56f9c1
Ran CosmicIntegration with small Compas data file so that the web version shows useful values
