+++ postProcessing/Tutorial/CosmicIntegration.py
@@ -121,7 +121,6 @@ import selection_effects
 # To create an instance of the COMPAS output class, we need to specify the following attributes:
 #
 #     path      = None
-#     fileName  = 'COMPAS_output.h5'
 #     
 #     lazyData  = True
 #
@@ -133,9 +132,6 @@ import selection_effects
 #     
 #     path to the h5-data. 
 #     
-# fileName:
-#
-#     name of the h5 data
 #
 # lazyData:
 #
@@ -1015,7 +1011,7 @@ print(P)
 # +
 # Create instance COMPAS data class
 
-COMPAS = ClassCOMPAS.COMPASData(path='COMPAS_CI_Output.h5')
+COMPAS = ClassCOMPAS.COMPASData(path=pathData)
 
 
 # +
@@ -1052,9 +1048,6 @@ MSSFR.logNormalPrescription ='Neijssel Phenomenological'
 # you need to recover when the systems were born
 #
 # First define our cosmology
-
-# +
-# see top notebook
 
 # +
 mergerRedshift = 0.2
@@ -1471,11 +1464,11 @@ for t in types:
     print('------total predicted rate for {} = {} mergers per year'.format(t, observedRate))
 # -
 
-# Note that his means that if you want to loop over 
+# Note that this means that if you want to loop over 
 # both DCO types and SFR prescriptions, it is more time-efficient
 # to loop over types first and then SFR prescriptions.
 #
-# With this you should be all set. good luck
+# With this you should be all set. Good luck!
 # For more info on how to create plots using this pipeline see
 # the plotting library in the postProcessing notes :)
 #     
@@ -1509,7 +1502,6 @@ for t in types:
 
 # +
 # First define the parameters
-path            = "COMPAS_CI_Output.h5"
 
 # For what DCO would you like the rate?  options: ALL, BHBH, BHNS NSNS
 dco_type        = "BBH"
@@ -1620,7 +1612,7 @@ if snr_step > 1.0:
 # ### Use functions in ClassCOMPAS to read your data
 
 # start by getting the necessary data from the COMPAS file
-COMPAS = ClassCOMPAS.COMPASData(path, fileName=filename, Mlower=m1_min, Mupper=m1_max, m2_min=m2_min, binaryFraction=fbin, suppress_reminder=True)
+COMPAS = ClassCOMPAS.COMPASData(path, Mlower=m1_min, Mupper=m1_max, m2_min=m2_min, binaryFraction=fbin, suppress_reminder=True)
 COMPAS.setCOMPASDCOmask(types=dco_type, withinHubbleTime=merges_hubble_time, pessimistic=pessimistic_CEE, noRLOFafterCEE=no_RLOF_after_CEE)
 COMPAS.setCOMPASData()
 COMPAS.set_sw_weights(weight_column)
@@ -1653,7 +1645,7 @@ if max(chirp_masses)*(1+max_redshift_detection) < Mc_max:
 # mostly a list of redsifts and their corresponding cosmological times
 
 # calculate the redshifts array and its equivalents
-redshifts, n_redshifts_detection, times, time_first_SF, distances, shell_volumes = CI.calculate_redshift_related_params(max_redshift, max_redshift_detection, redshift_step, z_first_SF)
+redshifts, n_redshifts_detection, times, time_first_SF, distances, shell_volumes = FCI.calculate_redshift_related_params(max_redshift, max_redshift_detection, redshift_step, z_first_SF)
 
 
 print('redshifts', redshifts, '\ntime_first_SF', time_first_SF, '\nshell_volumes', shell_volumes)
@@ -1667,7 +1659,7 @@ print('redshifts', redshifts, '\ntime_first_SF', time_first_SF, '\nshell_volumes
 #
 
[…]
