+++ postProcessing/Tutorial/CosmicIntegration.py
@@ -1,1832 +1,0 @@
-# ---
-# jupyter:
-#   jupytext:
-#     formats: ipynb,py:light
-#     text_representation:
-#       extension: .py
-#       format_name: light
-#       format_version: '1.5'
-#       jupytext_version: 1.11.2
-#   kernelspec:
-#     display_name: Python 3
-#     language: python
-#     name: python3
-# ---
-
-# # Introduction
-#
-# This section covers data processing related to Cosmic Integration (see [Neijssel et al. 2020](https://arxiv.org/abs/1906.08136)). These notes merely show how to call the functions and how to construct the pipeline. It does not offer any derivations.
-#
-# For these notes we assume that you have a sufficiently large COMPAS h5-data set with non-constant (preferably, randomly-sampled) metallicity and some number of double compact objects. If not, see Section 1 [Working With HDF5](./WorkingWithHDF5.py), or download some data from our [Zenodo database](https://zenodo.org/communities/compas/?page=1&size=20), though note that some of the Zenodo data may be out of date (you may have to rename some parameter names to get this file to run).
-#
-# The duration from the birth of the binary until the merger as a double compact object (DCO) can range from a few million years (lifetime of the stars) to more than 100 Gigayears depending on the evolution of the system.
-#
-# This means two things
-#
-# 1. DCOs merge at different redshifts
-# 2. Multiple DCOs merging at a specific redshift could have formed at different times.
-#     
-# We thus need to know the star formation that went into forming a single system. However, the star formation rate is non-constant over the lifetime of the universe. Furthermore, star formation is heavily dependent on the metallicity of the star forming gas, which also changes over the lifetime of the universe. Combined, we call this the metallicity-specific star formation rate (MSSFR).
-#
-# The cosmic-intergation pipeline tries to predict the population of DCO mergers assuming a model for the MSSFR.
-# Here we show how to use the combination of scripts. We show the purpose of each separate script in the pipeline and how to combine them to derive our results. 
-#
-# Note that although these pipelines are tailored for a COMPAS simulation, with some slight adjustments one could use these same pipelines for other (mock)-simulations provided all the information is given (see [Setting the Data](#1.-Setting-the-Data). 
-#
-# If you make use of any these pipelines, we would appreciate if you cite the following papers:
-#
-# Cosmic Integration https://arxiv.org/abs/1906.08136
-#
-# Selection effects https://arxiv.org/pdf/1711.06287
-
-# ## Material
-#         
-#
-# ### [1) Setting the Data ](#1.-Setting-the-Data)
-# Set the required columns and other information.
-#
-# ### [2) Setting the MSSFR ](#2.-Setting-the-MSSFR) 
-# Choose a model and plot it.
-#
-# ### [3) Setting the Selection Effects ](#3.-Setting-the-Selection-Effects)
-# Choose the sensitivity and calculate probability.
-#
-# ### [4) Rate at single redshift ](#4.-Rate-at-single-redshift)
-# Combine data, MSSFR, and selection effects.
-#
-# ### [5) Rate as function of redshift ](#5.-Rate-as-function-of-redshift)
-# Set the integral and combine results.
-#
-# ### [6) Fast Cosmic Integration ](#6.-Fast-Cosmic-Integration)
-# Do the whole routine in one go, using the phenomenological variation of the MSSFR and SFRD
-
-
-
-# ## For the following sections, you will need to have the following python packages installed. 
-# ## `numpy, h5py, pandas, matplotlib, astropy, scipy`
-
-
-# +
-import os, sys
-import numpy as np
-import matplotlib
[…]
