+++ postProcessing/Tutorial/DataAnalysis.py
@@ -1,506 +1,0 @@
-# ---
-# jupyter:
-#   jupytext:
-#     formats: ipynb,py:light
-#     text_representation:
-#       extension: .py
-#       format_name: light
-#       format_version: '1.5'
-#       jupytext_version: 1.11.2
-#   kernelspec:
-#     display_name: Python 3
-#     language: python
-#     name: python3
-# ---
-
-# # Introduction
-#
-# Most of the post-processing material presented here is written in python, and makes extensive use of the numpy package for rapid computation on large data arrays.
-# Here we show two important basics of python/numpy in the context of investigating your COMPAS simulation.
-
-
-# ## Material
-#
-# ### [1) Inspecting the data ](#1.-Inspecting-the-data)
-# Look at the data to see which parameters are available and check that it matches expectations.
-#
-# ### [2) Slicing the data ](#2.-Slicing-the-data)
-# Select specific systems and their parameters using seeds.
-#
-# ### [3) Visualizing the data ](#3.-Visualizing-the-data)
-# Binning and visualising your data.
-
-
-
-
-# ### For the following sections, you will need to have the following packages installed.
-# ### `numpy, h5py, time, matplotlib`
-
-# +
-#python libraries
-import os, sys
-import numpy as np               # for handling arrays
-import h5py as h5                # for reading the COMPAS data
-import time                      # for finding computation time
-import matplotlib.pyplot as plt  #for plotting
-
-# Import COMPAS specific scripts
-compasRootDir = os.environ['COMPAS_ROOT_DIR'] 
-sys.path.append(compasRootDir + 'postProcessing/PythonScripts')
-from compasUtils import printCompasDetails, getEventHistory, getEventStrings
-
-# Choose an output hdf5 file to work with
-pathToData = 'COMPAS_Tutorial_Output.h5'
-
-# This is known as an ipython magic command, and allows plots to be produced within the notebook
-# %matplotlib inline
-
-
-# -
-
-
-
-# # 1. Inspecting the data
-#
-# Often the first thing you want to do with new data is simply to look at it! Getting familiar with the data, including available parameters, size of the data file, etc. will help to inform how best to proceed with the analysis. We provide several useful functions for inspecting the data, `printCompasDetails`, `getEventHistory`, and `getEventStrings`.
-#
-# --
-#
-# If you want to create an alternative COMPAS_Output.h5, see Section 1 [Working With HDF5](./WorkingWithHDF5.ipynb), or download some data from our [Zenodo database](https://zenodo.org/communities/compas/?page=1&size=20).
-
-# *Note:* These cells may take a long time if you test them on large datasets.
-
-Data  = h5.File(pathToData)
-print(list(Data.keys()))
-
-# The output above represents the event categories available from the particular run. If you used the output produced in the previous tutorial, you should see `['BSE_Common_Envelopes', 'BSE_Double_Compact_Objects', 'BSE_RLOF', 'BSE_Supernovae', 'BSE_System_Parameters', 'Run_Details']`. Note that for smaller runs which do not produce any of a particular type of output, the output category will not be created. 
-#
-# Brief description of the categories:
-# - 'BSE_System_Parameters': Initial state of the binary
-# - 'BSE_RLOF': Any mass transfer events that occured within the binary
-# - 'BSE_Common_Envelopes': If any of the mass transfer events were unstable, details will be included here.
-# - 'BSE_Supernovae': Parameters and outcome of any supernovae that occured in the binary
[…]
