### 2021-11-30 TeamCOMPAS/COMPAS#715 @avigna
@HeloiseS thank you very much for the scripts you shared: they read great and will be used to create a revised version of the python code to submit COMPAS runs. We agree with most of what you suggest, and plan to address the suggestions in a revised version of the scripts. There are two items I want to comment on:

1. Refactor to a class: currently, the stroopwaffel script uses the pythonProgramOption class. Moreover, there might be additional and alternative approaches that might benefit from a pythonProgramOption class. While we will clean the code, we currently believe that we should keep pythonProgramOption as a class and only change later if wanted or needed (post JOSS review).
2. We will use a yaml file to create parse the options. For the reivew, we will upload a file with the default options. Once the JOSS review is done, I will open an enhancement "issue" to automatically generate a yaml file with the default COMPAS options/setting, in order avoid confussion and conflicting defaults.

If you do not have any objections with this, I will work on the cleaning and get back to y'all asap. Otherwise, we can keep fine tuning.
