### 2021-11-30 c4f5c4a
Resolve sudden mass drop at end of HG
