### 2021-11-30 786684d
Resolve sudden mass drop at end of HG
