+++ src/changelog.h
@@ -848,7 +848,11 @@
 // 02.25.10     JR - Nov 19, 2021    - Defect repairs:
 //                                      - clamp timestep returned in BaseStar::CalculateTimestep() to NUCLEAR_MINIMUM_TIMESTEP
 //                                      - change NUCLEAR_MINIMUM_TIMESTEP to 1 year (from 100 years) in constants.h
+// 02.26.00     IM - Nov 30, 2021    - Defect repairs:
+//                                      - only decrease effective initial mass for HG and HeHG stars on mass loss when this decrease would not drive an unphysical decrease in the core mass
+//                                      - change mass comparisons (e.g., mass vs. He flash mass threshold) to compare effective initial mass rather than current mass
+//                                      - minor code and comment cleanup
 
-const std::string VERSION_STRING = "02.25.10";
+const std::string VERSION_STRING = "02.26.00";
 
 # endif // __changelog_h__
