+++ preProcessing/runSubmit.py
@@ -1,0 +1,144 @@
+import numpy as np
+import subprocess
+import sys
+import os
+import pickle
+import itertools 
+from subprocess import call
+import yaml
+
+# Check if we are using python 3
+python_version = sys.version_info[0]
+print("python_version =", python_version)
+
+class pythonProgramOptions: # TODO: rename? | HFS oct 31 2021
+    """
+    A class to store and access COMPAS program options in python
+    """
+
+    def __init__(self, config_file='compasConfigDefault.yaml', grid_filename=None, randomSeedFileName='randomSeed.txt', output_directory=None):
+        # Do './COMPAS --help' to see all options
+        #-- Define variables
+
+        # environment variable COMPAS_EXECUTABLE_PATH is used for docker runs
+        # if COMPAS_EXECUTABLE_PATH is not set (== None) we assume this is an
+        # interactive run with python3
+        # if COMPAS_EXECUTABLE_PATH is set (!= None) we assume this is a run
+        # inside a docker container - we have different directories inside a 
+        # docker container (src, obj, bin), and the COMPAS executable resides
+        # in the bin directory (rather than the src directory)
+        
+        # Load yaml file with options
+        with open(config_file) as file:
+            # The FullLoader parameter handles the conversion from YAML
+            # scalar values to Python the dictionary format
+            config = yaml.load(file, Loader=yaml.FullLoader)
+        
+        self.booleanChoices = config['booleanChoices']
+        self.numericalChoices = config['numericalChoices']
+        self.stringChoices = config['stringChoices']
+        self.listChoices = config['listChoices']
+        
+        self.grid_filename = grid_filename
+        self.stringChoices['--grid']= self.grid_filename
+        
+        compas_executable_override = os.environ.get('COMPAS_EXECUTABLE_PATH')
+        print('compas_executable_override', compas_executable_override)
+        
+        if (compas_executable_override is None):
+            # Make sure you have set and exported the COMPAS_ROOT_DIR environment variable
+            git_directory = os.environ.get('COMPAS_ROOT_DIR')
+            self.compas_executable = os.path.join(git_directory, 'src/COMPAS')
+        else:
+            self.compas_executable = compas_executable_override
+
+        #TODO: add an option to read in the random seed form yaml? | HFS oct 31 2021
+        # AVG - 29/11/2021: Probably not due to the way I *think* we want to do sampling, but will double check.
+        if os.path.isfile(randomSeedFileName): 
+            self.numericalChoices['--random-seed'] = int(np.loadtxt(randomSeedFileName))
+        else:
+            self.numericalChoices['--random-seed'] = 0 # If you want a random seed, use: np.random.randint(2,2**63-1) 
+            # TODO: should random_seed be a parameter to give in __init__? | HFS oct 31 2021
+            # AVG - 29/11/2021: again, uncertain about it.
+
+        # environment variable COMPAS_LOGS_OUTPUT_DIR_PATH is used primarily for docker runs
+        # if COMPAS_LOGS_OUTPUT_DIR_PATH is set (!= None) it is used as the value for the
+        # --output-path option
+        # if COMPAS_LOGS_OUTPUT_DIR_PATH is not set (== None) the current working directory
+        # is used as the value for the --output-path option
+        compas_logs_output_override = os.environ.get('COMPAS_LOGS_OUTPUT_DIR_PATH')
+        
+        if (compas_logs_output_override is None):
+            self.stringChoices['--output-path'] = os.getcwd()
+            self.stringChoices['--output-container'] = output_directory   # names the directory to be created and in which log files are created.  Default in COMPAS is "COMPAS_Output"
+        else:
+            self.stringChoices['--output-path'] = compas_logs_output_override
+            self.stringChoices['--output-container'] = output_directory 
+        
+        
[…]
