+++ preProcessing/compasConfigDefault.yaml
@@ -1,0 +1,196 @@
+# yaml file which contains all default options from COMPAS
+
+booleanChoices:
+    --enable-warnings: False    # option to enable/disable warning messages | HFS oct 31 2021 SHOULD THAT BE AN __INIT__ OPTION?? 
+    --use-mass-loss: True
+    --mass-transfer: True
+    --detailed-output: True # WARNING! this creates a data heavy file
+    --evolve-unbound-systems: False
+    --population-data-printing: False
+    --rlof-printing: True
+    --circularise-binary-during-mass-transfer: True
+    --angular-momentum-conservation-during-circularisation: False
+    --pair-instability-supernovae: True
+    --pulsational-pair-instability: True
+    --quiet: False
+    --common-envelope-allow-main-sequence-survive: True # Allow main sequence stars to survive CE. Was previously False by default
+    --revised-energy-formalism-nandez-ivanova: False    # AVG: Need to either change to common-envelope-revised-energy-formalism-nandez-ivanova or remove.
+    --evolve-pulsars: False
+    --debug-to-file: False
+    --errors-to-file: False
+    --allow-rlof-at-birth: True # allow binaries that have one or both stars in RLOF at birth to evolve, particularly useful in the context of CHE binaries.
+    --allow-touching-at-birth: False    # record binaries that have stars touching at birth in output files
+    --switch-log: False
+    --check-photon-tiring-limit: False
+
+
+numericalChoices:
+    ### LOGISTICS 
+    --number-of-systems: 10 # number of systems per batch
+    --maximum-evolution-time: 13700.0   # Maximum physical time a system can be evolved [Myrs]
+    --maximum-number-timestep-iterations: 99999
+    --timestep-multiplier: 1    # Optional multiplier relative to default time step duration
+    --log-level: 0
+    --debug-level: 0
+    --hdf5-chunk-size: 100000
+    --hdf5-buffer-size: 1
+    #### BASIC STELLAR PROPERTIES 
+    --initial-mass: # initial mass for SSE
+    --initial-mass-min: 5.0 # Use 1.0 for LRNe, 5.0 for DCOs  [Msol] #HFS: dunno what that means - maybe it's in the manual?
+    --initial-mass-max: 150.0 # Stellar tracks extrapolated above 50 Msol (Hurley+2000) [Msol]
+    --initial-mass-power: 0.0
+    --metallicity: 0.0142   # metallicity for both SSE and BSE - Solar metallicity Asplund+2010
+    --metallicity-min: 0.0001
+    --metallicity-max: 0.03
+    --luminous-blue-variable-multiplier: 1.5
+    --overall-wind-mass-loss-multiplier: 1.0
+    --wolf-rayet-multiplier: 1.0
+    --cool-wind-mass-loss-multiplier: 1.0
+    --random-seed: # | HFS oct 31 2021 - IS DONE IN CLASS -> Add an option to read it in?
+    ###  BINARY PROPERTIES
+    --initial-mass-1:  # primary initial mass for BSE
+    --initial-mass-2:  # secondary initial mass for BSE
+    --mass-ratio: 
+    --mass-ratio-min: 0.01
+    --mass-ratio-max:  1.0
+    --minimum-secondary-mass: 0.1   # Brown dwarf limit  [Msol]
+    --eccentricity:  # eccentricity for BSE
+    --eccentricity-min: 0.0
+    --eccentricity-max: 1.0
+    --orbital-period:  # orbital period for BSE
+    --semi-major-axis:  # semi-major axis for BSE
[…]
