### 2021-11-30 TeamCOMPAS/COMPAS#715 @HeloiseS
That's cool. I asked several times where the class was used and no one mentioned strooowaffel so I wasn't aware of dependencies.

Sounds like you've got a plan!
