### 2021-11-30 7e28368
Fixed type line 920 in HG.cpp
