### 2021-11-30 TeamCOMPAS/COMPAS#698 @jeffriley
> was there anything left to do with this issue? If not, may I close it?

Maybe :-)  When I wrote H5copy I wrote it with the COMPAS_Output.h5 files in mind - I didn't provide functionality to combine COMPAS detailed output files.  I suppose there could be a use case to combine many detailed output files into a single file, but I think it would be rarely-used functionality.  The fact that H5copy does not have functionality to combine detailed output files, and that Heloise had detailed output files in the directory tree is what caused some of the error messages described above.

At a minimum I should probably document that H5copy does not copy detailed output files, and that no such files should be in the directory tree that H5copy will traverse to find files to copy.  I could also look at putting a check in the code and emitting a clearer error message.  If anybody thinks it's worthwhile, they could open an issue to add the functionality to H5copy.

EDIT: or maybe I could check for detailed output files and just ignore them.  Actually, check for any HDF5 files that are not clearly recognisable as COMPAS_Output files and ignore anything that is not a COMPAS_Output file.
