### 2021-12-01 12dc2ea
rearranged directory structure, now we have a utils/ directory containing PythonScripts, preProcessing tools, a Tutorial, and example_plots
