+++ utils/PythonScripts/CosmicIntegration/selection_effects.py
@@ -1,0 +1,317 @@
+#! /usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Created for Python 3
+
+@author: Sebastian M. Gaebel
+@email: sgaebel@star.sr.bham.ac.uk
+"""
+
+from __future__ import division, print_function
+import h5py
+import numpy as np
+import os
+import scipy.interpolate
+
+
+# Global variables to reuse objects for multiple calls
+_random_thetas = None
+_interpolator = None
+_sens = None
+
+def detection_probability(m1, m2, redshift, distance, snr_threshold,sensitivity='design'):
+    """
+    Returns the detection probability of a CBC event with given
+    masses and distance.
+
+    This function is a convenience function to generate the
+    interpolator with 'SNR_Grid_IMRPhenomPv2_FD_all_noise.hdf5'
+    and 'SimNoisePSDaLIGODesignSensitivityP1200087', redshift the
+    masses, rescale the SNR using the distance, and then call
+    'detection_probability_from_snr' to return the probability.
+
+    Parameters
+    ----------
+    m1, m2 : float > 0
+        Primary and secondary mass in source frame.
+    redshift : float >= 0
+        Redshift of the waveform through cosmology.
+    distance : float > 0
+        Luminosity Distance in Mpc to the source.
+    snr_threshold : float > 0
+        Threshold above which an event is considered to be detected.
+    sensitivity : str
+        Which detector sensitivity PSD to use. Options are 'design' and 'O1'
+
+    Returns
+    -------
+    out : float
+        Estimate of the detection probability.
+
+    Notes
+    -----
+    The interpolator is only initialized once and then stored in a
+    module level global variable to be reused.
+    """
+
+    global _interpolator
+    global _sens
+    if (_interpolator is None) or (_sens is not sensitivity):
+        _sens = sensitivity
+        _interpolator = SNRinterpolator(_sens)
+    interpolated_snr = _interpolator(m1*(1+redshift), m2*(1+redshift))
+    # SNR scales as 1/distance
+    interpolated_snr /= distance
+
+    return detection_probability_from_snr(snr_value=interpolated_snr,
+                                          snr_threshold=snr_threshold)
+
+def SNRinterpolator(sensitivity='design'):
+    """
+    Returns an Interpolator class instance for a given sensitivity
+    
+    This function is a convenience function to generate the
+    interpolator with 'SNR_Grid_IMRPhenomPv2_FD_all_noise.hdf5'
+    and 'SimNoisePSDaLIGODesignSensitivityP1200087'.
+    
+    Parameters
+    ----------
+    sensitivity : str
+    Which detector sensitivity PSD to use. Options are 'design' and 'O1'
+    
+    Returns
+    -------
+    out : Interpolator
+    Interpolator class instance
+    
+    Notes
+    -----
+    The interpolator is only initialized once and then stored in a
+    module level global variable to be reused.
+    """
+    path = os.path.dirname(os.path.abspath(__file__))
+    
+    if sensitivity == 'design':
+        hdfDatasetName = 'SimNoisePSDaLIGODesignSensitivityP1200087'
+    elif sensitivity == 'O1':
+        hdfDatasetName = 'P1500238_GW150914_H1-GDS-CALIB_STRAIN.txt'
+    elif sensitivity == 'O3':
+        hdfDatasetName = 'SimNoisePSDaLIGOMidHighSensitivityP1200087'
+
+    global _interpolator
+    global _sens
+    if (_interpolator is None) or (_sens is not sensitivity):
+        _sens = sensitivity
+        _interpolator = Interpolator(
+                                     path+'/SNR_Grid_IMRPhenomPv2_FD_all_noise.hdf5',
[…]
