+++ postProcessing/PythonScripts/CosmicIntegration/ClassMSSFR.py
@@ -1,725 +1,0 @@
-#!/usr/bin/env python3
-import numpy as np
-import scipy.special
-from   scipy.optimize    import newton
-from   astropy.cosmology import FlatLambdaCDM
-from   astropy.cosmology import WMAP9 #as cosmo
-import astropy.units as u
-
-
-
-
-class MSSFR(object):
-    """
-    This class is to calculate the metallicity specific star formation
-    rate in a specific metallicity bin.
-
-    It combines a
-    -  MZ -relation     : galaxy stellar mass - metallicity relatiion
-                          This translates a metallicity to a 
-                          a galaxy stellar mass which has the same 
-                          average metallicity.
-    -  SFR-prescription : Star formation rate prescription.
-                          Amount of solar mass that goes into forming
-                          stars.
-    -  GSMF function    : Galaxy stellar mass function
-                          A density function of the distribution of galaxy stellar masses.
-
-
-    The entire GSMF is the mass in all stars in all galaxies. Hence the fraction of the GSMF
-    is the fraction of the stellar mass in all galaxies at that time. 
-    We assume that the average cosmic SFR spreads evenly amongs all the galaxy stellar mass.
-    Meaning a galaxy with a stellar mass twice that of another galaxy, has twice
-    the SFR. Hence this method does not include local star burts (SMC/LMC for example)
-    
-
-    The overall idea. Given a metallicity bin with upper and lower bound.
-    Translate this to a galaxy stellar mass bin with upper and lower bound using
-    a MZ-relation. Then find the fraction of the GSMF density function that is taken by 
-    this glaxy stellar mass bin. We assumed that the fraction of the stellar mass is 
-    proportional to the fraction of the SFR. Hence we multiply this fraction by the SFR.
-    Note that because we take "cosmological averages" this method does not take into 
-    account galaxy specific metallicity distributions.
-
-
-    Alternatively we can also directly use a metallicity density/probability function
-    and calculate the fraction occupied by the metallicity bin in this distribution. This
-    is the logNormal distribution, which is still multiplied by the SFR and a 
-    'cosmological average'.
-
-    """
-
-    def __init__(self, verbose=False,\
-                 metallicityGrid=None,  binInLogSpace=True,\
-                 metallicityLowerLimit=1E-90, metallicityUpperLimit=1.,\
-                 logOH12sun=8.69,     solarMetallicity=0.0142, cosmo=None):
-
-        #if True we print statements
-        self.verbose               = verbose
-
-        #With regards to metallicity binning the integral
-        self.metallicityGrid       = metallicityGrid
-        self.metallicityBinEdges   = None  
-        self.binInLogSpace         = binInLogSpace 
-        #lower limit is not zero to avoid divide by zero error
-        self.metallicityLowerLimit = metallicityLowerLimit
-        self.metallicityUpperLimit = metallicityUpperLimit
-
-        print("ClassMSSFR: Remember to set self.SFRprescription() + subparts")
-        print("                            self.Zprescription()   +sub-parts")
-
-        if self.metallicityGrid is not None:
-            self.calculateMetallicityBinEdges()
-        else:
-            print("                            set self.calculateMetallicityBinEdges()")
-
-
-        #import constants default from is Asplund
-        self.logOH12sun            = logOH12sun        #fraction number density solar
-        self.solarMetallicity      = solarMetallicity  #fraction mass in metals solar
-
[…]
