+++ preProcessing/compasConfigDefault.yaml
@@ -1,94 +1,95 @@
 # yaml file which contains all default options from COMPAS
 
 booleanChoices:
-    --enable-warnings: False    # option to enable/disable warning messages | HFS oct 31 2021 SHOULD THAT BE AN __INIT__ OPTION?? 
-    --use-mass-loss: True
-    --mass-transfer: True
+    ### LOGISTICS
+    --debug-to-file: False
     --detailed-output: True # WARNING! this creates a data heavy file
+    --enable-warnings: False    # option to enable/disable warning messages
+    --errors-to-file: False    
     --evolve-unbound-systems: False
     --population-data-printing: False
+    --quiet: False    
     --rlof-printing: True
-    --circularise-binary-during-mass-transfer: True
+    --switch-log: False    
+    ### STELLAR PROPERTIES
+    --check-photon-tiring-limit: False    
+    --use-mass-loss: True
+    ### BINARY PROPERTIES
+    --allow-touching-at-birth: False    # record binaries that have stars touching at birth in output files    
+    ### MASS TRANSFER  
     --angular-momentum-conservation-during-circularisation: False
+    --allow-rlof-at-birth: True # allow binaries that have one or both stars in RLOF at birth to evolve, particularly useful in the context of CHE binaries.    
+    --circularise-binary-during-mass-transfer: True
+    --mass-transfer: True    
+    ### COMMON ENVELOPE
+    --common-envelope-allow-main-sequence-survive: True # Allow main sequence stars to survive CE. Was previously False by default
+    --revised-energy-formalism-nandez-ivanova: False
+    ### SUPERNOVAE, KICKS AND REMNANTS
     --pair-instability-supernovae: True
     --pulsational-pair-instability: True
-    --quiet: False
-    --common-envelope-allow-main-sequence-survive: True # Allow main sequence stars to survive CE. Was previously False by default
-    --revised-energy-formalism-nandez-ivanova: False    # AVG: Need to either change to common-envelope-revised-energy-formalism-nandez-ivanova or remove.
+    ### PULSAR PARAMETERS
     --evolve-pulsars: False
-    --debug-to-file: False
-    --errors-to-file: False
-    --allow-rlof-at-birth: True # allow binaries that have one or both stars in RLOF at birth to evolve, particularly useful in the context of CHE binaries.
-    --allow-touching-at-birth: False    # record binaries that have stars touching at birth in output files
-    --switch-log: False
-    --check-photon-tiring-limit: False
-
 
 numericalChoices:
-    ### LOGISTICS 
-    --number-of-systems: 10 # number of systems per batch
-    --maximum-evolution-time: 13700.0   # Maximum physical time a system can be evolved [Myrs]
-    --maximum-number-timestep-iterations: 99999
-    --timestep-multiplier: 1    # Optional multiplier relative to default time step duration
-    --log-level: 0
+    ### LOGISTICS
     --debug-level: 0
     --hdf5-chunk-size: 100000
     --hdf5-buffer-size: 1
-    #### BASIC STELLAR PROPERTIES 
+    --log-level: 0
[…]
