### 2021-12-01 8725754
Udated new scripts and deleted pythonSubmit.py
