### 2021-12-05 455c20d
Updates to ResolveEnvelopeLoss
