+++ src/HG.cpp
@@ -515,7 +515,7 @@ double HG::CalculateRadiusOnPhase(const double p_Mass, const double p_Tau, const
 double HG::CalculateRadialExtentConvectiveEnvelope() const {
 
 	BaseStar clone = *this;                         // clone this star so can manipulate without changes persisiting
-	clone.ResolveRemnantAfterEnvelopeLoss();        // update clone's attributes after envelope is lost
+	clone.ResolveEnvelopeLoss(true);                // update clone's attributes after envelope is lost
 
     return PPOW(m_Tau, 1.0 / 2.0) * (m_Radius - clone.Radius());
 }
@@ -800,7 +800,9 @@ double HG::ChooseTimestep(const double p_Time) const {
 
 
 /*
- * Resolve changes to the remnant after the star loses its envelope
+ * Modify the star after it loses its envelope
+ *
+ * Hurley et al. 2000, section 6 just before eq 76 and after Eq. 105
  *
  * Where necessary updates attributes of star (depending upon stellar type):
  *
@@ -816,33 +818,6 @@ double HG::ChooseTimestep(const double p_Time) const {
  *     - m_COCoreMass
  *     - m_Age
  *
- * Hurley et al. 2000, just after eq 105
- *
- * JR: todo: why is this different from ResolveEnvelopeLoss()?
- * JR: todo: original code: Star::radiusRemnantStarAfterLosingEnvelope() vs Star::modifyStarAfterLosingEnvelope(int stellarType, double mass)
- * JR: todo: why is stellar type changed for some types, but not others?  CheB and EAGB stars have stellar type changed, but no other types do...
- *
- *
- * STELLAR_TYPE ResolveRemnantAfterEnvelopeLoss()
- *
- * @return                                      Stellar type to which star should evolve
- */
-STELLAR_TYPE HG::ResolveRemnantAfterEnvelopeLoss() {
-
-    m_Radius = utils::Compare(m_Mass, m_MassCutoffs[static_cast<int>(MASS_CUTOFF::MHook)]) < 0
-                ? HeWD::CalculateRadiusOnPhase_Static(m_Mass)                   // reset mass/age parameters        JR: todo: why does the comment refer to mass/age?  Leftover from modifyStarAfterLosingEnvelope()?
-                : HeMS::CalculateRadiusAtZAMS_Static(m_Mass);                   // star evolves to Zero age Naked Helium Main Star and reset parameters
-
-    return m_StellarType;                                                       // no change
-}
-
-
-/*
- * Modify the star after it loses its envelope
- *
- * Hurley et al. 2000, section 6 just before eq 76
- *
- *
  * STELLAR_TYPE ResolveEnvelopeLoss()
  *
  * @return                                      Stellar Type to which star shoule evolve after losing envelope
@@ -855,13 +830,13 @@ STELLAR_TYPE HG::ResolveEnvelopeLoss(bool p_NoCheck) {
 
     if (p_NoCheck || utils::Compare(m_CoreMass, m_Mass) > 0) {                  // envelope loss
 
-        if (utils::Compare(m_Mass, massCutoffs(MHeF)) < 0) {                    // star evolves to Helium White Dwarf
+        if (utils::Compare(m_Mass0, massCutoffs(MHeF)) < 0) {                   // star evolves to Helium White Dwarf
 
             stellarType  = STELLAR_TYPE::HELIUM_WHITE_DWARF;
 
             m_Mass       = m_CoreMass;
             m_Radius     = HeWD::CalculateRadiusOnPhase_Static(m_Mass);
-            m_Age        = 0.0;                                                 // JR: see Hurley et al. 2000, discussion after eq 76
+            m_Age        = 0.0;                                                 // see Hurley et al. 2000, discussion after eq 76
         }
[…]
