### 2021-12-05 a2369d7
Updates to ResolveEnvelopeLoss
