### 2021-12-05 cb5f1a0
Updates to ResolveEnvelopeLoss
