### 2021-12-05 298779c
Updates to ResolveEnvelopeLoss
