### 2021-12-05 9e2258b
Updates to ResolveEnvelopeLoss
