### 2021-12-05 TeamCOMPAS/COMPAS#698 @ilyamandel
@jeffriley I agree that we should not be combining detailed output files, and should just document that.  Since I had #721 open, I just added a slight documentation tweak to it, namely, the H5 copy page now says:

This program copies ``COMPAS_Output.h5`` ``HDF5`` file(s) [but not ``Detailed_Ouput`` files] to a designated output ``HDF5`` file.

I am closing this issue.
