### 2021-12-05 TeamCOMPAS/COMPAS#680 @ilyamandel
@jeffriley -- see my question above; this issue ended up in a sufficiently long discussion that I can't figure out if there's actually a TODO item here. :)
