### 2021-12-06 6242cba
changed stellar_type(x) to stellar_type(x)>CE where applicable
