+++ src/constants.h
@@ -3027,7 +3027,7 @@ const ANY_PROPERTY_VECTOR BSE_COMMON_ENVELOPES_REC = {
     BINARY_PROPERTY::RADIUS_1_PRE_COMMON_ENVELOPE,
     BINARY_PROPERTY::RADIUS_1_POST_COMMON_ENVELOPE,
     BINARY_PROPERTY::STELLAR_TYPE_1_PRE_COMMON_ENVELOPE,
-    STAR_1_PROPERTY::STELLAR_TYPE,
+    BINARY_PROPERTY::STELLAR_TYPE_1_POST_COMMON_ENVELOPE,
     STAR_1_PROPERTY::LAMBDA_FIXED,
     STAR_1_PROPERTY::LAMBDA_NANJING,
     STAR_1_PROPERTY::LAMBDA_LOVERIDGE,
@@ -3044,7 +3044,7 @@ const ANY_PROPERTY_VECTOR BSE_COMMON_ENVELOPES_REC = {
     BINARY_PROPERTY::RADIUS_2_PRE_COMMON_ENVELOPE,
     BINARY_PROPERTY::RADIUS_2_POST_COMMON_ENVELOPE,
     BINARY_PROPERTY::STELLAR_TYPE_2_PRE_COMMON_ENVELOPE,
-    STAR_2_PROPERTY::STELLAR_TYPE,
+    BINARY_PROPERTY::STELLAR_TYPE_2_POST_COMMON_ENVELOPE,
     STAR_2_PROPERTY::LAMBDA_FIXED,
     STAR_2_PROPERTY::LAMBDA_NANJING,
     STAR_2_PROPERTY::LAMBDA_LOVERIDGE,
