### 2021-12-07 475a47d
changed references to 'Event' in constants.h to 'Bool' for clarity in the postProc
