### 2021-12-07 8fcdfc0
Revert "changed references to 'Event' in constants.h to 'Bool' for clarity in the postProc"
This reverts commit 475a47d432314476514fe1fbf6dbe3031da4e814.
