### 2021-12-10 910403c
Changes to online documentation for isses #695 and #680
