### 2021-12-10 2a0adb8
Changes to online documentation for isses #695 and #680
