### 2021-12-11 815a1e7
Changes to online documentation for isses #695 and #680
