+++ docs/online-docs/pages/User guide/Running COMPAS/running-via-docker-running.rst
@@ -149,3 +149,4 @@ to get the container id of interest, then type::
 .. [#f5] https://docs.docker.com/engine/reference/run/
 .. [#f6] https://docs.docker.com/engine/reference/run/#detached--d
 
+   
