### 2021-12-11 6a02abc
Changes to online documentation for isses #695 and #680
