### 2021-12-11 TeamCOMPAS/COMPAS#695 @jeffriley
Documentation changes in PR #724

Close when merged
