### 2021-12-11 TeamCOMPAS/COMPAS#680 @jeffriley
Documentation changes for docker issues in PR #724

@ilyamandel We should address the multiple pythonsubmit files - but maybe that can be address as part of the pythonsubmit refactor (PR #720)?
