### 2021-12-11 #3838 @ilyamandel
@christinahedges , @HeloiseS , @katiebreivik :

First of all, apologies for the delay in responding to some of the suggestions presented here.

Secondly, the ApJS submission of the longer accompanying COMPAS methods paper ( https://arxiv.org/abs/2109.10352 ) has now been accepted.

Thirdly, we have responded to the majority of the tasks identified here and/or added as issues on the COMPAS github page.  I believe the following are the only outstanding issues; please correct me if I am wrong.

1. Reformat pythonSubmit.py .  @avigna is working on that -- see https://github.com/TeamCOMPAS/COMPAS/pull/720 -- and just needs to update the documentation for that PR to be pulled.

2. Clean up some of the notebooks.  @reinhold-willcox is working on that -- see https://github.com/TeamCOMPAS/COMPAS/pull/718 -- and just needs to merge in some other changes for that PR to be pulled.

3. I realised that we have not yet addressed the two documentation issues identified by @katiebreivik in her review, https://github.com/openjournals/joss-reviews/issues/3838#issuecomment-978029500 .  I'll create a PR for them now.
