### 2021-12-11 a97508b
Adding info to start page and example to build page to match JOSS request
