### 2021-12-12 #3838 @arfon
> Secondly, the ApJS submission of the longer accompanying COMPAS methods paper ( https://arxiv.org/abs/2109.10352 ) has now been accepted.

FYI, I heard from AAS publishing that the DOI for that paper will be `10.3847/1538-4365/ac416c` so please add that information to the YAML header of your JOSS paper when you have a chance.
