### 2021-12-13 #3838 @ilyamandel
Thank you, @arfon -- paper.md on the JOSS branch updated with this DOI.
