### 2021-12-14 f8d490d
Update getting_started.md
