+++ docs/getting_started.md
@@ -8,8 +8,9 @@
     + [1.3 Setting up the Makefile and Compiling](#13-setting-up-the-makefile-and-compiling)
     + [1.4 Installing Python](#14-installing-python)
   * [2. Evolving your first binary](#2-evolving-your-first-binary)
-    + [2.1 Running COMPAS from a grid file](#21-running-compas-from-a-grid-file)
-    + [2.2 Examining detailed output](#22-examining-detailed-output)
+    + [2.1 Running COMPAS out-of-the-box](#21-running-compas-out-of-the-box)
+    + [2.2 Running COMPAS from a grid file](#22-running-compas-from-a-grid-file)
+    + [2.3 Examining detailed output](#23-examining-detailed-output)
   * [3. Further queries](#3-further-queries)
 
 ## 1. Installing COMPAS and Dependencies
@@ -124,15 +125,53 @@ If you do not have python3 installed, install it by following the instructions b
 
 
 ## 2. Evolving your first binary
-To start using COMPAS, you will need the python script `pythonSubmitDemo.py`, which specifies all the program options (physics assumptions, output types) and runs COMPAS in the terminal. Although the primary functionality of COMPAS is to evolve a whole population of binary stars rapidly, for now, let's focus on evolving a single stellar system and examining the detailed output.
+
+### 2.1 Running COMPAS out-of-the-box
+To start using COMPAS, and after the source code has been compiled, you need to go to `$COMPAS_ROOT_DIR/preProcessing/` and run the `runSubmit.py`, which reads the options from `compasConfigDefault.yaml` and overwrite the COMPAS defaults when specified.
+The `compasConfigDefault.yaml` specifies all the program options (physics assumptions, output types) and runs COMPAS in the terminal. 
+Your output should be similar to following.
+
+	$ python3 runSubmit.py
+	python_version = 3
+	compas_executable_override None
+	grid_filename None
+	$COMPAS_ROOT_DIR/src/COMPAS --LONG-COMMAND-LINE
+
+	COMPAS v02.25.10
+	Compact Object Mergers: Population Astrophysics and Statistics
+	by Team COMPAS (http://compas.science/index.html)
+	A binary star simulator
+
+	Start generating binaries at DATE
+
+	0: Unbound binary: (Main_Sequence_>_0.7 -> Neutron_Star) + (Main_Sequence_>_0.7 -> Main_Sequence_>_0.7)
+
+	Generated 1 of 1 binaries requested
+
+	Simulation completed
+
+	End generating binaries at DATE
+
+	Clock time = 0.264009 CPU seconds
+	Wall time  = 0000:00:00 (hhhh:mm:ss)
+
+If that is the case, you have succesfully installed COMPAS and ran your fist binary.
+
+As of v02.25.10, the `runSubmit.py` and `compasConfigDefault.yaml` combo are the preferred way to set up runs using python, instead of the prevously used `pythonSubmit.py` (and their adaptations).
+We will eventually update the documentation to reflect this, but for now the references to `pythonSubmit.py` or `pythonSubmitDemo.py` refer to the combined options and commands python script that has been tested and used in releases previous than v02.25.10, including those of the methods paper.
+The `pythonSubmit.py`, `pythonSubmitDemo.py`, and their functionality, will eventually be fully replaced by the `runSubmit.py` and `compasConfigDefault.yaml` combo. 
+
+### 2.2 Running COMPAS from a grid file
+
+We will now try to run a binary using a grid file. For this, you will need the python script `pythonSubmitDemo.py`, which specifies all the program options (physics assumptions, output types) and runs COMPAS in the terminal. For now, let's focus on evolving a single stellar system and examining the detailed output.
 
 To start, change to the `examples/methods_paper_plots/detailed_evolution/` directory:
 
[…]
