### 2021-12-14 630126c
Changed the default number of binaries to 1
