### 2021-12-14 991ed16
Implement radial change check for BSE mode
