+++ src/BaseBinaryStar.cpp
@@ -2246,12 +2246,84 @@ void BaseBinaryStar::EvolveOneTimestepPreamble(const double p_Dt) {
  *
  * @param   [IN]    p_Dt                        The suggested timestep to evolve
  */
-void BaseBinaryStar::EvolveOneTimestep(const double p_Dt) {
+EVOLUTION_STATUS BaseBinaryStar::EvolveOneTimestep(const double p_Dt) {
 
-    EvolveOneTimestepPreamble(p_Dt);
+    EVOLUTION_STATUS retVal = EVOLUTION_STATUS::CONTINUE;                                               // return status
 
-    m_Star1->AgeOneTimestep(p_Dt, true);    // Age the primary one timestep and switch to the new stellar type if necessary
-    m_Star2->AgeOneTimestep(p_Dt, true);    // Age the secondary one timestep and switch to the new stellar type if necessary
+    double dt = p_Dt;                                                                                   // timestep
+
+    if (OPTIONS->CheckRadialChange()) {                                                                 // check radial change?
+                                                                                                        // yes 
+        STELLAR_TYPE stellarType1;                                                                      // stellar type to which primary should switch
+        STELLAR_TYPE stellarType2;                                                                      // stellar type to which secondary should switch
+        
+        bool takeTimestep = false;                                                                      // flag to indicate whether timestep should be taken
+        int  retryCount   = 0;                                                                          // timestep/radial change retry count
+        while (!takeTimestep && retVal == EVOLUTION_STATUS::CONTINUE) {                                 // do this until a suitable timestep is found (or the maximum retry count is reached or an error occurs)
+        
+            m_Star1->SaveState();                                                                       // save the state of the primary - in case we want to revert
+            m_Star2->SaveState();                                                                       // save the state of the secondary - in case we want to revert
+
+            // calculate the minimum timestep - maximum of minimum dynamical timescale for each constituent star and the absolute minimum timestep
+            double minTimestep = std::max(std::min(m_Star1->CalculateDynamicalTimescale(), m_Star2->CalculateDynamicalTimescale()), ABSOLUTE_MINIMUM_TIMESTEP);
+
+            stellarType1 = m_Star1->AgeOneTimestep(dt, false);                                          // Age the primary one time step - modify stellar attributes as appropriate, but do not switch stellar type
+            stellarType2 = m_Star2->AgeOneTimestep(dt, false);                                          // Age the secondary one time step - modify stellar attributes as appropriate, but do not switch stellar type
+
+            // determine if the timestep should be taken
+            // don't take the timestep if we stepped too far
+
+            takeTimestep = true;                                                                        // flag to determine if the timestep should be taken
+            if ((utils::Compare(m_Star1->CalculateRadialChange(), MAXIMUM_RADIAL_CHANGE) >= 0) || 
+                (utils::Compare(m_Star2->CalculateRadialChange(), MAXIMUM_RADIAL_CHANGE) >= 0)) {       // too much change?
+
+                if (utils::Compare(dt, minTimestep) <= 0) {                                             // yes - already at or below minimum timestep?
+                    takeTimestep = true;                                                                // yes - just take the last timestep
+                    SHOW_WARN(ERROR::TIMESTEP_BELOW_MINIMUM);                                           // announce the problem if required and plough on regardless...
+                }
+                else {                                                                                  // not at or below dynamical - reduce timestep and try again
+                    retryCount++;                                                                       // increment retry count
+                    if (retryCount > MAX_TIMESTEP_RETRIES) {                                            // too many retries?
+                        takeTimestep = true;                                                            // yes - take the last timestep anyway
+                        SHOW_WARN(ERROR::TIMESTEP_BELOW_MINIMUM);                                       // announce the problem if required and plough on regardless...
+                    }
[…]
