### 2021-12-15 #3838 @ilyamandel
Dear @HeloiseS , @katiebreivik , @christinahedges:

We've now pulled the remaining key changes addressing the issues that you brought  up (thanks to @avigna , @jeffriley , @reinhold-willcox et al.).

Of course, things are still far from perfect -- there are nearly 50 outstanding issues listed at https://github.com/TeamCOMPAS/COMPAS/issues , some of them related the topics raised in the review, including post-processing, documentation, etc.  It's a living code, so it will never reach a state where we are completely happy with it.  But the reviews have been incredibly helpful in making it better -- **thank you**!

Please let us know whether you think we can proceed from here on the understanding that we'll continue to improve the code along the lines highlighted in the review, or whether there are still significant issues you'd like to see addressed first.

P.S. I will be offline from 18 December through 10 January, so I apologise in advance for not responding to your messages during this time.  Best wishes for the holiday season!
