### 2021-12-15 #3838 @christinahedges
@HeloiseS and @katiebreivik, it looks to me that the authors have completed all of the major comments from the review. @HeloiseS it seems like the team are addressing Python submit file, which I believe @ilyamandel is close to merging.

Are the reviewers happy to sign off on this submission? If so, please say in the comments, and I will then recommend the submission for acceptance.
