### 2021-12-15 3d32ea3
Added check for mass transfer in previous timestep when deciding whether to check radial change
