### 2021-12-16 869ad48
updated Makefile to point at the correct notebooks directory (#730)
