### 2021-12-16 b19c7ca
first round of changes, plus added temp ppowtest file
