### 2021-12-16 86f24f9
first round of changes, plus added temp ppowtest file
