+++ src/CHeB.cpp
@@ -94,7 +94,7 @@ double CHeB::CalculateLambdaDewi() const {
 
     double lambda3 = std::min(-0.9, 0.58 + (0.75 * log10(m_Mass))) - (0.08 * log10(m_Luminosity));                          // (A.4) Claeys+2014
     double lambda1 = std::min(lambda3, std::min(0.8, 1.25 - (0.15 * log10(m_Luminosity))));                                 // (A.5) Top, Claeys+2014
-	double lambda2 = 0.42 * PPOW(m_RZAMS / m_Radius, 0.4);                                                                  // (A.2) Claeys+2014
+	double lambda2 = 0.42 * PPOW(m_RZAMS / m_Radius, 0.4);                                                                  // (A.2) Claeys+2014          // RTW - Consider replacing this with a 2/5 root function (somehow) to avoid NaNs if the base is negative
 	double envMass = utils::Compare(m_CoreMass, 0.0) > 0 && utils::Compare(m_Mass, m_CoreMass) > 0 ? m_Mass - m_CoreMass : 0.0;
 
     double lambdaCE;
@@ -778,9 +778,9 @@ double CHeB::CalculateRadiusRho(const double p_Mass, const double p_Tau) const {
 
     double ty_tx = ty - tx;
 
-    double one   = PPOW(log((Ry / Rmin)), 1.0 / 3.0);
+    double one   = std::cbrt(log((Ry / Rmin)));
     double two   = (p_Tau - tx) / ty_tx;
-    double three = PPOW(log((Rx / Rmin)), 1.0 / 3.0);
+    double three = std::cbrt(log((Rx / Rmin)));
     double four  = (ty - p_Tau) / ty_tx;
 
     return (one * two) - (three * four);
