### 2021-12-16 3c84ffb
fixed easy exponents
