+++ src/WhiteDwarfs.cpp
@@ -31,5 +31,7 @@ double WhiteDwarfs::CalculateLuminosityOnPhase_Static(const double p_Mass, const
  * @return                                      Radius of a White Dwarf in Rsol (since WD is ~ Earth sized, expect answer around 0.009)
  */
 double WhiteDwarfs::CalculateRadiusOnPhase_Static(const double p_Mass) {
-    return std::max(NEUTRON_STAR_RADIUS, 0.0115 * sqrt(PPOW(MCH / p_Mass, 2.0 / 3.0) - PPOW(p_Mass / MCH, 2.0 / 3.0)));
+    double MCH_Mass_one_third  = std::cbrt(MCH / p_Mass); 
+    double MCH_Mass_two_thirds = MCH_Mass_one_third* MCH_Mass_one_third
+    return std::max(NEUTRON_STAR_RADIUS, 0.0115 * sqrt((MCH_Mass_two_thirds - 1.0/MCH_Mass_two_thirds )));
 }
