### 2021-12-16 2767b9c
fixed easy exponents
