+++ src/utils.cpp
@@ -172,7 +172,7 @@ namespace utils {
         double a_cubed_SI_top    = G * ((p_Mass1 * MSOL_TO_KG) + (p_Mass2 * MSOL_TO_KG)) * p_Period * p_Period * SECONDS_IN_DAY * SECONDS_IN_DAY;
         double a_cubed_SI_bottom = 4.0 * M_PI * M_PI;
         double a_cubed_SI        = a_cubed_SI_top / a_cubed_SI_bottom;
-        double a_SI              = PPOW(a_cubed_SI, 1.0 / 3.0);
+        double a_SI              = std::cbrt(a_cubed_SI); 
 
         return a_SI / AU;
     }
