+++ src/MainSequence.cpp
@@ -348,7 +348,7 @@ double MainSequence::CalculateDeltaR(const double p_Mass) const {
     double deltaR;
 
     if (utils::Compare(p_Mass, massCutoffs(MHook)) <= 0) deltaR = 0.0;   // this really is supposed to be 0
-    else if (utils::Compare(p_Mass, a[42])         <= 0) deltaR = a[43] * PPOW(((p_Mass - massCutoffs(MHook)) / (a[42] - massCutoffs(MHook))), 0.5);
+    else if (utils::Compare(p_Mass, a[42])         <= 0) deltaR = a[43] * sqrt((p_Mass - massCutoffs(MHook)) / (a[42] - massCutoffs(MHook)));
     else if (utils::Compare(p_Mass, 2.0)            < 0) deltaR = a[43] + ((m_RConstants[static_cast<int>(R_CONSTANTS::B_DELTA_R)] - a[43]) * PPOW(((p_Mass - a[42]) / (2.0 - a[42])), a[44]));
     else {
         // pow() is slow - use multiplication (sqrt() is faster than pow())
@@ -483,7 +483,7 @@ double MainSequence::CalculateRadiusOnPhase(const double p_Mass, const double p_
  * @return                                      Radial extent of the star's convective envelope in Rsol
  */
 double MainSequence::CalculateRadialExtentConvectiveEnvelope() const {
-    return utils::Compare(m_Mass, 0.35) <= 0 ? m_Radius * PPOW(1.0 - m_Tau, 1.0 / 4.0) : 0.0;
+    return utils::Compare(m_Mass, 0.35) <= 0 ? m_Radius * sqrt(sqrt(1.0 - m_Tau)) : 0.0;
 }
 
 
