### 2021-12-16 72b00c6
changed sqrt to std::sqrt everywhere
