### 2021-12-16 482cfc1
changed sqrt to std::sqrt everywhere
