+++ src/HeMS.cpp
@@ -56,12 +56,12 @@ void HeMS::CalculateTimescales(const double p_Mass, DBL_VECTOR &p_Timescales) {
 double HeMS::CalculateLuminosityAtZAMS_Static(const double p_Mass) {
 
     // pow() is slow - use multiplication (sqrt() is much faster than pow())
-    double m_0_5   = sqrt(p_Mass);
+    double m_0_5   = std::sqrt(p_Mass);
     double m_3     = p_Mass * p_Mass * p_Mass;
     double m_6     = m_3 * m_3;
     double m_7_5   = m_6 * p_Mass * m_0_5;
     double m_9     = m_6 * m_3;
-    double m_10_25 = m_9 * p_Mass * sqrt(m_0_5);
+    double m_10_25 = m_9 * p_Mass * std::sqrt(m_0_5);
 
     return (15262.0 * m_10_25) / (m_9 + (29.54 * m_7_5) + (31.18 * m_6) + 0.0469);
 }
@@ -303,7 +303,7 @@ double HeMS::CalculateLifetimeOnPhase_Static(const double p_Mass) {
     // pow() is slow - use multiplication (sqrt() is much faster than pow())
     double m_4   = p_Mass * p_Mass * p_Mass * p_Mass;
     double m_6   = m_4 * p_Mass * p_Mass;
-    double m_6_5 = m_6 * sqrt(p_Mass);
+    double m_6_5 = m_6 * std::sqrt(p_Mass);
 
     return (0.4129 + (18.81 * m_4) + (1.853 * m_6)) / m_6_5;
 }
