+++ src/BaseStar.cpp
@@ -1212,7 +1212,7 @@ double BaseStar::CalculateLuminosityAtZAMS(const double p_MZAMS) {
 
     // pow() is slow - use multiplication where it makes sense
     // sqrt() is much faster than pow()
-    double m_0_5 = sqrt(p_MZAMS);
+    double m_0_5 = std::sqrt(p_MZAMS);
     double m_2   = p_MZAMS * p_MZAMS;
     double m_3   = m_2 * p_MZAMS;
     double m_5   = m_3 * m_2;
@@ -1295,8 +1295,8 @@ double BaseStar::CalculateRadiusAtZAMS(const double p_MZAMS) const {
 #define coeff(x) m_RCoefficients[static_cast<int>(R_Coeff::x)]  // for convenience and readability - undefined at end of function
 
     // pow() is slow - use multiplication where it makes sense
-    // sqrt() is much faster than pow()
-    double m_0_5  = sqrt(p_MZAMS);
+    // std::sqrt() is much faster than pow()
+    double m_0_5  = std::sqrt(p_MZAMS);
     double m_2    = p_MZAMS * p_MZAMS;
     double m_2_5  = m_2 * m_0_5;
     double m_6    = m_2 * m_2 * m_2;
@@ -1458,7 +1458,7 @@ double BaseStar::CalculateMassLossRateNieuwenhuijzenDeJager() const {
     double rate = 0.0;
     if (utils::Compare(m_Luminosity, NJ_MINIMUM_LUMINOSITY) > 0) {      // check for minimum luminosity
         double smoothTaper = min(1.0, (m_Luminosity - 4000.0) / 500.0); // Smooth taper between no mass loss and mass loss
-        rate = sqrt((m_Metallicity / ZSOL)) * smoothTaper * 9.6E-15 * PPOW(m_Radius, 0.81) * PPOW(m_Luminosity, 1.24) * PPOW(m_Mass, 0.16);
+        rate = std::sqrt((m_Metallicity / ZSOL)) * smoothTaper * 9.6E-15 * PPOW(m_Radius, 0.81) * PPOW(m_Luminosity, 1.24) * PPOW(m_Mass, 0.16);
     } else {
         rate = 0.0;
     }
@@ -1477,7 +1477,7 @@ double BaseStar::CalculateMassLossRateNieuwenhuijzenDeJager() const {
  */
 double BaseStar::CalculateMassLossRateLBV(const LBV_PRESCRIPTION p_LBV_prescription) {
     double rate = 0.0;
-    double HD_limit_factor = m_Radius * sqrt(m_Luminosity) * 1.0E-5;                                                            // calculate factor by which you are above the HD limit
+    double HD_limit_factor = m_Radius * std::sqrt(m_Luminosity) * 1.0E-5;                                                            // calculate factor by which you are above the HD limit
     if ((utils::Compare(m_Luminosity, LBV_LUMINOSITY_LIMIT_STARTRACK) > 0) && (utils::Compare(HD_limit_factor, 1.0) > 0)) {     // check if luminous blue variable
 		m_LBVphaseFlag = true;                                                                                                  // mark the star as LBV
         m_DominantMassLossRate = MASS_LOSS_TYPE::LUMINOUS_BLUE_VARIABLE;
@@ -1959,7 +1959,7 @@ double BaseStar::CalculateThermalMassAcceptanceRate(const double p_Radius) const
  * @return                                      Effective temperature of the star (Tsol)
  */
[…]
