### 2021-12-17 4eb2e6b
added options for initial stellar type, still need to add sanity check
