+++ src/Options.cpp
@@ -191,6 +191,10 @@ void Options::OptionValues::Initialise() {
     m_InitialMassFunctionMax                                        = 150.0;
     m_InitialMassFunctionPower                                      = 0.0;
 
+    // Initial stellar type options
+    m_InitialStellarType                                            = 1;                        // Default Main Sequence
+    m_InitialStellarType1                                           = 1;                        // Default Main Sequence
+    m_InitialStellarType2                                           = 1;                        // Default Main Sequence
 
     // Initial mass ratio
     m_MassRatio                                                     = 1.0;
@@ -200,7 +204,6 @@ void Options::OptionValues::Initialise() {
     m_MassRatioDistributionMax                                      = 1.0;
 
     m_MinimumMassSecondary                                          = 0.1;
-
 
     // Initial orbit options
     m_SemiMajorAxis                                                 = 0.1;
@@ -780,6 +783,23 @@ bool Options::AddOptions(OptionValues *p_Options, po::options_description *p_Opt
             po::value<int>(&p_Options->m_DebugLevel)->default_value(p_Options->m_DebugLevel),                                                                                                     
             ("Determines which print statements are displayed for debugging (default = " + std::to_string(p_Options->m_DebugLevel) + ")").c_str()
         )
+
+        (
+            "initial-stellar-type",                                            
+            po::value<int>(&p_Options->m_InitialStellarType)->default_value(p_Options->m_InitialStellarType),                                                                          
+            ("Initial stellar type (Hurley type) for the star (SSE) (default = " + std::to_string(p_Options->m_InitialStellarType) + ")").c_str()
+        )
+        (
+            "initial-stellar-type-1",                                            
+            po::value<int>(&p_Options->m_InitialStellarType1)->default_value(p_Options->m_InitialStellarType1),                                                                          
+            ("Initial stellar type (Hurley type) for the primary star (BSE) (default = " + std::to_string(p_Options->m_InitialStellarType1) + ")").c_str()
+        )
+        (
+            "initial-stellar-type-2",                                            
+            po::value<int>(&p_Options->m_InitialStellarType2)->default_value(p_Options->m_InitialStellarType2),
+            ("Initial stellar type (Hurley type) for the secondary star (BSE) (default = " + std::to_string(p_Options->m_InitialStellarType2) + ")").c_str()
+        )
+
         (
             "grid-start-line",                                                 
             po::value<std::streamsize>(&p_Options->m_GridStartLine)->default_value(p_Options->m_GridStartLine),                                                                                                     
@@ -1993,6 +2013,11 @@ std::string Options::OptionValues::CheckAndSetOptions() {
         COMPLAIN_IF(m_InitialMassFunctionMin < MINIMUM_INITIAL_MASS, "Minimum initial mass (--initial-mass-min) must be >= " + std::to_string(MINIMUM_INITIAL_MASS) + " Msol");
         COMPLAIN_IF(m_InitialMassFunctionMax > MAXIMUM_INITIAL_MASS, "Maximum initial mass (--initial-mass-max) must be <= " + std::to_string(MAXIMUM_INITIAL_MASS) + " Msol");
         COMPLAIN_IF(m_InitialMassFunctionMax <= m_InitialMassFunctionMin, "Maximum initial mass (--initial-mass-max) must be > Minimum initial mass (--initial-mass-min)");
+
+        // RTW TODO
[…]
