### 2021-12-17 738c6cc
added parameters to handle p_InitialStellarType as input to a star
