### 2021-12-17 971298a
added stellar type into assignment for single stars
