### 2021-12-17 41bfece
added stellar type assignment for binary star components
