+++ src/BaseBinaryStar.cpp
@@ -106,6 +106,14 @@ BaseBinaryStar::BaseBinaryStar(const unsigned long int p_Seed, const long int p_
 
             mass2 = mass1 * q;                                                                                                          // calculate mass2 using mass ratio                                                                     
         }
+    
+        STELLAR_TYPE initialStellarType1 = OPTIONS->OptionSpecified("initial-stellar-type-1") == 1                                      // user specified primary stellar type?
+                        ? OPTIONS->InitialStellarType1()                                                                                // yes, use it
+                        : STELLAR_TYPE::MS_GT_07;                                                                                       // no, set it to default value
+
+        STELLAR_TYPE initialStellarType2 = OPTIONS->OptionSpecified("initial-stellar-type-2") == 1                                      // user specified secondary stellar type?
+                        ? OPTIONS->InitialStellarType2()                                                                                // yes, use it
+                        : STELLAR_TYPE::MS_GT_07;                                                                                       // no, set it to default value
 
         double metallicity = OPTIONS->OptionSpecified("metallicity") == 1                                                               // user specified metallicity?
                                 ? OPTIONS->Metallicity()                                                                                // yes, use it
@@ -151,12 +159,12 @@ BaseBinaryStar::BaseBinaryStar(const unsigned long int p_Seed, const long int p_
         // binary star contains two instances of star to hold masses, radii and luminosities.
         // star 1 initially more massive
         m_Star1 = OPTIONS->OptionSpecified("rotational-frequency-1") == 1                                                               // user specified primary rotational frequency?
-                    ? new BinaryConstituentStar(m_RandomSeed, mass1, metallicity, kickParameters1, OPTIONS->RotationalFrequency1() * SECONDS_IN_YEAR) // yes - use it (convert from Hz to cycles per year - see BaseStar::CalculateZAMSAngularFrequency())
-                    : new BinaryConstituentStar(m_RandomSeed, mass1, metallicity, kickParameters1);                                     // no - let it be calculated
+                    ? new BinaryConstituentStar(m_RandomSeed, mass1, initialStellarType1, metallicity, kickParameters1, OPTIONS->RotationalFrequency1() * SECONDS_IN_YEAR) // yes - use it (convert from Hz to cycles per year - see BaseStar::CalculateZAMSAngularFrequency())
+                    : new BinaryConstituentStar(m_RandomSeed, mass1, initialStellarType1, metallicity, kickParameters1);                                     // no - let it be calculated
 
         m_Star2 = OPTIONS->OptionSpecified("rotational-frequency-2") == 1                                                               // user specified secondary rotational frequency?
-                    ? new BinaryConstituentStar(m_RandomSeed, mass2, metallicity, kickParameters2, OPTIONS->RotationalFrequency2() * SECONDS_IN_YEAR) // yes - use it (convert from Hz to cycles per year - see BaseStar::CalculateZAMSAngularFrequency())
-                    : new BinaryConstituentStar(m_RandomSeed, mass2, metallicity, kickParameters2);                                     // no - let it be calculated
[…]
