### 2021-12-17 9186d0f
Cleaned up initial stellar type call, to take integer input but convert it to stellar type for getters
