+++ src/Options.cpp
@@ -192,9 +192,9 @@ void Options::OptionValues::Initialise() {
     m_InitialMassFunctionPower                                      = 0.0;
 
     // Initial stellar type options
-    m_InitialStellarType                                            = 1;                        // Default Main Sequence
-    m_InitialStellarType1                                           = 1;                        // Default Main Sequence
-    m_InitialStellarType2                                           = 1;                        // Default Main Sequence
+    m_InitialStellarTypeNum                                         = 1;                         // Default Main Sequence
+    m_InitialStellarTypeNum1                                        = 1;                         // Default Main Sequence
+    m_InitialStellarTypeNum2                                        = 1;                         // Default Main Sequence
 
     // Initial mass ratio
     m_MassRatio                                                     = 1.0;
@@ -786,18 +786,18 @@ bool Options::AddOptions(OptionValues *p_Options, po::options_description *p_Opt
 
         (
             "initial-stellar-type",                                            
-            po::value<int>(&p_Options->m_InitialStellarType)->default_value(p_Options->m_InitialStellarType),                                                                          
-            ("Initial stellar type (Hurley type) for the star (SSE) (default = " + std::to_string(p_Options->m_InitialStellarType) + ")").c_str()
+            po::value<int>(&p_Options->m_InitialStellarTypeNum)->default_value(p_Options->m_InitialStellarTypeNum),                                                                          
+            ("Initial stellar type (Hurley type) for the star (SSE) (default = " + std::to_string(p_Options->m_InitialStellarTypeNum) + ")").c_str()
         )
         (
             "initial-stellar-type-1",                                            
-            po::value<int>(&p_Options->m_InitialStellarType1)->default_value(p_Options->m_InitialStellarType1),                                                                          
-            ("Initial stellar type (Hurley type) for the primary star (BSE) (default = " + std::to_string(p_Options->m_InitialStellarType1) + ")").c_str()
+            po::value<int>(&p_Options->m_InitialStellarTypeNum1)->default_value(p_Options->m_InitialStellarTypeNum1),                                                                          
+            ("Initial stellar type (Hurley type) for the primary star (BSE) (default = " + std::to_string(p_Options->m_InitialStellarTypeNum1) + ")").c_str()
         )
         (
             "initial-stellar-type-2",                                            
-            po::value<int>(&p_Options->m_InitialStellarType2)->default_value(p_Options->m_InitialStellarType2),
-            ("Initial stellar type (Hurley type) for the secondary star (BSE) (default = " + std::to_string(p_Options->m_InitialStellarType2) + ")").c_str()
+            po::value<int>(&p_Options->m_InitialStellarTypeNum2)->default_value(p_Options->m_InitialStellarTypeNum2),
+            ("Initial stellar type (Hurley type) for the secondary star (BSE) (default = " + std::to_string(p_Options->m_InitialStellarTypeNum2) + ")").c_str()
         )
 
         (
@@ -2015,9 +2015,9 @@ std::string Options::OptionValues::CheckAndSetOptions() {
[…]
