### 2021-12-17 f5b4e3a
Cleaned up function calls and PROGRAM_OPTIONS settings for custom initial stellar type
