### 2021-12-17 cc4d0aa
code now compiles, and the stars get set correctly (according to stdout), but they are not internally set correctly - the timestep calculations do not use the stellar types, they use the default (min) in BaseStar
