+++ src/Star.cpp
@@ -41,6 +41,8 @@ Star::Star(const unsigned long int p_RandomSeed,
     //      if star2 non MS, set it
     //      else switch if low mass
 
+    std::cout << "Stellar type is : " << static_cast<int>(p_InitialStellarType) << std::endl;
+
     if (utils::IsOneOf(p_InitialStellarType, { STELLAR_TYPE::MS_GT_07 })) {                                             // Configure CHE or LMMS if stellar type not explicitly set
         if (OPTIONS->CHEMode() != CHE_MODE::NONE && utils::Compare(m_Star->Omega(), m_Star->OmegaCHE()) >= 0) {         // CHE?
             (void)SwitchTo(STELLAR_TYPE::CHEMICALLY_HOMOGENEOUS, true);                                                 // yes
@@ -54,7 +56,7 @@ Star::Star(const unsigned long int p_RandomSeed,
         }
     }
     else {
-        (void)SwitchTo(p_InitialStellarType, true);                                                               // MS > 0.7 Msol
+        (void)SwitchTo(p_InitialStellarType, true);                                                                     // Switch to user supplied type
     }
 
     m_SaveStar = nullptr;
