### 2021-12-17 411abda
added check in Star::SwitchTo to switch also if setting the initial StellarType, now the code seems to run properly
