+++ src/Options.cpp
@@ -191,10 +191,13 @@ void Options::OptionValues::Initialise() {
     m_InitialMassFunctionMax                                        = 150.0;
     m_InitialMassFunctionPower                                      = 0.0;
 
-    // Initial stellar type options
-    m_InitialStellarTypeNum                                         = 1;                         // Default Main Sequence
-    m_InitialStellarTypeNum1                                        = 1;                         // Default Main Sequence
-    m_InitialStellarTypeNum2                                        = 1;                         // Default Main Sequence
+    // Supplied initial stellar type 
+    m_SuppliedStellarType.type                                      = INITIAL_STELLAR_TYPE::STAR;                            // Default Main Sequence
+    m_SuppliedStellarType.typeString                                = INITIAL_STELLAR_TYPE_LABEL.at(m_InitialStellarType);   // Default Main Sequence
+    m_SuppliedStellarType1.type                                     = INITIAL_STELLAR_TYPE::STAR;                            // Default Main Sequence
+    m_SuppliedStellarType1.typeString                               = INITIAL_STELLAR_TYPE_LABEL.at(m_InitialStellarType);   // Default Main Sequence
+    m_SuppliedStellarType2.type                                     = INITIAL_STELLAR_TYPE::STAR;                            // Default Main Sequence
+    m_SuppliedStellarType2.typeString                               = INITIAL_STELLAR_TYPE_LABEL.at(m_InitialStellarType);   // Default Main Sequence
 
     // Initial mass ratio
     m_MassRatio                                                     = 1.0;
@@ -786,18 +789,18 @@ bool Options::AddOptions(OptionValues *p_Options, po::options_description *p_Opt
 
         (
             "initial-stellar-type",                                            
-            po::value<int>(&p_Options->m_InitialStellarTypeNum)->default_value(p_Options->m_InitialStellarTypeNum),                                                                          
-            ("Initial stellar type (Hurley type) for the star (SSE) (default = " + std::to_string(p_Options->m_InitialStellarTypeNum) + ")").c_str()
+            po::value<int>(&p_Options->m_SuppliedStellarType)->default_value(p_Options->m_SuppliedStellarType),                                                                          
+            ("Initial stellar type (Hurley type) for the star (SSE) (default = " + std::to_string(p_Options->m_SuppliedStellarType) + ")").c_str()
         )
         (
             "initial-stellar-type-1",                                            
-            po::value<int>(&p_Options->m_InitialStellarTypeNum1)->default_value(p_Options->m_InitialStellarTypeNum1),                                                                          
-            ("Initial stellar type (Hurley type) for the primary star (BSE) (default = " + std::to_string(p_Options->m_InitialStellarTypeNum1) + ")").c_str()
+            po::value<int>(&p_Options->m_SuppliedStellarType1)->default_value(p_Options->m_SuppliedStellarType1),                                                                          
+            ("Initial stellar type (Hurley type) for the primary star (BSE) (default = " + std::to_string(p_Options->m_SuppliedStellarType1) + ")").c_str()
         )
         (
[…]
