### 2021-12-20 d9bfb70
fixed references to hurley stellar type numbers in initial stellar type option. Now users submit a string referring to the stellar type
