+++ src/BaseBinaryStar.cpp
@@ -107,13 +107,9 @@ BaseBinaryStar::BaseBinaryStar(const unsigned long int p_Seed, const long int p_
             mass2 = mass1 * q;                                                                                                          // calculate mass2 using mass ratio                                                                     
         }
     
-        STELLAR_TYPE initialStellarType1 = OPTIONS->OptionSpecified("initial-stellar-type-1") == 1                                      // user specified primary stellar type?
-                        ? OPTIONS->InitialStellarType1()                                                                                // yes, use it
-                        : STELLAR_TYPE::MS_GT_07;                                                                                       // no, set it to default value
-
-        STELLAR_TYPE initialStellarType2 = OPTIONS->OptionSpecified("initial-stellar-type-2") == 1                                      // user specified secondary stellar type?
-                        ? OPTIONS->InitialStellarType2()                                                                                // yes, use it
-                        : STELLAR_TYPE::MS_GT_07;                                                                                       // no, set it to default value
+        // RTW
+        STELLAR_TYPE initialStellarType1 = OPTIONS->InitialStellarType1();                                                              // Get primary stellar type, possibly user specified
+        STELLAR_TYPE initialStellarType2 = OPTIONS->InitialStellarType2();                                                              // Get secondary stellar type, possibly user specified
 
         double metallicity = OPTIONS->OptionSpecified("metallicity") == 1                                                               // user specified metallicity?
                                 ? OPTIONS->Metallicity()                                                                                // yes, use it
@@ -296,6 +292,7 @@ void BaseBinaryStar::SetRemainingValues() {
         // newly-assigned rotational frequencies
         // but only if the stellar type was not otherwise explicitly set
 
+        // RTW
         // star 1
         if (m_Star1->IsOneOf({ STELLAR_TYPE::MS_LTE_07, STELLAR_TYPE::MS_GT_07, STELLAR_TYPE::CHEMICALLY_HOMOGENEOUS })) {
             if (utils::Compare(m_Star1->Omega(), m_Star1->OmegaCHE()) >= 0) {                                                                               // star 1 CH?
@@ -2387,7 +2384,6 @@ EVOLUTION_STATUS BaseBinaryStar::Evolve() {
             if (evolutionStatus == EVOLUTION_STATUS::CONTINUE) {                                                                            // continue evolution?
 
                 dt = std::min(m_Star1->CalculateTimestep(), m_Star2->CalculateTimestep()) * OPTIONS->TimestepMultiplier();                  // new timestep
-
                 if ((m_Star1->IsOneOf({ STELLAR_TYPE::MASSLESS_REMNANT }) || m_Star2->IsOneOf({ STELLAR_TYPE::MASSLESS_REMNANT })) || dt < NUCLEAR_MINIMUM_TIMESTEP) {
                     dt = NUCLEAR_MINIMUM_TIMESTEP;                                                                                          // but not less than minimum
 		}
