+++ src/Options.cpp
@@ -191,13 +191,14 @@ void Options::OptionValues::Initialise() {
     m_InitialMassFunctionMax                                        = 150.0;
     m_InitialMassFunctionPower                                      = 0.0;
 
-    // Supplied initial stellar type 
-    m_SuppliedStellarType.type                                      = INITIAL_STELLAR_TYPE::STAR;                            // Default Main Sequence
-    m_SuppliedStellarType.typeString                                = INITIAL_STELLAR_TYPE_LABEL.at(m_InitialStellarType);   // Default Main Sequence
-    m_SuppliedStellarType1.type                                     = INITIAL_STELLAR_TYPE::STAR;                            // Default Main Sequence
-    m_SuppliedStellarType1.typeString                               = INITIAL_STELLAR_TYPE_LABEL.at(m_InitialStellarType);   // Default Main Sequence
-    m_SuppliedStellarType2.type                                     = INITIAL_STELLAR_TYPE::STAR;                            // Default Main Sequence
-    m_SuppliedStellarType2.typeString                               = INITIAL_STELLAR_TYPE_LABEL.at(m_InitialStellarType);   // Default Main Sequence
+    // RTW
+    // Initial initial stellar type 
+    m_InitialStellarType.type                                      = STELLAR_TYPE::STAR;                                 // Default reverts to Main Sequence later 
+    m_InitialStellarType.typeString                                = INITIAL_STELLAR_TYPE_LABEL.at(m_InitialStellarType.type);   // Default reverts to Main Sequence later
+    m_InitialStellarType1.type                                     = STELLAR_TYPE::STAR;                                 // Default reverts to Main Sequence later
+    m_InitialStellarType1.typeString                               = INITIAL_STELLAR_TYPE_LABEL.at(m_InitialStellarType.type);   // Default reverts to Main Sequence later
+    m_InitialStellarType2.type                                     = STELLAR_TYPE::STAR;                                 // Default reverts to Main Sequence later
+    m_InitialStellarType2.typeString                               = INITIAL_STELLAR_TYPE_LABEL.at(m_InitialStellarType.type);   // Default reverts to Main Sequence later
 
     // Initial mass ratio
     m_MassRatio                                                     = 1.0;
@@ -785,22 +786,6 @@ bool Options::AddOptions(OptionValues *p_Options, po::options_description *p_Opt
             "debug-level",                                                 
             po::value<int>(&p_Options->m_DebugLevel)->default_value(p_Options->m_DebugLevel),                                                                                                     
             ("Determines which print statements are displayed for debugging (default = " + std::to_string(p_Options->m_DebugLevel) + ")").c_str()
-        )
-
-        (
-            "initial-stellar-type",                                            
-            po::value<int>(&p_Options->m_SuppliedStellarType)->default_value(p_Options->m_SuppliedStellarType),                                                                          
-            ("Initial stellar type (Hurley type) for the star (SSE) (default = " + std::to_string(p_Options->m_SuppliedStellarType) + ")").c_str()
-        )
-        (
-            "initial-stellar-type-1",                                            
-            po::value<int>(&p_Options->m_SuppliedStellarType1)->default_value(p_Options->m_SuppliedStellarType1),                                                                          
[…]

[…]
