### 2021-12-20 0fcb204
cleaned up call for initial stellar type in main.cpp (it was still using previous stellar type num functionality)
