### 2021-12-20 8a15926
cleaned up temporary comments and string list of available stellar types
