### 2021-12-20 5a341ed
added initial stellar types to program options documentation
