### 2021-12-21 #3838 @ilyamandel
I think these are bib file references that aren't actually used in the JOSS
paper.  Jeff, perhaps you could simply extract only the necessary
references into a downselected bib file.  Or ask Simon to do that.
Ilya, trying to send a GitHub comment via email from my phone from Eden
(Eden, New South Wales ;-) ).

On Wed, Dec 22, 2021, 4:31 AM Christina Hedges ***@***.***>
wrote:

> Hi @jeffriley <https://github.com/jeffriley>, that's great. If we don't
> get to this till after the holiday that's totally fine, I think that's the
> last piece before I can recommend it for acceptance 👍
>
