### 2021-12-24 d23f2c8
fixed duplicates and description in notebook
