+++ utils/example_plots/methods_paper_plots/fig_16_Chirpmass_distribution/pythonSubmit.py
@@ -1,776 +1,0 @@
-import numpy as np
-import sys
-import os
-from subprocess import call
-import re
-import ntpath
-
-#### DISCLAIMER: This script uses the `pythonSubmit.py` format
-#### that has been replaced by the `runSubmit.py` and 
-#### `compasConfigDefault.yaml` combo as of v02.25.10.
-#### The `pythonSubmit.py` format will eventually become deprecated.
-
-# Check if we are using python 3
-python_version = sys.version_info[0]
-print("python_version =", python_version)
-
-
-class pythonProgramOptions:
-    """
-    A class to store and access COMPAS program options in python
-    """
-
-    # Do './COMPAS --help' to see all options
-    #-- Define variables
-
-    # environment variable COMPAS_EXECUTABLE_PATH is used for docker runs
-    # if COMPAS_EXECUTABLE_PATH is not set (== None) we assume this is an
-    # interactive run with python3
-    # if COMPAS_EXECUTABLE_PATH is set (!= None) we assume this is a run
-    # inside a docker container - we have different directories inside a 
-    # docker container (src, obj, bin), and the COMPAS executable resides
-    # in the bin directory (rather than the src directory)
-    compas_executable_override = os.environ.get('COMPAS_EXECUTABLE_PATH')
-
-    if (compas_executable_override is None):
-        git_directory = os.environ.get('COMPAS_ROOT_DIR')
-        compas_executable = os.path.join(git_directory, 'src/COMPAS')
-    else:
-        compas_executable = compas_executable_override
-
-    enable_warnings = False                                     # option to enable/disable warning messages
-
-    number_of_systems = 10  #number of systems per batch
-
-    populationPrinting = False
-
-    randomSeedFileName = 'randomSeed.txt'
-    if os.path.isfile(randomSeedFileName):
-        random_seed = int(np.loadtxt(randomSeedFileName))
-    else:
-        random_seed = 0 # If you want a random seed, use: np.random.randint(2,2**63-1)
-
-    # environment variable COMPAS_LOGS_OUTPUT_DIR_PATH is used primarily for docker runs
-    # if COMPAS_LOGS_OUTPUT_DIR_PATH is set (!= None) it is used as the value for the
-    # --output-path option
-    # if COMPAS_LOGS_OUTPUT_DIR_PATH is not set (== None) the current working directory
-    # is used as the value for the --output-path option
-    compas_logs_output_override = os.environ.get('COMPAS_LOGS_OUTPUT_DIR_PATH')
-
-    if (compas_logs_output_override is None):
-        output = os.getcwd()
-        output_container = None                                 # names the directory to be created and in which log files are created.  Default in COMPAS is "COMPAS_Output"
-    else:
-        output = compas_logs_output_override
-        output_container = None
-
-    # environment variable COMPAS_INPUT_DIR_PATH is used primarily for docker runs
-    # if COMPAS_INPUT_DIR_PATH is set (!= None) it is prepended to input filenames
-    # (such as grid_filename and logfile_definitions)
-    # if COMPAS_INPUT_DIR_PATH is not set (== None) the current working directory
-    # is prepended to input filenames
-    compas_input_path_override = os.environ.get('COMPAS_INPUT_DIR_PATH')
-
-    #-- option to make a grid of hyperparameter values at which to produce populations.
-    #-- If this is set to true, it will divide the number_of_binaries parameter equally
-    #-- amoungst the grid points (as closely as possible). See the hyperparameterGrid method below
-    #-- for more details. If this is set to True, some hyperparameter values defined in this method'gridOutputs/'+str(i)
-    #-- will be overwritten
-    hyperparameterGrid = False
-    hyperparameterList = False
-    shareSeeds = False
-
[…]
