### 2022-01-01 6cb9e13
fixed bitwise operators to booleans
