### 2022-01-05 b27d344
(lambda_enhancements) add enhanced Nanjing lambda option
