+++ src/TPAGB.cpp
@@ -105,7 +105,7 @@ double TPAGB::CalculateLambdaDewi() const {
  *
  * @return                                      Nanjing lambda for use in common envelope
  */
-double TPAGB::CalculateLambdaNanjingEnhanced(const int p_MassInd, const int p_Zind) {
+double TPAGB::CalculateLambdaNanjingEnhanced(const int p_MassInd, const int p_Zind) const {
 
 	DBL_VECTOR maxBG    = {};                                                       // [0] = maxB, [1] = maxG
 	DBL_VECTOR lambdaBG = {};                                                       // [0] = lambdaB, [1] = lambdaG
@@ -389,19 +389,19 @@ double TPAGB::CalculateLambdaNanjingEnhanced(const int p_MassInd, const int p_Zi
  * This function good for TPAGB stars.
  *
  *
- * double CalculateLambdaNanjing()
+ * double CalculateLambdaNanjing(const double p_Mass, const double p_Metallicity)
  *
  * @return                                      Nanjing lambda for use in common envelope
  */
-double TPAGB::CalculateLambdaNanjing() const {
+double TPAGB::CalculateLambdaNanjing(const double p_Mass, const double p_Metallicity) const {
 
 	DBL_VECTOR maxBG    = {};                                                       // [0] = maxB, [1] = maxG
 	DBL_VECTOR lambdaBG = {};                                                       // [0] = lambdaB, [1] = lambdaG
 	DBL_VECTOR a        = {};                                                       // 0..5 a_coefficients
 	DBL_VECTOR b        = {};                                                       // 0..5 b_coefficients
 
-    if (utils::Compare(m_Metallicity, LAMBDA_NANJING_ZLIMIT) > 0) {                 // Z>0.5 Zsun: popI
-        if (utils::Compare(m_MZAMS, 1.5) < 0) {
+    if (utils::Compare(p_Metallicity, LAMBDA_NANJING_ZLIMIT) > 0) {                 // Z>0.5 Zsun: popI
+        if (utils::Compare(p_Mass, 1.5) < 0) {
             maxBG = { 2.5, 1.5 };
             if (utils::Compare(m_Radius, 200.0) > 0) lambdaBG = { 0.05, 0.05 };
             else  {
@@ -409,7 +409,7 @@ double TPAGB::CalculateLambdaNanjing() const {
                 lambdaBG   = { tmp, tmp };
             }
         }
-        else if (utils::Compare(m_MZAMS, 2.5) < 0) {
+        else if (utils::Compare(p_Mass, 2.5) < 0) {
             maxBG = { 4.0, 2.0 };
             if (utils::Compare(m_Radius, 340.0) > 0) lambdaBG = { 3.589970, 0.514132 };
             else {
@@ -417,7 +417,7 @@ double TPAGB::CalculateLambdaNanjing() const {
                 b = { 0.48271, 0.00584, -6.22051E-05, 2.41531E-07, -3.1872E-10, 0.0 };
             }
         }
-        else if (utils::Compare(m_MZAMS, 3.5) < 0) {
+        else if (utils::Compare(p_Mass, 3.5) < 0) {
             maxBG = { 500.0, 10.0 };
             if (utils::Compare(m_Radius, 400.0) > 0) lambdaBG = { 116.935557, 0.848808 };
             else {
[…]
