+++ src/Log.cpp
@@ -2352,81 +2352,6 @@ hid_t Log::CreateHDF5Dataset(const string p_Filename, const hid_t p_GroupId, con
 
 
 /*
- * Create a dataset subordinate to a group in an HDF5 file
- * 
- * 
- * 
- * 
- * hid_t Log::CreateHDF5Dataset(const string p_Filename, const hid_t p_GroupId, const string p_DatasetName, const hid_t p_H5DataType, const string p_UnitsStr, const size_t p_HDF5ChunkSize)
- *
- * @param   [IN]    p_Filename                  The filename of the HDF5 file (for error logging)
- * @param   [IN]    p_GroupId                   The group id under which the dataset should be created
- * @param   [IN]    p_DatasetName               The dataset name (this (generally) corresponds to the COMPAS column header)
- * @param   [IN]    p_H5DataType                The HDF5 datatype for the dataset
- * @param   [IN]    p_UnitsStr                  The units string to be associated with the dataset (generally corresponds to COMPAS units string)
- * @param   [IN]    p_HDF5ChunkSize             Chunk size for this dataset
- * @return                                      HDF5 dataset id (-1 indicates failure)
- */
-hid_t Log::CreateHDF5Dataset(const string p_Filename, const hid_t p_GroupId, const string p_DatasetName, const hid_t p_H5DataType, const string p_UnitsStr, const size_t p_HDF5ChunkSize) {
-
-    hid_t h5Dset = -1;                                                                                              // datset id - return value
-
-    // create a 1-d HDF5 dataspace
-    hsize_t h5Dims[1]    = {0};                                                                                     // initially 0, but...
-    hsize_t h5MaxDims[1] = {H5S_UNLIMITED};                                                                         // ... unlimited
-    hid_t   h5Dspace     = H5Screate_simple(1, h5Dims, h5MaxDims);                                                  // create the dataspace
-    hid_t   h5CPlist     = H5Pcreate(H5P_DATASET_CREATE);                                                           // create the dataset creation property list
-    (void)H5Pset_alloc_time(h5CPlist, H5D_ALLOC_TIME_INCR);                                                         // allocate space on disk incrementally
-    (void)H5Pset_layout(h5CPlist, H5D_CHUNKED);                                                                     // must be chunked when using unlimited dimensions
-
-    hsize_t h5ChunkDims[1] = {p_HDF5ChunkSize};                                                                     // chunk size - affects performance
-    herr_t h5Result = H5Pset_chunk(h5CPlist, 1, h5ChunkDims);                                                       // set chunk size
-    if (h5Result < 0) {                                                                                             // ok?
-        Squawk("ERROR: Unable to set chunk size for HDF5 container file " + p_Filename);                            // no - announce error
-    }
-    else {                                                                                                          // yes - chunk size set ok
-        // create HDF5 dataset
-        string h5DsetName = p_DatasetName;                                                                          // dataset name 
-        h5DsetName        = utils::trim(h5DsetName);                                                                // remove leading and trailing blanks
-        h5Dset            = H5Dcreate(p_GroupId,                                                                    // create the dataset in group p_GroupId
-                                      h5DsetName.c_str(),                                                           // dataset name
-                                      p_H5DataType,                                                                 // datatype
-                                      h5Dspace,                                                                     // dataspace
-                                      H5P_DEFAULT,                                                                  // dataset link property list                                                                     
-                                      h5CPlist,                                                                     // dataset creation property list
-                                      H5P_DEFAULT);                                                                 // dataset access property list
-        if (h5Dset < 0) {                                                                                           // dataset created ok?
-            Squawk("ERROR: Unable to create HDF5 dataSet " + h5DsetName + " for file " + p_Filename);               // no - announce error
-        }
-        else {                                                                                                      // yes - dataset created ok
-            // create attribute for units
[…]
