+++ src/HeWD.h
@@ -46,19 +46,18 @@ protected:
 
 
     // member functions - alphabetically
-    double  CalculateLambdaDewi()                                                   { return BaseStar::CalculateLambdaDewi(); }                                     // Not supported - use BaseStar
-    double  CalculateLambdaNanjing()                                                { return BaseStar::CalculateLambdaNanjing; }                 // Not supported - use BaseStar     JR: todo: check this (type 10 not mentioned as not supported in original code)
-    double  CalculateLambdaNanjingEnhanced()                                        { return CalculateLambdaNanjing; }
+    double  CalculateLambdaDewi()                                                         { return BaseStar::CalculateLambdaDewi(); }                                     // Not supported - use BaseStar
+    double  CalculateLambdaNanjing(const double p_Mass, const double p_Metallicity) const { return BaseStar::CalculateLambdaNanjing(0.0, 0.0); }                          // Not supported - use BaseStar (0.0 are dummy values)    JR: todo: check this (type 10 not mentioned as not supported in original code)
     double  CalculateLuminosityOnPhase(const double p_Mass,
-                                               const double p_Time,
-                                               const double p_Metallicity) const    { return CalculateLuminosityOnPhase_Static(p_Mass, p_Time, p_Metallicity); }
-    double  CalculateLuminosityOnPhase() const                                      { return CalculateLuminosityOnPhase_Static(m_Mass, m_Age, m_Metallicity); }     // Use class member variables
+                                       const double p_Time,
+                                       const double p_Metallicity) const                  { return CalculateLuminosityOnPhase_Static(p_Mass, p_Time, p_Metallicity); }
+    double  CalculateLuminosityOnPhase() const                                            { return CalculateLuminosityOnPhase_Static(m_Mass, m_Age, m_Metallicity); }     // Use class member variables
 
-    double  CalculateRadiusOnPhase(const double p_Mass) const                       { return CalculateRadiusOnPhase_Static(p_Mass); }
-    double  CalculateRadiusOnPhase() const                                          { return CalculateRadiusOnPhase(m_Mass); }                                      // Use class member variables
-    std::tuple <double, STELLAR_TYPE> CalculateRadiusAndStellarTypeOnPhase() const  { return BaseStar::CalculateRadiusAndStellarTypeOnPhase(); }
+    double  CalculateRadiusOnPhase(const double p_Mass) const                             { return CalculateRadiusOnPhase_Static(p_Mass); }
+    double  CalculateRadiusOnPhase() const                                                { return CalculateRadiusOnPhase(m_Mass); }                                      // Use class member variables
+    std::tuple <double, STELLAR_TYPE> CalculateRadiusAndStellarTypeOnPhase() const        { return BaseStar::CalculateRadiusAndStellarTypeOnPhase(); }
 
-    bool    ShouldEvolveOnPhase() const                                             { return true; }                                                                // Always it seems...  JR: todo: check this
+    bool    ShouldEvolveOnPhase() const                                                   { return true; }                                                                // Always it seems...  JR: todo: check this
 };
 
 #endif // __HeWD_h__

[…]
