### 2022-01-06 8b43fa3
(lambda_enhancements) remove unintended changes from Options.cpp
