### 2022-01-08 148fff2
(lambda_enhancements) fix cCalculateLambdaNanjing headers
