### 2022-01-08 7b33c5a
(lambda_enhancements) fix cCalculateLambdaNanjing headers
