+++ src/changelog.h
@@ -859,7 +859,12 @@
 // 02.26.02     RTW - Dec 17, 2021   - Defect repair, Code cleanup:
 //                                      - Changed all occurences of PPOW(base, 1.0/3.0) with std::cbrt, as the former could not handle negative bases
 //                                      - Changed all occurences of sqrt with std::sqrt for consistency with the above change
+// 02.27.00     ML - Jan 08, 2021    - Enhancements:
+//                                      - Add enhanced Nanjing lambda option that continuously extrapolates beyond radial range
+//                                      - Add Nanjing lambda option to switch between calculation using rejuvenated mass and true birth mass
+//                                      - Add Nanjing lambda mass and metallicity interpolation options
+//                                      - No change in default behaviour
 
-const std::string VERSION_STRING = "02.26.02";
+const std::string VERSION_STRING = "02.27.00";
 
 # endif // __changelog_h__
