### 2022-01-09 d438ac1
(compasUtils) bug fix to getSNevents()
