### 2022-01-09 26d75d9
(compasUtils) make getEventHistory() work for COMPAS output with unordered seeds
