### 2022-01-09 7b75a11
Cleaned up treatment of HG donors having CONVECTIVE envelopes in LEGACY; fixes an issue with missed failed CEs introduced in 02.25.01
