### 2022-01-09 TeamCOMPAS/COMPAS#715 @ilyamandel
Since pythonSubmit has been replaced, I am closing this issue.
