### 2022-01-10 b902b58
Fixed logical flaw in if/else
