### 2022-01-10 4120001
Fixed logical flaw in if/else
