### 2022-01-12 b462800
(lambda_enhancements) add lambda enhancement options to compasConfigDefault
