+++ utils/preProcessing/compasConfigDefault.yaml
@@ -23,6 +23,10 @@ booleanChoices:
     --mass-transfer: True    
     ### COMMON ENVELOPE
     --common-envelope-allow-main-sequence-survive: True # Allow main sequence stars to survive CE. Was previously False by default
+    --common-envelope-lambda-nanjing-enhanced: False
+    --common-envelope-lambda-nanjing-interpolate-in-mass: False
+    --common-envelope-lambda-nanjing-interpolate-in-metallicity: False
+    --common-envelope-lambda-nanjing-use_rejuvenated-mass: False
     --revised-energy-formalism-nandez-ivanova: False
     ### SUPERNOVAE, KICKS AND REMNANTS
     --pair-instability-supernovae: True
