### 2022-01-12 712e7db
(lambda_enhancements) eliminate extraneous exit points in functions
