+++ src/BaseStar.cpp
@@ -1093,33 +1093,36 @@ double BaseStar::CalculateLambdaLoveridgeEnergyFormalism(const double p_EnvMass,
  * @return                                      Common envelope lambda parameter
  */ 
 double BaseStar::EvaluateLambdaNanjing() const {
+
     double mass = m_MZAMS;
+    double lambda = 0.0;
     if (OPTIONS->CommonEnvelopeLambdaNanjingUseRejuvenatedMass()) {mass = m_Mass0;}                              // Use rejuvenated mass to calculate lambda instead of true birth mass
     
     if (OPTIONS->CommonEnvelopeLambdaNanjingEnhanced()) {                                                        // If using enhanced Nanjing lambda's
         if (OPTIONS->CommonEnvelopeLambdaNanjingInterpolateInMass()) {
             if (OPTIONS->CommonEnvelopeLambdaNanjingInterpolateInMetallicity()) {
-                return BaseStar::CalculateMassAndZInterpolatedLambdaNanjing(mass, m_Metallicity);
+                lambda = BaseStar::CalculateMassAndZInterpolatedLambdaNanjing(mass, m_Metallicity);
             }
             else {
                 int Zind = 0;
                 if (utils::Compare(m_Metallicity, LAMBDA_NANJING_ZLIMIT) < 0) {Zind = 0;} else {Zind = 1;}
-                return BaseStar::CalculateMassInterpolatedLambdaNanjing(mass, Zind);
+                lambda = BaseStar::CalculateMassInterpolatedLambdaNanjing(mass, Zind);
             }
         }
         else {
             int massInd = BaseStar::FindLambdaNanjingNearestMassIndex(mass);                                    // Do not interpolate in mass, so need to use nearest mass bin
             if (OPTIONS->CommonEnvelopeLambdaNanjingInterpolateInMetallicity()) {
-                return BaseStar::CalculateZInterpolatedLambdaNanjing(m_Metallicity, massInd);
+                lambda = BaseStar::CalculateZInterpolatedLambdaNanjing(m_Metallicity, massInd);
             }
             else {
                 int Zind = 0;
                 if (utils::Compare(m_Metallicity, LAMBDA_NANJING_ZLIMIT) < 0) {Zind = 0;} else {Zind = 1;}
-                return BaseStar::CalculateLambdaNanjingEnhanced(massInd, Zind);
+                lambda = BaseStar::CalculateLambdaNanjingEnhanced(massInd, Zind);
             }
         }
     }
-    else { return CalculateLambdaNanjing(mass, m_Metallicity); }
+    else { lambda = CalculateLambdaNanjing(mass, m_Metallicity); }
+    return lambda
 }
 
 
@@ -1133,18 +1136,20 @@ double BaseStar::EvaluateLambdaNanjing() const {
  */ 
 double BaseStar::CalculateMassAndZInterpolatedLambdaNanjing(const double p_Mass, const double p_Z) const {
 
+    double lambda = 0.0;
     if (utils::Compare(m_Metallicity, LAMBDA_NANJING_POPII_Z) < 0) {
-        return BaseStar::CalculateMassInterpolatedLambdaNanjing(p_Mass, 0);                    // Use lambda for pop. II metallicity
+        lambda = BaseStar::CalculateMassInterpolatedLambdaNanjing(p_Mass, 0);                    // Use lambda for pop. II metallicity
     }
     else if (utils::Compare(m_Metallicity, LAMBDA_NANJING_POPI_Z) > 0) {
-        return BaseStar::CalculateMassInterpolatedLambdaNanjing(p_Mass, 1);                    // Use lambda for pop. I metallicity
[…]
