+++ src/CHeB.cpp
@@ -468,11 +468,11 @@ double CHeB::CalculateLambdaNanjingEnhanced(const int p_MassInd, const int p_Zin
  * This function good for CHeB stars.
  *
  *
- * double CalculateLambdaNanjing(const double p_Mass, const double p_Metallicity)
+ * double CalculateLambdaNanjingStarTrack(const double p_Mass, const double p_Metallicity)
  *
  * @return                                      Nanjing lambda for use in common envelope
  */
-double CHeB::CalculateLambdaNanjing(const double p_Mass, const double p_Metallicity) const {
+double CHeB::CalculateLambdaNanjingStarTrack(const double p_Mass, const double p_Metallicity) const {
 
 	DBL_VECTOR maxBG    = {};                                                       // [0] = maxB, [1] = maxG
 	DBL_VECTOR lambdaBG = {};                                                       // [0] = lambdaB, [1] = lambdaG
