+++ src/TPAGB.cpp
@@ -389,11 +389,11 @@ double TPAGB::CalculateLambdaNanjingEnhanced(const int p_MassInd, const int p_Zi
  * This function good for TPAGB stars.
  *
  *
- * double CalculateLambdaNanjing(const double p_Mass, const double p_Metallicity)
+ * double CalculateLambdaNanjingStarTrack(const double p_Mass, const double p_Metallicity)
  *
  * @return                                      Nanjing lambda for use in common envelope
  */
-double TPAGB::CalculateLambdaNanjing(const double p_Mass, const double p_Metallicity) const {
+double TPAGB::CalculateLambdaNanjingStarTrack(const double p_Mass, const double p_Metallicity) const {
 
 	DBL_VECTOR maxBG    = {};                                                       // [0] = maxB, [1] = maxG
 	DBL_VECTOR lambdaBG = {};                                                       // [0] = lambdaB, [1] = lambdaG
