+++ src/EAGB.cpp
@@ -371,11 +371,11 @@ double EAGB::CalculateLambdaNanjingEnhanced(const int p_MassInd, const int p_Zin
  * This function good for EAGB stars.
  *
  *
- * double CalculateLambdaNanjing(const double p_Mass, const double p_Metallicity)
+ * double CalculateLambdaNanjingStarTrack(const double p_Mass, const double p_Metallicity)
  *
  * @return                                      Nanjing lambda for use in common envelope
  */
-double EAGB::CalculateLambdaNanjing(const double p_Mass, const double p_Metallicity) const {
+double EAGB::CalculateLambdaNanjingStarTrack(const double p_Mass, const double p_Metallicity) const {
 
 	DBL_VECTOR maxBG    = {};                                                       // [0] = maxB, [1] = maxG
 	DBL_VECTOR lambdaBG = {};                                                       // [0] = lambdaB, [1] = lambdaG
