### 2022-01-14 #3838 @ilyamandel
Thank you, @christinahedges !

The version of record for JOSS is v02.27.00:
https://github.com/TeamCOMPAS/COMPAS/releases/tag/v02.27.00

By archiving, do you mean creating a Zenodo archive of the repo?  I haven't done that before, but will attempt to create one (though possibly not until after the weekend).
