### 2022-01-17 4b7f128
Updated pythonSubmit -> runSubmit in git workflow and docker rst
