### 2022-01-17 066f1ad
Modified pythonSubmit -> runSubmit in /online-docs/pages/User guide/Running COMPAS
