+++ online-docs/pages/User guide/Running COMPAS/running-via-docker-running.rst
@@ -1,27 +1,34 @@
 Running the COMPAS Docker image
 ===============================
 
-COMPAS can be configured as usual via command line arguments passed to the COMPAS executable or via a ``pythonSubmit.py`` file in the 
+COMPAS can be configured as usual via command line arguments passed to the COMPAS executable or via a ``runSubmit.py`` file in the 
 ``Docker`` environment.
 
 
-via pythonSubmit.py
+via runSubmit.py
 -------------------
 
-To run COMPAS via a ``pythonSubmit.py`` file, type::
+To run COMPAS via a ``runSubmit.py`` file, type::
 
     docker run                                                  \
         --rm                                                    \
         -it                                                     \
         -v $(pwd)/compas-input:/app/COMPAS/config               \
         -v $(pwd)/compas-logs:/app/COMPAS/logs                  \
-        -v $(pwd)/pythonSubmit.py:/app/starts/pythonSubmit.py   \
+        -v $(pwd)/runSubmit.py:/app/starts/runSubmit.py   \
         -e COMPAS_EXECUTABLE_PATH=/app/COMPAS/bin/COMPAS        \
         -e COMPAS_INPUT_DIR_PATH=/app/COMPAS/config             \
         -e COMPAS_LOGS_OUTPUT_DIR_PATH=/app/COMPAS/logs         \
         teamcompas/compas                                       \
-        python3 /app/starts/pythonSubmit.py                     
+        python3 /app/starts/runSubmit.py                     
 
+
+NOTE: if you decide to execute using ``runSubmit.py``, you will need 
+a ``compasConfigDefault.yaml``  file in the same directory. This file 
+can be find in the same directory as he ``runSubmit.py``, and contains
+the default COMPAS choices for stellar and binary physics. This choices
+can be changed by modifying the options availabe in the ``compasConfigDefault.yaml`` 
+file.
 
 Breaking down this command:
 
@@ -39,7 +46,7 @@ mount ``<path-on-host>`` to ``<path-in-container>``\ [#f3]_. |br|
 
 This time we not only want to read the COMPAS input files (i.e. grid file and/or logfile-definitions file) on the
 host from the container, and get the output from COMPAS in the container onto the host machine, we also want to 
-supply a ``pythonSubmit.py`` to the container from the host machine.
+supply a ``runSubmit.py`` to the container from the host machine.
 
 **-e VAR_NAME=value** |br|
 set the environment variable ``VAR_VAME`` to `value`\ [#f4]_.
@@ -47,14 +54,14 @@ set the environment variable ``VAR_VAME`` to `value`\ [#f4]_.
 **teamcompas/compas** |br|
 the image to run.
 
-**python3 /app/starts/pythonSubmit.py** |br|
+**python3 /app/starts/runSubmit.py** |br|
 the command to run when the container starts.
 
 
 via the command line
 --------------------
 
-To run the COMPAS executable from the command line (i.e. without ``pythonSubmit.py``), type::
+To run the COMPAS executable from the command line (i.e. without ``runSubmit.py``), type::
 
     docker run                                      \
         --rm                                        \
@@ -102,10 +109,10 @@ forces logs to go to the directory that is mapped to the host machine.
 Environment variables
 ---------------------
 
-Three new environment variables are used in ``pythonSubmit.py``.  These environment variables are used primarily in the ``Docker``
[…]
