+++ online-docs/pages/User guide/Tutorial/example-compas-run.rst
@@ -1,26 +1,29 @@
 Tutorial: simple COMPAS run
 ===========================
 
-This tutorial assumes that you have already built the COMPAS executable as described in :doc:`../../Getting started/building-COMPAS`.
+NOTE: we are currently updating our documentation and will include a ``compasConfigDefault.yaml`` for the demo asap.
 
-For this example you will need the python script ``pythonSubmitDemo.py``, which specifies all the program options (physics assumptions, 
-output types) and runs COMPAS in the terminal. Although the primary functionality of COMPAS is to evolve a whole population of binary 
-stars rapidly, for now, let's focus on evolving a single stellar system and examining the detailed output.
+..
+    This tutorial assumes that you have already built the COMPAS executable as described in :doc:`../../Getting started/building-COMPAS`.
 
-If you haven't yet defined the ``COMPAS_ROOT_DIR`` environment variable, do that now::
+    For this example you will need the python script ``pythonSubmitDemo.py``, which specifies all the program options (physics assumptions, 
+    output types) and runs COMPAS in the terminal. Although the primary functionality of COMPAS is to evolve a whole population of binary 
+    stars rapidly, for now, let's focus on evolving a single stellar system and examining the detailed output.
 
-    export COMPAS_ROOT_DIR=path-to-compas
+    If you haven't yet defined the ``COMPAS_ROOT_DIR`` environment variable, do that now::
 
-where `path-to-compas` should be replaced with the path to the parent directory of the COMPAS `src` directory. Depending upon your system,
-for the ``export`` command to take effect, it may be necessary to either restart your session or execute the following command::
+        export COMPAS_ROOT_DIR=path-to-compas
 
-    source ~/.bashrc
+    where `path-to-compas` should be replaced with the path to the parent directory of the COMPAS `src` directory. Depending upon your system,
+    for the ``export`` command to take effect, it may be necessary to either restart your session or execute the following command::
 
-To start, change to the ``examples/methods_paper_plots/detailed_evolution/`` directory::
+        source ~/.bashrc
 
-  cd $COMPAS_ROOT_DIR/examples/methods_paper_plots/detailed_evolution/
+    To start, change to the ``examples/methods_paper_plots/detailed_evolution/`` directory::
 
-where you will find the script ``pythonSubmitDemo.py`` for this demo.
+      cd $COMPAS_ROOT_DIR/examples/methods_paper_plots/detailed_evolution/
+
+    where you will find the script ``pythonSubmitDemo.py`` for this demo.
 
 
 .. toctree::
