+++ online-docs/pages/User guide/docker.rst
@@ -113,7 +113,7 @@ from the host machine.
 
 NOTE: if you decide to execute using ``runSubmit.py``, you will need 
 a ``compasConfigDefault.yaml``  file in the same directory. This file 
-can be find in the same directory as he ``runSubmit.py``, and contains
+can be find in the same directory as the ``runSubmit.py``, and contains
 the default COMPAS choices for stellar and binary physics. This choices
 can be changed by modifying the options availabe in the ``compasConfigDefault.yaml`` 
 file.
