### 2022-01-17 98c485d
Modified pythonSubmit -> runSubmit in remaining rst files
