### 2022-01-17 24d4c28
Update README.md
Removed reference to docs branch that is no longer relevant.
