### 2022-01-17 2cd2595
updated pythonSubmitDemo to use OPTIMISTIC CE assumption, and common_envelope_allow_radiative_envelope_survive to True to reproduce the methods paper figure
