### 2022-01-18 7e26ee0
fixed immediate RLOF setting in pythonSubmitDemo.py
