+++ online-docs/pages/User guide/Running COMPAS/running-via-docker-running.rst
@@ -25,8 +25,8 @@ To run COMPAS via a ``runSubmit.py`` file, type::
 
 NOTE: if you decide to execute using ``runSubmit.py``, you will need 
 a ``compasConfigDefault.yaml``  file in the same directory. This file 
-can be find in the same directory as he ``runSubmit.py``, and contains
-the default COMPAS choices for stellar and binary physics. This choices
+can be find in the same directory as the ``runSubmit.py``, and contains
+the default COMPAS choices for stellar and binary physics. These choices
 can be changed by modifying the options availabe in the ``compasConfigDefault.yaml`` 
 file.
 
