### 2022-01-18 54188e5
Update docker.rst
Typo fixes
