### 2022-01-18 cbaaf85
Update running-via-python.rst
Typo fixes
