### 2022-01-18 b87ec49
Update README.md
Removed unnecessary bib lines
