### 2022-01-18 #3838 @whedon
:wave: @openjournals/joss-eics, this paper is ready to be accepted and published.
