### 2022-01-18 #3838 @christinahedges
Thank you so much to everyone involved in this review, the paper has now been recommended for acceptance! I believe the paper draft looks good, and the archive and version look correct. The EIC will now be able to step in. Thanks again for your hard work and patience, this one was a very long review process, but I believe there was some very helpful discussions had! Congratulations @ilyamandel!
