### 2022-01-19 8030306
added fastforwarding function to correctly assign stellar attributes to non-MS stars
