+++ src/WhiteDwarfs.h
@@ -34,7 +34,6 @@ public:
 
 protected:
 
-
     // member functions - alphabetically
             double      CalculateCOCoreMassOnPhase() const                  { return m_COCoreMass; }                                                // NO-OP
 
