+++ src/NS.h
@@ -57,6 +57,15 @@ protected:
         CalculateAndSetPulsarParameters();
     }
 
+    void FastForward() {                                                                                                                                                        // Set stellar attributes for stars initialized to this stellar type
+
+        m_Radius                                   = CalculateRadiusOnPhase();
+        m_Luminosity                               = CalculateLuminosityOnPhase();
+    
+        m_InitialStellarType                       = m_StellarType;
+        m_StellarTypePrev                          = m_StellarType;
+    }
+
     double m_MomentOfInertia;                                                                                                                                       // in CGS g cm^2
     double m_AngularMomentum;                                                                                                                                       // Current angular momentum in (Msol AU^2 yr-1)
     
