### 2022-01-19 e3afcbb
added initial radius and initial luminosity parameters to track those values even when the initial stellar type changes (so ZAMS values always correspond to H ZAMS).
