+++ src/HeWD.h
@@ -49,6 +49,8 @@ protected:
         m_Radius                                   = CalculateRadiusOnPhase();
         m_Luminosity                               = CalculateLuminosityOnPhase();
     
+        m_InitialLuminosity                        = m_Luminosity;
+        m_InitialRadius                            = m_Radius;
         m_InitialStellarType                       = m_StellarType;
         m_StellarTypePrev                          = m_StellarType;
     }
