### 2022-01-19 09f42c2
added initial L and R to STAR_PROPS
