+++ src/Star.cpp
@@ -36,7 +36,6 @@ Star::Star(const unsigned long int p_RandomSeed,
         (void)FastForward();                                                                                            // for non-MS stellar types, need to fast forward their attributes
     }
     else {                                                                                                              // no -
-    //if (utils::IsOneOf(p_InitialStellarType, {STELLAR_TYPE::STAR})) {                                                  // Stellar type is default?
         if (OPTIONS->CHEMode() != CHE_MODE::NONE && utils::Compare(m_Star->Omega(), m_Star->OmegaCHE()) >= 0) {             // CHE?
             (void)SwitchTo(STELLAR_TYPE::CHEMICALLY_HOMOGENEOUS, true);                                                     // yes
         }
