### 2022-01-20 #3838 @kyleniemeyer
Hi @ilyamandel, I am doing some final checks before accepting. There are just a few minor issues with the paper to address:

- I noticed an apparent discrepancy between author names on the paper and Zenodo archive: the paper has "Floris Kummer" while Zenodo says "Kummer, Florian". Could you please correct whichever of these is wrong? (Also, it would be great if you could add ORCIDs for the authors if at all possible, so that name issues are less important—but I recognize that might be challenging with such a large author list.)
- It looks like journal names are missing from the final paper PDF for a handful of references. Normally this happens because people are using custom journal macros that JOSS doesn't recognize. (For example, your `.bib` file has `{\apjs}`, etc.) Could you please replace these macros so that the journal titles show up properly?
