+++ paper.bib
@@ -20,7 +20,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Moe:2017ApJS,
        author = {{Moe}, Maxwell and {Di Stefano}, Rosanne},
         title = "{Mind Your Ps and Qs: The Interrelation between Period (P) and Mass-ratio (Q) Distributions of Binary Stars}",
-      journal = {\apjs},
+      journal = {ApJS},
      keywords = {binaries: close, binaries: general, stars: evolution, stars: formation, stars: massive, stars: statistics, Astrophysics - Solar and Stellar Astrophysics},
          year = "2017",
         month = "Jun",
@@ -41,7 +41,7 @@ archivePrefix = {arXiv},
 	{Acernese}, F. and {Ackley}, K. and {Adams}, C. and {Adams}, T. and 
 	{Addesso}, P. and {Adhikari}, R.~X. and et al.},
     title = "{Observation of Gravitational Waves from a Binary Black Hole Merger}",
-  journal = {\prl},
+  journal = {Phys.~Rev.~Lett},
 archivePrefix = "arXiv",
    eprint = {1602.03837},
  primaryClass = "gr-qc",
@@ -73,7 +73,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Scenario,
        author = {{Lipunov}, V.~M. and {Postnov}, K.~A. and {Prokhorov}, M.~E.},
         title = "{The Scenario Machine: restrictions on key parameters of binary evolution.}",
-      journal = {\aap},
+      journal = {A\&A},
      keywords = {STARS: BINARIES: GENERAL, BINARIES: CLOSE STARS: EVOLUTION, STARS: NEUTRON, GALAXY: STELLAR CONTENT},
          year = "1996",
         month = "Jun",
@@ -86,7 +86,7 @@ archivePrefix = {arXiv},
 @ARTICLE{IBiS,
        author = {{Tutukov}, A. and {Yungelson}, L.},
         title = "{Double-degenerate semidetached binaries with helium secondaries: cataclysmic variables, supersoft X-ray sources, supernovae and accretion-induced collapses}",
-      journal = {\mnras},
+      journal = {MNRAS},
      keywords = {ACCRETION, ACCRETION DISCS, BINARIES: CLOSE, NOVAE, CATACLYSMIC VARIABLES, SUPERNOVAE: GENERAL, X-RAYS: GENERAL},
          year = "1996",
         month = "Jun",
@@ -102,7 +102,7 @@ archivePrefix = {arXiv},
 @ARTICLE{SeBa,
        author = {{Portegies Zwart}, S.~F. and {Verbunt}, F.},
         title = "{Population synthesis of high-mass binaries.}",
-      journal = {\aap},
+      journal = {A\&A},
      keywords = {STARS: BINARIES: CLOSE, STARS: NEUTRON, STARS: SUPERNOVAE, STARS: EVOLUTION},
          year = "1996",
         month = "May",
@@ -115,7 +115,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Hurley:2002rf,
        author = {{Hurley}, Jarrod R. and {Tout}, Christopher A. and {Pols}, Onno R.},
         title = "{Evolution of binary stars and the effect of tides on binary populations}",
-      journal = {\mnras},
+      journal = {MNRAS},
      keywords = {METHODS: ANALYTICAL, METHODS: STATISTICAL, BINARIES: GENERAL, STARS: EVOLUTION, STARS: VARIABLES: OTHER, GALAXIES: STELLAR CONTENT, Astrophysics},
          year = "2002",
         month = "Feb",
@@ -135,7 +135,7 @@ archivePrefix = {arXiv},
          {Rasio}, Frederic A. and {Taam}, Ronald E. and {Zezas}, Andreas and
[…]
