### 2022-01-21 #3838 @ilyamandel
@kyleniemeyer:
- Thank you for catching the copy/paste mistake to Zenodo.  Name corrected, all the ORCIDs I know added.
- Bibliography updated on the JOSS branch.
