### 2022-01-21 #3838 @kyleniemeyer
@arfon @openjournals/dev whoops, I forgot that I needed to ping you about manually fixing the PDF to handle the large author list... hope this doesn't cause any headaches post-acceptance.
