### 2022-01-21 #3838 @kyleniemeyer
Many thanks to @katiebreivik and @HeloiseS for reviewing this submission, and @christinahedges for editing it.

(I am going to leave the issue open until that PDF formatting issue is resolved.)
