### 2022-01-23 #3838 @arfon
> @arfon @openjournals/dev whoops, I forgot that I needed to ping you about manually fixing the PDF to handle the large author list... hope this doesn't cause any headaches post-acceptance.

No problem. I've just fixed this. @ilyamandel – the updated paper might take ~24 hours to show as updated on the JOSS site due to caching we have in place.
